[split-svgs] Split out svgs
