+++ assets/components/ctaLink/ctaLink.jsx
@@ -3,7 +3,7 @@
 // ----- Imports ----- //
 
 import React from 'react';
-import { SvgArrowRightStraight } from 'components/svg/svg';
+import { SvgArrowRightStraight } from 'components/svgs/arrowRightStraight';
 import { clickSubstituteKeyPressHandler } from 'helpers/utilities';
 import uuidv4 from 'uuid';
 import { addQueryParamToURL } from 'helpers/url';
