+++ assets/components/directDebit/directDebitPopUpForm/directDebitPopUpForm.jsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { closeDirectDebitPopUp, resetDirectDebitFormError } from 'components/directDebit/directDebitActions';
 import DirectDebitForm from 'components/directDebit/directDebitForm/directDebitForm';
-import { SvgCross } from 'components/svg/svg';
+import { SvgCross } from 'components/svgs/cross';
 
 import type { Phase } from 'components/directDebit/directDebitActions';
