+++ assets/components/socialShare/socialShare.jsx
@@ -6,7 +6,8 @@ import React from 'react';
 
 import type { Node } from 'react';
 
-import { SvgFacebook, SvgTwitter } from 'components/svg/svg';
+import { SvgFacebook } from 'components/svgs/facebook';
+import { SvgTwitter } from 'components/svgs/twitter';
 
 
 // ---- Types ----- //
