--- assets/components/svg/svg.jsx
@@ -1,818 +1,0 @@
-// @flow
-
-// ----- Imports ----- //
-
-import React from 'react';
-
-
-// ----- SVGs ----- //
-
-// American Express credit card logo.
-function SvgAmexLogo() {
-
-  return (
-    <svg
-      className="svg-amex-logo"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 36 24"
-      preserveAspectRatio="xMinYMid"
-    >
-      <defs>
-        <path id="a" d="M36.125 30.18H0V15h36.125z" />
-      </defs>
-      <g transform="translate(-1 -6)" fill="none" fillRule="evenodd">
-        <rect fill="#246EA9" x="1" y="6" width="35.25" height="23.25" rx="2.25" />
-        <path d="M36.25 16.5V33H1V0h35.25v16.5" />
-        <path d="M1 17.276h1.547l.348-.842h.78l.347.842h3.042v-.642l.27.646h1.578l.27-.654v.65h7.557l-.003-1.38h.146c.103.004.132.014.132.183v1.196h3.908v-.322c.315.172.806.322 1.45.322h1.646l.35-.842h.782l.344.842h3.167v-.798l.482.798h2.536V12H29.17v.625l-.35-.625h-2.58v.625L25.916 12h-3.482c-.583 0-1.098.08-1.51.308V12H18.52v.308c-.266-.233-.624-.308-1.023-.308H8.72l-.594 1.363-.6-1.363H4.758v.625L4.452 12H2.096L1 14.506v2.77zm29.875-.744h-1.28l-1.715-2.837v2.837h-1.84l-.354-.843h-1.882l-.337.842h-1.06c-.443 0-.998-.098-1.31-.417-.32-.323-.484-.757-.484-1.445 0-.562.1-1.075.487-1.477.294-.304.75-.444 1.375-.444h.872v.808h-.853c-.332 0-.52.05-.7.226-.152.16-.26.46-.26.857 0 .41.08.698.253.89.135.148.39.194.628.194h.406l1.276-2.975h1.353l1.53 3.577V12.75h1.382l1.588 2.634V12.75h.925v3.782zm-11.628 0h.923V12.75h-.923v3.782zm-.524-2.76c0 .596-.403.913-.636 1.005.197.078.365.208.443.32.13.185.154.358.154.69v.745h-.913l-.003-.477c0-.224.023-.554-.146-.734-.13-.135-.328-.162-.656-.162h-.973v1.374h-.9V12.75h2.077c.463 0 .804.014 1.098.182.287.167.458.414.458.84zm-4.262 2.76h-3.02V12.75h3.02v.788h-2.114v.68h2.062v.777l-2.062-.002v.757h2.115v.782zm-3.695 0H9.84l-.006-2.96-1.31 2.96h-.793l-1.317-2.966v2.966h-1.84l-.346-.843H2.343l-.35.842H1.01l1.618-3.783h1.346L5.51 16.33V12.75h1.48l1.182 2.568 1.088-2.57h1.505v3.784zM25.378 14.9l-.628-1.506-.62 1.506h1.248zm-7.84-.597c-.122.07-.27.075-.446.075h-1.1v-.84h1.11c.162 0 .327.006.434.068.118.056.19.173.19.333 0 .167-.07.298-.188.363zM3.9 14.9l-.617-1.506-.617 1.506H3.9z" fill="#FFFFFE" />
-        <mask id="b" fill="#fff">
-          <use xlinkHref="#a" />
-        </mask>
-        <path d="M22.273 19.998c0 1.05-.793 1.27-1.593 1.27h-1.144v1.268h-1.778l-1.13-1.25-1.17 1.25h-3.63V18.75h3.684l1.128 1.24 1.165-1.24h2.926c.73 0 1.543.2 1.543 1.248zm-7.28 1.746H12.74v-.75h2.01v-.772h-2.01v-.685h2.295l1 1.102-1.042 1.104zm3.63.45l-1.405-1.544 1.405-1.49v3.033zm2.098-1.697h-1.184v-.96h1.195c.332 0 .56.128.56.46 0 .33-.217.5-.57.5zm6.18-1.746h3.06v.787h-2.145v.685h2.09v.772h-2.09v.75l2.145.006v.786H26.9V18.75zm-1.157 2.026c.203.075.37.207.45.32.128.185.15.356.15.69v.75h-.918v-.476c0-.226.02-.562-.148-.738-.133-.138-.337-.17-.672-.17h-.98v1.384h-.918V18.75h2.113c.466 0 .804.02 1.103.183.288.17.47.41.47.84 0 .6-.408.91-.65 1.003zm-.553-.468c-.12.07-.274.077-.45.077h-1.116v-.848h1.132c.16 0 .326.002.434.065.118.063.19.177.19.34 0 .164-.072.293-.19.366zm8.264.232c.183.183.277.414.277.8 0 .818-.516 1.196-1.438 1.196h-1.786v-.813h1.778c.176 0 .297-.022.374-.093.062-.057.108-.14.108-.245 0-.11-.047-.2-.11-.25-.073-.06-.167-.083-.327-.087-.858-.03-1.93.026-1.93-1.17 0-.548.35-1.127 1.31-1.127h1.834v.807h-1.676c-.17 0-.278.008-.37.067-.1.062-.137.155-.137.273 0 .144.087.242.202.284.096.033.197.042.353.045l.496.01c.5.014.837.098 1.044.305zm3.657 1.51c-.224.324-.657.486-1.243.486h-1.77v-.813h1.762c.174 0 .296-.022.37-.093.063-.057.11-.14.11-.245 0-.11-.047-.2-.112-.25-.07-.06-.164-.083-.322-.087-.86-.03-1.933.026-1.933-1.17 0-.548.353-1.127 1.312-1.127h1.826V18h-1.695c-.513 0-.885.12-1.147.308V18h-2.51c-.397 0-.867.098-1.092.308l.002-.308h-4.48v.308c-.357-.252-.957-.308-1.235-.308h-2.95v.308c-.288-.268-.91-.308-1.294-.308h-3.307l-.757.808-.708-.808H11v5.28h4.844l.78-.82.736.82h2.986v-1.242h.295c.394.005.865-.007 1.275-.18v1.42h2.465v-1.375h.117c.153 0 .168.008.168.155v1.22h7.485c.472 0 .966-.12 1.243-.337v.34h2.375c.495 0 .976-.067 1.343-.244v-.986zm0-2.493h-1.67c-.168 0-.28.008-.37.067-.094.062-.134.155-.134.273 0 .144.08.242.202.284.095.033.198.042.352.045l.495.01c.5.014.835.098 1.043.305.032.028.058.064.083.096v-1.08z" fill="#FFFFFE" mask="url(#b)" />
-      </g>
-    </svg>
-  );
-
-}
-
-// The right-pointing arrow commonly used on CTAs.
-function SvgArrowRightStraight() {
-
-  return (
-    <svg
-      className="svg-arrow-right-straight"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 20 17.89"
-      preserveAspectRatio="xMinYMid"
-    >
-      <path d="M20 9.35l-9.08 8.54-.86-.81 6.54-7.31H0V8.12h16.6L10.06.81l.86-.81L20 8.51v.84z" />
-    </svg>
-  );
-
-}
-
-// The generic credit card logo used on checkouts.
-function SvgCreditCard() {
-
-  return (
-    <svg
-      className="svg-credit-card"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 44 28.1"
-      preserveAspectRatio="xMinYMid"
-    >
-      <path d="M0 10.1v15l3 3h38.1l2.9-3.2V10.1M44 6.1V3l-2.9-3H3L0 3v3.1" />
-    </svg>
-  );
-
-}
-
-// Exclamation mark, used on error messages.
-function SvgExclamation() {
-
-  return (
-    <svg
-      className="svg-exclamation"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 2.2 8.9"
-      preserveAspectRatio="xMinYMid"
-    >
-      <path d="M1.1 8.9C.5 8.9 0 8.4 0 7.8c0-.6.5-1.1 1.1-1.1.6 0 1.1.5 1.1 1.1 0 .6-.5 1.1-1.1 1.1zM.6 0h1.1l.4.4-.4 5.2H.6L.2.4.6 0z" />
-    </svg>
-  );
-
-}
-
-// Alternate error exclamation mark, used on Direct Debit pop up.
-function SvgExclamationAlternate() {
-
-  return (
-    <svg
-      className="svg-exclamation-alternate"
-      width="22"
-      height="22"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g transform="translate(1 1)" fill="none">
-        <circle stroke="#E00000" cx="10" cy="10" r="10" />
-        <path
-          d="M8.786 4.832l.518 6.674h1.428l.536-6.674-.518-.546H9.304l-.518.546zm2.643 9.491c0-.775-.643-1.409-1.447-1.409a1.42 1.42 0 0 0-1.41 1.41c0 .756.642 1.39 1.41 1.39.786 0 1.447-.634 1.447-1.39z"
-          fill="#E00000"
-        />
-      </g>
-    </svg>
-  );
-
-}
-
-// The Facebook logo, in monochrome.
-function SvgFacebook() {
-
-  return (
-    <svg className="svg-facebook" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" preserveAspectRatio="xMinYMid">
-      <path d="M10.897 8.371h-2.229v5.943h-2.154v-5.943h-1.485v-2.154h1.485v-1.783c0-1.412.817-2.748 2.971-2.748.892 0 1.485.074 1.485.074v2.229h-1.337c-.743 0-.892.371-.892.966v1.337h2.229l-.074 2.08z" />
-    </svg>
-  );
-
-}
-
-// The Globe
-function SvgGlobe() {
-
-  return (
-    <svg className="svg-globe" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 96 96" preserveAspectRatio="xMinYMid">
-      <path d="M23.7,22.4 L45.8,22.4 L45.8,0.5 C40.7,0.7 35.6,1.8 30.8,3.7 C29.3,6 28,8.5 27,11.2 C25.6,14.8 24.5,18.6 23.7,22.4" />
-      <path d="M22,9.2 C22.2,8.7 22.4,8.2 22.6,7.7 C16.6,11.4 11.6,16.4 7.8,22.4 L18.2,22.4 C19.1,17.9 20.4,13.5 22,9.2" />
-      <path d="M15.8,50.7 L0.5,50.7 C0.8,56.8 2.3,62.7 4.9,68.2 L17.2,68.2 C16.3,62.4 15.9,56.5 15.8,50.7" />
-      <path d="M17.2,27.8 L4.9,27.8 C2.3,33.3 0.8,39.2 0.5,45.3 L15.8,45.3 C15.9,39.4 16.4,33.6 17.2,27.8" />
-      <path d="M78.8,22.4 L88.2,22.4 C84.7,16.9 80.1,12.2 74.7,8.5 C74.8,8.7 74.9,9 75,9.2 C76.6,13.5 77.9,17.9 78.8,22.4" />
-      <path d="M81.2,45.3 L95.5,45.3 C95.2,39.2 93.7,33.3 91.1,27.8 L79.7,27.8 C80.6,33.6 81.1,39.4 81.2,45.3" />
-      <path d="M73.3,22.4 C72.5,18.6 71.4,14.8 70,11.2 C69.1,8.7 67.9,6.4 66.5,4.1 C61.6,2.1 56.4,0.9 51.2,0.6 L51.2,22.5 L73.3,22.5 L73.3,22.4 Z" />
-      <path d="M51.2,68.2 L74.3,68.2 C75.2,62.4 75.7,56.6 75.8,50.7 L51.2,50.7 L51.2,68.2 Z" />
-      <path d="M79.7,68.2 L91.1,68.2 C93.7,62.7 95.2,56.8 95.5,50.7 L81.2,50.7 C81.1,56.5 80.6,62.4 79.7,68.2" />
-      <path d="M73.3,73.5 L51.2,73.5 L51.2,95.4 C56.5,95.1 61.6,93.9 66.5,91.9 C67.9,89.7 69.1,87.3 70,84.8 C71.4,81.1 72.5,77.4 73.3,73.5" />
-      <path d="M23.7,73.5 C24.5,77.3 25.6,81.1 27,84.8 C28,87.5 29.3,90 30.8,92.4 C35.6,94.3 40.7,95.3 45.8,95.6 L45.8,73.6 L23.7,73.5 Z" />
-      <path d="M45.8,27.8 L22.6,27.8 C21.7,33.6 21.2,39.4 21.1,45.3 L45.8,45.3 L45.8,27.8 Z" />
-      <path d="M75,86.7 C74.9,87 74.8,87.2 74.7,87.4 C80.1,83.7 84.7,79 88.2,73.5 L78.8,73.5 C77.9,78 76.6,82.4 75,86.7" />
-      <path d="M75.8,45.3 C75.7,39.4 75.2,33.6 74.3,27.8 L51.2,27.8 L51.2,45.3 L75.8,45.3 Z" />
-      <path d="M18.2,73.5 L7.8,73.5 C11.6,79.4 16.6,84.5 22.6,88.2 C22.4,87.7 22.1,87.2 22,86.7 C20.4,82.4 19.1,78 18.2,73.5" />
-      <path d="M21.1,50.7 C21.2,56.6 21.7,62.4 22.6,68.2 L45.7,68.2 L45.7,50.7 L21.1,50.7 Z" />
-    </svg>
-  );
-}
-
-// The main guardian logo.
-function SvgGuardianLogo() {
-
-  return (
-    <svg className="svg-guardian-logo" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 324.999 104.42" preserveAspectRatio="xMinYMid">
-      <path fill="#121212" d="M72.381 55.621l5.539-2.883V9.17h-4.188L63.493 22.755h-1.151l.646-15.138h44.387l.645 15.138h-1.222L96.772 9.17h-4.295v43.474l5.576 2.943v1.469H72.381zM113.322 53.693V5.418l-4.303-1.721v-.899L124.587 0h1.642v22.842l.43-.365c3.442-3.012 8.389-4.948 13.337-4.948 6.813 0 9.824 3.845 9.824 11.017v25.147l3.639 1.982.002 1.463h-20.348v-1.455l3.657-1.989v-25.23c0-3.944-1.722-5.521-4.948-5.521-2.151 0-4.006.674-5.378 1.776v29.043l3.585 1.98v1.396h-20.361V55.75l3.654-2.057zM165.513 38.869c.422 7.979 3.996 14.154 12.469 14.154 4.094 0 7.002-1.898 9.742-3.343v1.571c-2.117 2.894-7.484 6.963-14.968 6.963-13.131 0-19.838-7.3-19.838-19.938 0-12.354 7.343-20.048 19.203-20.048 11.155 0 16.944 5.576 16.944 20.261v.379h-23.552zm-.213-1.838l11.551-.706c0-9.885-1.693-16.445-5.082-16.445-3.601-.001-6.469 7.62-6.469 17.151M0 76.391c0-21.012 13.907-28.493 29.401-28.493 6.575 0 12.773 1.058 16.249 2.494l.303 14.663h-1.478l-9.096-14.188c-1.558-.669-3.035-.931-5.756-.931-8.239 0-12.447 9.515-12.315 25.118.157 18.669 3.401 27.142 10.959 27.142 1.965 0 3.477-.303 4.535-.756V81.454l-4.988-2.858V76.92h24.111v1.785l-4.915 2.749v19.728c-4.08 1.588-10.958 3.099-18.214 3.099C11.186 104.28 0 96.118 0 76.391M50.87 66.585v-1.212l16.219-2.859 1.777.148v31.951c0 3.852 1.852 5.035 4.962 5.035 1.999 0 3.81-.761 5.258-2.482V68.51l-4.444-1.925v-1.259l16.219-2.859 1.63.147v36.624l4.369 1.835v1.165l-15.997 1.967-1.629-.148v-4.812h-.445c-2.962 2.739-7.109 5.109-12.145 5.109-7.776 0-11.331-4.592-11.331-11.553V68.51l-4.443-1.925zM154.449 62.446l1.333.147-.004 11.828h.37c1.732-8.672 5.554-11.907 10.22-11.907.741 0 1.555.073 1.999.298v12.1c-.741-.222-2.073-.296-3.332-.296-3.703 0-6.425.669-8.813 1.761l.004 23.378 3.687 2.046.001 1.518h-21.027v-1.503l3.787-2.056V67.862l-4.444-1.325v-1.083l16.219-3.008zM196.163 63.443V50.967l-4.443-1.556v-.999l16.367-3.009 1.555.223v53.596l4.518 1.64v1.381l-16.145 2.178-1.258-.148v-4.435h-.371c-2.369 2.368-5.629 4.517-10.737 4.517-8.813 0-15.257-6.739-15.257-20.513 0-14.518 7.479-21.657 18.811-21.657 3.257-.001 5.701.592 6.96 1.258m-.023 34.315V65.72c-1.036-.666-1.786-1.489-4.473-1.396-4.377.155-7.08 6.764-7.08 18.539 0 10.589 1.943 16.509 7.776 16.314 1.629-.054 2.963-.637 3.777-1.419M232.846 62.409l1.407.149v37.197l3.697 2.046.002 1.517h-21.027v-1.504l3.775-2.055V68.436l-4.518-1.777v-1.241l16.664-3.009zm1.482-10.043c0 3.926-3.333 6.888-7.185 6.888-3.998 0-7.109-2.962-7.109-6.888 0-3.925 3.111-6.962 7.109-6.962 3.852 0 7.185 3.037 7.185 6.962M283.532 99.759V68.092l-4.443-1.555v-1.536l16.145-3.008 1.629.147v4.74h.444c3.481-3.11 8.666-5.11 13.774-5.11 7.037 0 10.146 3.332 10.146 10.738v27.188l3.771 2.103v1.517h-21.026v-1.503l3.776-2.055V73.25c0-4.074-1.776-5.702-5.109-5.702-2.148 0-3.919.544-5.555 1.767v30.433l3.703 2.053v1.517h-21.035v-1.503l3.78-2.056zM259.661 79.857v-5.311c0-8.003-1.745-10.622-6.693-10.622-.58 0-1.09.073-1.673.146l-8.803 11.931h-1.235V65.017c3.782-1.164 8.51-2.547 14.767-2.547 10.768 0 17.024 2.983 17.024 12.004v25.899l3.855 1.019v1.018c-1.528.946-4.584 1.819-7.93 1.819-5.311 0-7.857-1.746-9.021-4.656h-.363c-2.256 3.055-5.457 4.802-10.476 4.802-6.402 0-10.767-4.001-10.767-10.913 0-6.692 4.146-10.331 12.586-11.931l8.729-1.674zm0 17.896V81.75l-2.691.217c-4.221.364-5.748 3.057-5.748 9.021 0 6.475 2.111 8.147 5.094 8.147 1.673 0 2.619-.509 3.345-1.382M119.175 79.857v-5.311c0-8.003-1.745-10.622-6.693-10.622-.581 0-1.09.073-1.673.146l-8.803 11.931h-1.236V65.017c3.783-1.164 8.511-2.547 14.767-2.547 10.768 0 17.024 2.983 17.024 12.004v25.899l3.856 1.019v1.018c-1.528.946-4.584 1.819-7.93 1.819-5.31 0-7.857-1.746-9.021-4.656h-.364c-2.255 3.055-5.456 4.802-10.475 4.802-6.402 0-10.767-4.001-10.767-10.913 0-6.692 4.146-10.331 12.586-11.931l8.729-1.674zm0 17.896V81.75l-2.691.217c-4.221.364-5.748 3.057-5.748 9.021 0 6.475 2.111 8.147 5.093 8.147 1.672 0 2.619-.509 3.346-1.382" />
-    </svg>
-  );
-
-}
-
-// A padlock icon used to signify that checkouts are secure.
-function SvgLock() {
-
-  return (
-    <svg className="svg-lock" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 9 10.57" preserveAspectRatio="xMinYMid">
-      <path d="M8.17 3.68C8.17 1.65 6.52 0 4.5 0S.83 1.65.83 3.68H0V9.3l1.28 1.28h6.45L9 9.3V3.68h-.83zM4.5 1.57c1.12 0 2.1.9 2.1 2.1H2.4c0-1.12.98-2.1 2.1-2.1z" />
-    </svg>
-  );
-
-}
-
-// The standard PayPal logo.
-function SvgPaypalLogo() {
-
-  return (
-    <svg
-      className="svg-paypal-logo"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 37 24"
-      preserveAspectRatio="xMinYMid"
-    >
-      <g fillRule="nonzero" fill="none">
-        <path d="M34.823.7H2.177C.975.7 0 1.685 0 2.9v18.83c0 1.215.975 2.2 2.177 2.2h32.646c1.202 0 2.177-.985 2.177-2.2V2.9c0-1.215-.975-2.2-2.177-2.2z" fill="#F8F8F8" />
-        <path d="M24.775 9.563c-.31-.368-.86-.563-1.59-.563h-2.183c-.15 0-.276.11-.3.26l-.882 5.653c-.017.11.068.212.18.212h1.12c.104 0 .193-.077.21-.18l.25-1.604c.023-.148.15-.258.3-.258h.69c1.437 0 2.267-.703 2.483-2.096.098-.61.004-1.088-.278-1.423zm-1.338 1.502c-.12.79-.718.79-1.297.79h-.33l.233-1.476c.013-.09.09-.156.18-.156h.15c.394 0 .766 0 .958.227.115.136.15.337.107.615zM30.176 10.85H29.1c-.09 0-.17.07-.183.16l-.048.302-.076-.11c-.233-.34-.752-.452-1.27-.452-1.19 0-2.203.903-2.4 2.17-.103.63.042 1.235.4 1.656.328.388.796.55 1.354.55.957 0 1.488-.618 1.488-.618l-.048.3c-.018.114.07.217.185.217h.97c.152 0 .283-.113.307-.265l.58-3.692c.02-.113-.07-.217-.184-.217zm-1.5 2.1c-.102.617-.59 1.03-1.212 1.03-.312 0-.562-.1-.723-.29-.157-.19-.217-.458-.167-.758.097-.61.592-1.038 1.205-1.038.305 0 .553.102.717.294.165.194.23.464.182.763zM32.79 9h-1.174c-.103 0-.192.066-.208.155l-1.042 5.758c-.02.11.08.212.21.212h1.046c.174 0 .322-.11.348-.26l1.027-5.653c.02-.11-.08-.212-.208-.212z" fill="#179BD7" />
-        <path d="M8.957 9.563C8.647 9.195 8.097 9 7.367 9H5.184c-.15 0-.276.11-.3.26l-.882 5.653c-.017.11.068.212.18.212h1.042c.15 0 .276-.11.3-.26l.238-1.524c.023-.148.15-.258.3-.258h.69c1.437 0 2.267-.703 2.483-2.096.098-.61.004-1.088-.278-1.423zm-1.34 1.502c-.118.79-.716.79-1.295.79h-.33l.232-1.476c.014-.09.09-.156.18-.156h.15c.394 0 .766 0 .958.227.115.136.15.337.106.615zM14.36 10.85h-1.077c-.09 0-.17.07-.184.16l-.05.302-.074-.11c-.233-.34-.752-.452-1.27-.452-1.19 0-2.203.903-2.4 2.17-.104.63.042 1.235.4 1.656.327.388.796.55 1.354.55.957 0 1.488-.618 1.488-.618l-.048.3c-.018.114.07.217.185.217h.97c.152 0 .282-.113.307-.265l.58-3.692c.02-.113-.07-.217-.183-.217zm-1.5 2.1c-.104.617-.592 1.03-1.214 1.03-.312 0-.56-.1-.722-.29-.16-.19-.22-.458-.168-.758.096-.61.592-1.038 1.205-1.038.306 0 .554.102.718.294.164.194.23.464.182.763z" fill="#253B80" />
-        <path d="M19.628 10.75h-1.096c-.105 0-.203.048-.263.128l-1.512 2.052-.64-1.972c-.042-.123-.165-.208-.305-.208h-1.077c-.13 0-.222.118-.18.232l1.207 3.265-1.135 1.477c-.09.116 0 .276.155.276h1.094c.104 0 .202-.047.26-.125l3.647-4.85c.087-.116-.002-.275-.156-.275z" fill="#253B80" />
-      </g>
-    </svg>
-  );
-
-}
-
-// The mastercard logo.
-function SvgMastercardLogo() {
-
-  return (
-    <svg
-      className="svg-mastercard-logo"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 36 24"
-      preserveAspectRatio="xMinYMid"
-    >
-      <g fill="none" fillRule="evenodd">
-        <rect fill="#333" width="35.25" height="23.25" rx="2.25" />
-        <path d="M10.748 20.047V18.9c.017-.193-.053-.384-.193-.523-.14-.138-.333-.212-.533-.2-.265-.017-.52.108-.662.325-.13-.21-.37-.336-.623-.325-.225-.013-.44.088-.568.267v-.223h-.39v1.83h.383V19c-.014-.13.03-.257.122-.35.092-.095.222-.144.355-.137.272 0 .426.172.426.482v1.054h.382V19c-.013-.13.032-.257.124-.35.092-.094.22-.144.354-.137.28 0 .43.172.43.482v1.054l.392-.003zm6.145-1.826h-.687v-.554h-.384v.555h-.395v.333h.395v.867c0 .425.154.677.64.677.18 0 .358-.048.51-.14l-.117-.327c-.11.064-.235.1-.363.104-.197 0-.288-.123-.288-.306v-.878h.687l.003-.33zm3.502-.045c-.208-.005-.402.097-.513.267v-.22h-.375v1.83h.38v-1.034c0-.302.15-.497.417-.497.088 0 .175.015.256.047l.127-.346c-.094-.032-.193-.05-.292-.05v.005zm-5.295.19c-.232-.133-.5-.2-.77-.19-.476 0-.768.218-.768.574 0 .297.213.477.635.534l.197.027c.225.032.347.108.347.21 0 .143-.164.235-.445.235-.225.007-.446-.06-.627-.192l-.197.287c.242.16.527.242.818.238.544 0 .846-.245.846-.593 0-.35-.23-.48-.64-.538l-.197-.027c-.177-.023-.335-.076-.335-.19 0-.116.15-.227.375-.227.208.003.41.057.59.157l.173-.304zm5.72.766c-.014.258.085.51.273.69.188.184.446.28.712.268.246.012.488-.066.678-.218l-.197-.287c-.142.104-.315.162-.493.164-.33-.03-.58-.3-.58-.62 0-.322.25-.59.58-.62.178 0 .35.058.493.163l.197-.287c-.19-.152-.432-.23-.678-.218-.266-.013-.524.084-.712.266-.188.183-.287.434-.274.692v.01zm-2.656-.955c-.552 0-.932.382-.932.956-.017.255.078.504.26.686.184.182.438.28.7.272.28.01.557-.082.773-.257l-.197-.27c-.156.12-.35.188-.548.19-.287.01-.534-.192-.57-.468h1.41v-.153c0-.575-.36-.957-.89-.957l-.006.002zm0 .34c.13-.004.256.043.35.13.093.087.146.207.147.333h-1.01c.015-.266.246-.472.52-.463h-.006zm-5.11.618v-.912h-.377v.222c-.15-.177-.376-.277-.612-.267-.547 0-.99.43-.99.96 0 .532.443.962.99.962.236.01.463-.09.612-.268v.22h.378v-.917zm-1.527 0c-.013-.16.043-.318.155-.436.113-.118.27-.184.436-.184.33.03.582.3.582.62 0 .322-.252.59-.582.622-.165 0-.323-.066-.436-.184-.112-.117-.168-.275-.155-.435v-.002zm14.57-.956c-.207-.005-.4.097-.512.267v-.223h-.375v1.83h.38V19.02c0-.302.15-.497.417-.497.088 0 .175.015.256.047l.127-.345c-.094-.032-.193-.05-.292-.05v.005zm3.042 1.61c.024 0 .05.005.073.014.023.01.044.023.062.04.017.015.03.035.04.056.02.045.02.097 0 .142-.01.02-.023.04-.04.057-.018.017-.04.03-.062.04-.023.01-.05.014-.074.014-.078-.002-.147-.045-.18-.112-.02-.045-.02-.096 0-.142.01-.02.025-.04.042-.057.018-.017.04-.03.062-.04.023-.01.048-.017.075-.018v.005zm0 .324c.018 0 .038-.003.056-.01.017-.008.032-.02.045-.032.052-.054.052-.137 0-.19-.012-.014-.027-.024-.044-.03-.018-.01-.038-.013-.057-.013-.02 0-.04.004-.06.012-.016.007-.032.017-.046.03-.05.054-.05.137 0 .19.014.015.03.025.047.032.02.005.04.008.06.007v.004zm.01-.227c.02 0 .037.004.052.015.012.01.018.025.017.04 0 .013-.005.026-.014.035-.012.01-.026.016-.042.017l.057.063h-.044l-.053-.063h-.018v.063h-.037v-.168l.082-.002zm-.043.032v.046h.043c.008.004.016.004.024 0 .003-.004.003-.01 0-.016.003-.005.003-.012 0-.017-.008-.002-.016-.002-.024 0l-.043-.012zm-4.463-.778v-.916h-.378v.222c-.15-.177-.376-.277-.612-.267-.547 0-.99.43-.99.96 0 .532.443.962.99.962.236.01.463-.09.612-.268v.22h.378v-.913zm-1.528 0c-.013-.16.043-.318.156-.436.112-.117.27-.183.436-.183.33.03.58.3.58.62 0 .322-.25.59-.58.622-.166 0-.324-.068-.436-.185-.113-.118-.17-.276-.156-.436v-.003zm5.338 0v-1.655h-.378v.957c-.15-.18-.376-.28-.612-.27-.547 0-.99.43-.99.962 0 .53.443.96.99.96.236.01.463-.09.612-.266v.22h.378v-.91zm-1.528 0c-.013-.16.043-.318.156-.436.112-.117.27-.183.436-.183.315.047.547.31.545.62-.002.31-.238.57-.553.614-.165 0-.323-.065-.436-.182-.112-.118-.17-.275-.156-.435l.008.004z" fill="#FFF" fillRule="nonzero" />
-        <path fill="#FF5F00" fillRule="nonzero" d="M14.44 4.15h6.27v11.047h-6.27z" />
-        <path d="M14.893 9.672c0-2.163.988-4.205 2.683-5.54-2.88-2.27-7.015-1.94-9.5.76-2.484 2.7-2.484 6.862 0 9.562 2.485 2.7 6.62 3.03 9.5.76-1.696-1.336-2.685-3.38-2.683-5.542z" fill="#EB001B" fillRule="nonzero" />
-        <path d="M28.285 13.94v-.222h.084v-.045h-.214v.045h.084v.22h.045zm.414 0v-.267h-.066l-.075.19-.076-.19h-.065v.266h.044v-.203l.07.174h.048l.07-.173v.202h.047zM28.94 9.674c0 2.698-1.536 5.16-3.955 6.338-2.418 1.18-5.295.87-7.41-.8 1.695-1.335 2.684-3.377 2.684-5.54 0-2.16-.99-4.204-2.684-5.54 2.114-1.667 4.99-1.978 7.41-.798 2.418 1.18 3.953 3.64 3.953 6.338v.002z" fill="#F79E1B" fillRule="nonzero" />
-      </g>
-    </svg>
-  );
-
-}
-
-// A monochrome version of the Twitter logo.
-function SvgTwitter() {
-
-  return (
-    <svg className="svg-twitter" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" preserveAspectRatio="xMinYMid">
-      <path d="M13.553 5.05v.399c0 3.746-2.79 8.051-7.892 8.051-1.594 0-3.029-.478-4.224-1.275.239 0 .478.08.638.08 1.275 0 2.471-.478 3.428-1.196-1.196 0-2.232-.797-2.631-1.913.159 0 .319.08.558.08l.718-.08c-1.275-.239-2.232-1.434-2.232-2.79.399.239.797.319 1.275.319-.718-.478-1.275-1.355-1.275-2.312 0-.478.159-1.037.399-1.434 1.355 1.674 3.428 2.87 5.74 2.95-.08-.239-.08-.399-.08-.638 0-1.594 1.275-2.79 2.79-2.79.797 0 1.514.319 1.993.877.638-.08 1.196-.319 1.753-.638-.239.638-.638 1.196-1.196 1.514.558-.08 1.116-.239 1.594-.399-.319.319-.797.797-1.355 1.196l-.001-.001z" />
-    </svg>
-  );
-
-}
-
-// A user icon, with a head and body.
-function SvgUser() {
-
-  return (
-    <svg className="svg-user" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" preserveAspectRatio="xMinYMid">
-      <path d="M10 8.3c1.55 0 3.4-1.75 3.4-3.9S12.15 1 10 1 6.6 2.25 6.6 4.4s2 3.9 3.4 3.9zm5.86 3.4l-.86-.82c-1.67-.58-3.14-.88-5-.88-1.87 0-3.32.3-5 .86l-.86.85L2 18.16l.86.84h14.28l.86-.86-2.14-6.42v-.02z" />
-    </svg>
-  );
-
-}
-
-// A video camera icon.
-function SvgVideo() {
-
-  return (
-    <svg className="svg-video" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 12.728" preserveAspectRatio="xMinYMid">
-      <path d="M0 1.818v9.091l1.818 1.819h10.682v-12.728h-10.682zM18.75.165l-5 5v2.5l5 5.001h1.25v-12.501z" />
-    </svg>
-  );
-
-}
-
-// Just the 'P' of the PayPal logo.
-function SvgPaypalPLogo() {
-
-  return (
-    <svg className="svg-paypal-p-logo" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 23" preserveAspectRatio="xMinYMid">
-      <path d="M.555 20.444c-.58-.166-.503-.638-.439-1.035.239-1.578.497-3.156.749-4.727.713-4.477 1.432-8.96 2.154-13.449C3.174.287 3.471 0 4.42 0c2.8 0 5.6 0 8.387.096.976.065 1.93.312 2.813.728a4.171 4.171 0 0 1 2.246 4.926 15.908 15.908 0 0 0-2.736-.581c-2.18-.083-4.368 0-6.548 0-.749 0-1.026.242-1.142.977-.331 2.032-.658 4.066-.98 6.102-.059.357-.072.715-.104 1.073-.052.236-.11.473-.149.709-.316 1.955-.645 3.904-.935 5.865-.065.434-.252.569-.677.562-1.368-.013-2.697-.013-4.04-.013z" className="paypal-p-logo-0" />
-      <path d="M.555 20.444c1.348 0 2.69-.038 4.039 0 .425 0 .612-.127.677-.562.297-1.955.645-3.91.935-5.865.04-.236.097-.473.149-.709a1.23 1.23 0 0 1 1.168-.53h2.516a8.516 8.516 0 0 0 4.89-1.527c1.82-1.348 2.58-3.278 2.955-5.412 1.632.901 2.097 2.377 1.897 4.077a10.972 10.972 0 0 1-.742 2.92c-.955 2.293-2.845 3.398-5.265 3.698-.606.077-1.22.084-1.832.103-.73 0-1 .223-1.123.932-.245 1.432-.464 2.875-.69 4.313-.142.907-.361 1.099-1.29 1.099h-3.31c-.542 0-.742-.25-.645-.754.097-.505.168-1.067.264-1.693H2.806c-.748 0-1.516-.038-2.251-.09z" className="paypal-p-logo-1" />
-      <path d="M17.877 5.859c-.36 2.133-1.135 4.057-2.954 5.411a8.516 8.516 0 0 1-4.89 1.508c-.84.032-1.678 0-2.517 0a1.23 1.23 0 0 0-1.168.505c.033-.358.046-.722.104-1.074.322-2.036.65-4.07.98-6.101.123-.735.394-.971 1.142-.978 2.18 0 4.368-.057 6.549 0 .93.126 1.848.333 2.742.62a.272.272 0 0 1 .012.109z" className="paypal-p-logo-2" />
-    </svg>
-  );
-
-}
-
-// A Visa logo.
-function SvgVisaLogo() {
-
-  return (
-    <svg
-      className="svg-visa-logo"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 36 24"
-    >
-      <defs>
-        <rect id="a" width="35.25" height="23.25" rx="2.25" />
-      </defs>
-      <g fill="none" fillRule="evenodd">
-        <mask id="b" fill="#fff">
-          <use xlinkHref="#a" />
-        </mask>
-        <use fill="#FFF" xlinkHref="#a" />
-        <path fill="#193065" mask="url(#b)" d="M-18 5.357h57v-5.62h-57" />
-        <path fill="#EFAC3C" mask="url(#b)" d="M-8.25 23.316H43.5v-5.102H-8.25" />
-        <path d="M16.92 7.65l-1.574 8.29h-1.904l1.575-8.29h1.904zm8.014 5.354l1.003-3.113.577 3.114h-1.58zm2.126 2.937h1.76l-1.537-8.29h-1.625c-.366 0-.675.24-.81.61l-2.86 7.68h2l.398-1.237h2.443l.23 1.238zm-4.972-2.705c.008-2.188-2.686-2.31-2.668-3.287.006-.297.257-.613.808-.694.272-.04 1.025-.072 1.878.37l.334-1.758c-.458-.186-1.048-.366-1.782-.366-1.883 0-3.207 1.126-3.218 2.74-.012 1.194.946 1.86 1.667 2.256.742.407.99.668.988 1.03-.005.557-.593.803-1.14.812-.958.017-1.513-.29-1.956-.524l-.346 1.818c.446.23 1.267.43 2.117.44 2.003 0 3.312-1.114 3.318-2.837zm-7.89-5.584l-3.085 8.29H9.1L7.58 9.326c-.09-.407-.17-.556-.45-.728-.46-.28-1.215-.542-1.88-.705l.045-.24h3.24c.414 0 .785.31.88.844l.802 4.798 1.98-5.643H14.2z" fill="#193065" mask="url(#b)" />
-      </g>
-    </svg>
-  );
-
-}
-
-// A chevron symbol pointing up, used on buttons.
-function SvgChevronUp() {
-
-  return (
-    <svg className="svg-chevron-up" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30" preserveAspectRatio="xMinYMid">
-      <path d="M25.8999253 18.4499643 15.4499609 8 14.4499643 8 4 18.4499643 4.97499667 19.3999611 14.9499626 10.9999898 24.9249286 19.3999611Z" />
-    </svg>
-  );
-
-}
-
-function SvgCross() {
-
-  return (
-    <svg
-      className="svg-cross"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="-2 -2 38 38"
-      preserveAspectRatio="xMinYMid"
-    >
-      <path d="M16.525 19.87L30.865 33 33 30.87 19.9 16.5 33 2.13 30.866 0l-14.34 13.13L2.133.05 0 2.18 13.15 16.5 0 30.82l2.134 2.13" fillRule="evenodd" />
-    </svg>
-  );
-
-}
-
-function SvgMovingCircle() {
-
-  return (
-    <svg
-      className="svg-moving-circle"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 75 30"
-      preserveAspectRatio="xMinYMid"
-    >
-      <g fill="#FB0" filRule="evenodd">
-        <circle fillOpacity=".1" cx="60" cy="15" r="15" />
-        <circle fillOpacity=".4" cx="30" cy="15" r="15" />
-        <circle fillOpacity=".2" cx="46" cy="15" r="15" />
-        <circle cx="15" cy="15" r="15" />
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgDirectDebitSymbol() {
-  return (
-    <svg
-      className="svg-direct-debit-symbol"
-      width="35"
-      height="26"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M18.254 9.655v13.73s-7.444-1.436-7.444-6.893c0-4.353 5.51-6.272 7.444-6.837zm-14.618 6.55c0 3.2 5.371 7.878 14.119 8 .345 0 10.282-.41 10.129-11.817-.129-9.582-6.834-11.26-8.978-11.552V.252c8.548.58 15.27 6.078 15.27 12.75 0 7.06-7.52 12.804-16.766 12.804C8.163 25.806.68 20.06.68 13.003.681 5.945 8.163.2 17.41.2c.283 0 .565.005.844.016V9.4c-4.474-.346-14.618 1.897-14.618 6.805z"
-        fill="#212121"
-      />
-    </svg>
-  );
-
-}
-
-function SvgDirectDebitSymbolAndText() {
-  return (
-    <svg className="svg-direct-debit-symbol-and-text" width="86" height="28" xmlns="http://www.w3.org/2000/svg">
-      <g>
-        <g
-          stroke="null"
-          clipRule="evenodd"
-          fillRule="evenodd"
-        >
-          <path d="M40.13 6.054h4.382s4.34-.088 4.295 3.856c-.043 2.936-4.207 3.375-4.207 3.375h-4.47V6.054zM53.24 20.21c.124.573.538 1.205 1.922 1.27 1.84.044 2.893-.658 2.893-.658v1.359s-2.016.482-2.98.482c-1.008 0-3.989-.22-3.989-2.98 0-2.235 2.455-2.936 3.857-2.936 2.454.087 3.637.964 3.594 3.462 0 .036-4.039.01-5.296 0zm.006-1.097h3.363s0-1.27-1.622-1.27c-1.22 0-1.62.657-1.741 1.27zm10.638-1.373c.486-.527 1.303-.95 2.674-.95 2.717 0 3.155 2.279 3.111 2.893-.044.613-.263 2.936-3.418 2.936-1.191 0-1.92-.406-2.366-.911v.736h-2.148v-7.889h2.148v3.185h-.001zm1.84 3.652c1.272-.044 1.841-1.14 1.841-1.665 0-.527-.307-1.622-1.928-1.622-1.259 0-1.641.95-1.753 1.415v.572c.11.473.495 1.3 1.84 1.3zm18.54-.219l.043 1.227s-1.008.22-1.665.22c-.658 0-2.367-.395-2.367-2.06v-2.543H78.48V16.88h1.796V15.52h2.192v1.36h1.796v1.138h-1.796v2.192s0 1.096.876 1.096.92-.132.92-.132zM77.601 7.237h2.541v6.092h2.148V7.237h2.586V5.966h-7.275v1.27zm-.877.307l.263-1.272s-1.095-.482-3.024-.482c-1.884 0-4.909 1.052-4.909 3.9 0 2.806 3.2 3.77 5.172 3.77 1.973 0 2.63-.306 2.63-.306l-.219-1.403s-.876.263-2.367.263c-1.446-.044-2.936-1.183-2.936-2.454.044-1.14 1.359-2.41 3.068-2.41 1.314 0 2.322.394 2.322.394zM57.353 6.01s3.595-.176 3.595 1.928c0 1.797-2.06 1.929-2.06 1.929s.57.088.832.482c.22.35 1.666 2.98 1.666 2.98h-2.192s-.35-.438-.526-.745c-.13-.306-.526-.964-.526-.964s-.307-1.095-1.358-1.095c-.027 0-.456-.017-.79-.03v2.835h-2.103V6.01h3.462zm-1.358 3.287h1.14s1.577-.088 1.665-1.052c.044-.921-1.534-1.052-1.534-1.052h-1.27v2.104zm8.81 1.14h3.505v-1.36h-3.506V7.412h3.813V6.009h-5.96v7.32h6.004v-1.315h-3.856v-1.578zm8.545 5.609h2.148v-1.49H73.35v1.49zm0 6.399h2.148v-5.566H73.35v5.566zm-23.403-9.116h2.104v-7.32h-2.104v7.32zM40.13 15.17h4.382s4.34-.088 4.295 3.856c-.043 2.937-4.207 3.375-4.207 3.375h-4.47V15.17zm2.147 5.873h1.402s2.893-.176 2.893-2.41c0-2.236-3.068-2.455-3.068-2.455h-1.227v4.865zm0-9.116h1.402s2.893-.22 2.893-2.411c0-2.28-3.068-2.455-3.068-2.455h-1.227v4.866zM20.451 10.41v14.664s-8.503-1.533-8.503-7.362c0-4.65 6.295-6.698 8.503-7.301zm0-.273c-5.11-.37-16.698 2.026-16.698 7.268 0 3.418 6.136 8.415 16.128 8.546.395 0 11.746-.439 11.57-12.622C31.306 3.096 23.647 1.303 21.197.99V.367c9.764.62 17.443 6.492 17.443 13.62 0 7.538-8.59 13.673-19.152 13.673C8.925 27.66.378 21.524.378 13.986S8.924.313 19.487.313c.323 0 .644.005.964.017v9.807z" />
-        </g>
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgCirclesHeroDesktop() {
-
-  return (
-    <svg
-      className="svg-circles-hero-desktop"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 1400 406"
-      preserveAspectRatio="xMinYMid slice"
-    >
-      <g fill="none" fillRule="evenodd">
-        <path d="M18.512 320.347c6.889-12.381 18.47-22.136 33.143-26.343A55.1 55.1 0 0 1 81 293.685c-12.986 13.825-18.576 33.973-12.954 53.579s21.039 33.73 39.378 38.572a55.1 55.1 0 0 1-25.056 15.28c-14.673 4.208-29.663 2.073-42.067-4.776-12.515-6.912-22.398-18.622-26.657-33.477-4.26-14.855-2.084-30.022 4.867-42.516" fill="#00B2FF" opacity=".6" />
-        <path d="M-57.133 342.729c6.83-12.768 18.588-22.86 33.578-27.159 14.671-4.207 29.662-2.072 42.067 4.777-6.951 12.494-9.128 27.661-4.868 42.516 4.26 14.855 14.143 26.565 26.658 33.476-6.89 12.383-18.472 22.137-33.143 26.344-14.99 4.298-30.31 1.97-42.867-5.239-12.132-6.963-21.69-18.48-25.857-33.015-4.168-14.535-2.166-29.366 4.432-41.7" fill="#00B2FF" opacity=".4" />
-        <g transform="scale(-1 1) rotate(16 -980.416 -636.061)">
-          <path d="M56.6 57.688l.003-.006.006-.012.006.018-.001.006-.008.014-.006-.02zm42.075 38.006c.485-.54.96-1.092 1.426-1.654l.14-.169c.458-.557.905-1.124 1.343-1.7l.161-.213c.419-.556.829-1.122 1.23-1.698.07-.102.14-.206.213-.309.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044.003-.04.003-.077.004-.115a56.828 56.828 0 0 0-3.112-21.48c-.132-.374-.277-.739-.415-1.111-.002 0-.002-.001-.002-.003a56.725 56.725 0 0 0-7.067-13.304l-.089-.123a54.93 54.93 0 0 0-1.67-2.201l-.205-.26a56.782 56.782 0 0 0-1.863-2.197l-.05-.056a56.419 56.419 0 0 0-1.952-2.066c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.067-.033-.132-.07-.201-.104l-.003.008c-.005-.003-.009-.005-.014-.006l.006-.012c-.226-.11-.457-.204-.684-.31a60.424 60.424 0 0 0-2.674-1.167 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006-.644.008-1.288.027-1.933.058l-.281.012a59.28 59.28 0 0 0-2.183.16l-.318.031a58.288 58.288 0 0 0-2.222.26c-.711.097-1.423.211-2.133.338-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35l.002.002-.014.004-.003-.006c-.064.023-.127.05-.193.072-.822.293-1.637.598-2.444.928-.01.003-.018.008-.029.01a58.98 58.98 0 0 0-2.47 1.083c-.175.08-.343.168-.516.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303c-.495.467-.983.946-1.464 1.433l-.223.226c-.499.51-.987 1.032-1.467 1.564-.052.057-.105.114-.156.173a55.969 55.969 0 0 0-1.43 1.658l-.134.163a55.554 55.554 0 0 0-1.504 1.913 53.951 53.951 0 0 0-1.445 2.009 55.887 55.887 0 0 0-1.09 1.65c-.097.151-.192.305-.286.457a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496a54.546 54.546 0 0 0-.148 2.159A56.796 56.796 0 0 0 3.18 76.44c.13.375.277.74.415 1.111a56.72 56.72 0 0 0 8.827 15.631l.206.261a59.275 59.275 0 0 0 1.861 2.197l.05.055a56.419 56.419 0 0 0 1.953 2.066c.228.229.465.45.697.676.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.07.033.134.071.202.104l.004-.007a.03.03 0 0 0 .012.006l-.006.012c.228.11.46.205.688.312a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 1.934-.057c.093-.005.186-.008.28-.012.727-.04 1.457-.094 2.186-.16l.313-.032a54.834 54.834 0 0 0 2.09-.24c.046-.007.09-.011.134-.017a58.17 58.17 0 0 0 2.135-.34c.156-.028.31-.058.465-.087a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35l-.001-.002.013-.005.003.008c.066-.024.13-.051.196-.076a56.368 56.368 0 0 0 2.472-.937c.832-.34 1.654-.703 2.47-1.081.174-.082.342-.169.514-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305c.494-.466.982-.943 1.46-1.429l.227-.23a55.362 55.362 0 0 0 1.623-1.737z" fill="#00B2FF" />
-        </g>
-        <path d="M81 293.686a55.1 55.1 0 0 1 25.056-15.28 55.106 55.106 0 0 1 27.542-.77 56.503 56.503 0 0 0-15.337 35.722c-1.545 27.482 16.671 51.53 42.26 58.228a55.096 55.096 0 0 1-23.75 13.932 55.1 55.1 0 0 1-29.347.317c-18.338-4.84-33.756-18.966-39.378-38.57-5.622-19.607-.032-39.755 12.955-53.58" fill="#00B2FF" opacity=".8" />
-        <path d="M1311.685 287.418c-2.147 30.695-28.682 53.845-59.27 51.706-30.586-2.14-53.642-28.757-51.495-59.452 2.146-30.695 28.682-53.845 59.269-51.706 30.586 2.139 53.642 28.756 51.496 59.452" fill="#FF1921" opacity=".1" />
-        <path d="M1446.65 304.256c-2.147 30.696-28.683 53.845-59.27 51.707-30.586-2.14-53.642-28.757-51.496-59.452 2.147-30.695 28.683-53.845 59.27-51.706 30.586 2.139 53.642 28.756 51.495 59.451" fill="#FF1921" opacity=".05" />
-        <path d="M1139.038 216.524c-15.556-1.088-30.056 4.375-40.841 14.037 9.018 10.693 14.058 24.783 13.004 39.867-1.055 15.084-8.007 28.336-18.426 37.67 9.335 11.068 22.934 18.496 38.49 19.584 30.587 2.139 57.123-21.011 59.27-51.706 2.146-30.695-20.91-57.313-51.497-59.452" fill="#FF1921" opacity=".2" />
-        <path d="M1019.552 224.396c-9.313-10.68-22.67-17.812-37.896-18.877a55.1 55.1 0 0 0-28.771 5.79c15.576 10.822 25.233 29.367 23.81 49.714-1.422 20.346-13.566 37.367-30.498 45.916a55.1 55.1 0 0 0 27.686 9.738c15.226 1.064 29.445-4.14 40.154-13.42 10.805-9.362 18.037-22.87 19.115-38.287 1.078-15.415-4.204-29.8-13.6-40.574" fill="#FF1921" opacity=".6" />
-        <path d="M1098.196 230.56c-9.335-11.068-22.934-18.496-38.49-19.584-15.226-1.064-29.445 4.14-40.155 13.42 9.397 10.775 14.68 25.158 13.602 40.574-1.078 15.416-8.312 28.925-19.116 38.288 9.314 10.678 22.67 17.812 37.896 18.876 15.556 1.088 30.056-4.375 40.841-14.037 10.42-9.333 17.373-22.585 18.428-37.669 1.054-15.084-3.987-29.174-13.006-39.867" fill="#FF1921" opacity=".4" />
-        <g transform="rotate(4 -2339.192 11743.357)">
-          <path d="M100.101 94.04a59.115 59.115 0 0 0 1.483-1.87c.58-.768.99-1.334 1.39-1.91l.214-.309c.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044a56.995 56.995 0 0 0-3.108-21.594c-.132-.375-.277-.74-.415-1.112a56.746 56.746 0 0 0-8.829-15.631 57.603 57.603 0 0 0-2.067-2.457 56.73 56.73 0 0 0-2.002-2.122c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.438-.224-.67-.318-.896-.425a60.424 60.424 0 0 0-2.674-1.166 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006a55.331 55.331 0 0 0-9.07.859c-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35-1.03.365-1.845.67-2.652 1a57.84 57.84 0 0 0-2.5 1.093c-.174.08-.342.168-.515.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303a57.085 57.085 0 0 0-4.74 5.054 57.078 57.078 0 0 0-4.174 5.734c-.096.152-.19.306-.285.458a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496-.06.68-.109 1.362-.143 2.044A56.963 56.963 0 0 0 3.18 76.44c.13.376.277.74.415 1.112a56.742 56.742 0 0 0 8.827 15.632 59.92 59.92 0 0 0 2.067 2.457 56.73 56.73 0 0 0 2.003 2.12c.228.23.465.451.697.677.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.44.225.672.32.9.427a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 9.072-.859c.156-.027.31-.057.465-.086a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35a60.552 60.552 0 0 0 2.642-.997 58.305 58.305 0 0 0 2.512-1.096c.173-.082.34-.169.513-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305a56.786 56.786 0 0 0 3.151-3.22 60.168 60.168 0 0 0 1.585-1.83z" fill="#FF1921" />
-        </g>
-        <path d="M952.885 211.309a55.1 55.1 0 0 0-27.685-9.738 55.106 55.106 0 0 0-27.1 4.975 56.503 56.503 0 0 1 22.43 31.752c7.224 26.56-5.594 53.87-29.232 65.742a55.096 55.096 0 0 0 26.129 8.69 55.1 55.1 0 0 0 28.771-5.792c16.931-8.548 29.075-25.57 30.498-45.915 1.423-20.347-8.234-38.892-23.811-49.714" fill="#FF1921" opacity=".8" />
-        <path d="M1351.417-24.959c6.921 29.982-11.687 59.878-41.562 66.776-29.876 6.897-59.706-11.817-66.628-41.798-6.922-29.982 11.686-59.878 41.562-66.776 29.875-6.897 59.706 11.817 66.628 41.798" fill="#CE147C" opacity=".1" />
-        <path d="M1166.71-39.496c-15.194 3.508-27.464 12.972-34.953 25.364 11.751 7.59 20.69 19.59 24.092 34.324 3.402 14.733.628 29.438-6.607 41.41 12.164 7.856 27.34 10.983 42.535 7.475 29.875-6.897 48.483-36.794 41.561-66.775-6.921-29.982-36.752-48.695-66.628-41.798" fill="#CE147C" opacity=".2" />
-        <path d="M1054.746 2.966c-12.028-7.49-26.886-10.406-41.759-6.972a55.1 55.1 0 0 0-25.821 13.948c18.06 5.796 32.717 20.707 37.305 40.58 4.589 19.874-2.048 39.702-15.74 52.828a55.1 55.1 0 0 0 29.322 1.217c14.873-3.433 26.949-12.568 34.477-24.572 7.595-12.113 10.561-27.146 7.085-42.203-3.476-15.058-12.732-27.269-24.869-34.826" fill="#CE147C" opacity=".6" />
-        <path d="M1131.757-14.132c-12.164-7.855-27.34-10.983-42.535-7.475-14.871 3.434-26.947 12.568-34.476 24.573 12.136 7.557 21.394 19.768 24.87 34.825 3.476 15.058.509 30.091-7.087 42.204 12.03 7.489 26.888 10.405 41.76 6.972 15.194-3.508 27.463-12.972 34.953-25.365 7.234-11.972 10.01-26.678 6.608-41.41-3.401-14.734-12.342-26.735-24.093-34.324" fill="#CE147C" opacity=".4" />
-        <g transform="rotate(-13 564.614 -3706.003)">
-          <path d="M100.101 94.04a59.115 59.115 0 0 0 1.483-1.87c.58-.768.99-1.334 1.39-1.91l.214-.309c.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044a56.995 56.995 0 0 0-3.108-21.594c-.132-.375-.277-.74-.415-1.112a56.746 56.746 0 0 0-8.829-15.631 57.603 57.603 0 0 0-2.067-2.457 56.73 56.73 0 0 0-2.002-2.122c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.438-.224-.67-.318-.896-.425a60.424 60.424 0 0 0-2.674-1.166 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006a55.331 55.331 0 0 0-9.07.859c-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35-1.03.365-1.845.67-2.652 1a57.84 57.84 0 0 0-2.5 1.093c-.174.08-.342.168-.515.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303a57.085 57.085 0 0 0-4.74 5.054 57.078 57.078 0 0 0-4.174 5.734c-.096.152-.19.306-.285.458a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496-.06.68-.109 1.362-.143 2.044A56.963 56.963 0 0 0 3.18 76.44c.13.376.277.74.415 1.112a56.742 56.742 0 0 0 8.827 15.632 59.92 59.92 0 0 0 2.067 2.457 56.73 56.73 0 0 0 2.003 2.12c.228.23.465.451.697.677.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.44.225.672.32.9.427a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 9.072-.859c.156-.027.31-.057.465-.086a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35a60.552 60.552 0 0 0 2.642-.997 58.305 58.305 0 0 0 2.512-1.096c.173-.082.34-.169.513-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305a56.786 56.786 0 0 0 3.151-3.22 60.168 60.168 0 0 0 1.585-1.83z" fill="#CE147C" />
-        </g>
-        <path d="M987.166 9.943a55.1 55.1 0 0 0-29.322-1.218 55.106 55.106 0 0 0-24.46 12.68 56.503 56.503 0 0 1 30.731 23.808c14.675 23.287 10.401 53.15-8.733 71.415a55.096 55.096 0 0 0 27.528.67 55.1 55.1 0 0 0 25.82-13.95c13.693-13.124 20.33-32.953 15.742-52.825-4.588-19.874-19.245-34.785-37.306-40.58" fill="#CE147C" opacity=".8" />
-        <path d="M970.018 391.233c-15.195-3.508-30.371-.38-42.535 7.475 7.235 11.972 10.009 26.678 6.608 41.41-3.402 14.734-12.341 26.735-24.092 34.324 7.489 12.393 19.758 21.856 34.953 25.364 29.875 6.898 59.706-11.816 66.628-41.797 6.922-29.982-11.686-59.879-41.562-66.776" fill="#AB8958" opacity=".4" />
-        <path d="M850.62 380.064c-7.528-12.004-19.603-21.14-34.476-24.573a55.1 55.1 0 0 0-29.323 1.218c13.692 13.126 20.33 32.953 15.741 52.827-4.588 19.873-19.245 34.785-37.305 40.58a55.1 55.1 0 0 0 25.82 13.949c14.873 3.433 29.732.517 41.76-6.973 12.136-7.557 21.392-19.768 24.869-34.825 3.476-15.058.51-30.09-7.086-42.203" fill="#C28A2F" opacity=".8" />
-        <path d="M927.332 398.456c-7.489-12.393-19.758-21.857-34.953-25.365-14.871-3.433-29.73-.516-41.759 6.973 7.595 12.112 10.563 27.145 7.087 42.203-3.477 15.057-12.734 27.268-24.87 34.825 7.529 12.005 19.605 21.14 34.476 24.573 15.195 3.508 30.371.38 42.535-7.475 11.75-7.59 20.692-19.59 24.093-34.323s.626-29.44-6.609-41.411" fill="#C28A2F" opacity=".6" />
-        <g transform="rotate(13 -1081.625 3262.394)">
-          <path d="M56.6 57.688l.003-.006.006-.012.006.018-.001.006-.008.014-.006-.02zm42.075 38.006c.485-.54.96-1.092 1.426-1.654l.14-.169c.458-.557.905-1.124 1.343-1.7l.161-.213c.419-.556.829-1.122 1.23-1.698.07-.102.14-.206.213-.309.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044.003-.04.003-.077.004-.115a56.828 56.828 0 0 0-3.112-21.48c-.132-.374-.277-.739-.415-1.111-.002 0-.002-.001-.002-.003a56.725 56.725 0 0 0-7.067-13.304l-.089-.123a54.93 54.93 0 0 0-1.67-2.201l-.205-.26a56.782 56.782 0 0 0-1.863-2.197l-.05-.056a56.419 56.419 0 0 0-1.952-2.066c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.067-.033-.132-.07-.201-.104l-.003.008c-.005-.003-.009-.005-.014-.006l.006-.012c-.226-.11-.457-.204-.684-.31a60.424 60.424 0 0 0-2.674-1.167 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006-.644.008-1.288.027-1.933.058l-.281.012a59.28 59.28 0 0 0-2.183.16l-.318.031a58.288 58.288 0 0 0-2.222.26c-.711.097-1.423.211-2.133.338-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35l.002.002-.014.004-.003-.006c-.064.023-.127.05-.193.072-.822.293-1.637.598-2.444.928-.01.003-.018.008-.029.01a58.98 58.98 0 0 0-2.47 1.083c-.175.08-.343.168-.516.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303c-.495.467-.983.946-1.464 1.433l-.223.226c-.499.51-.987 1.032-1.467 1.564-.052.057-.105.114-.156.173a55.969 55.969 0 0 0-1.43 1.658l-.134.163a55.554 55.554 0 0 0-1.504 1.913 53.951 53.951 0 0 0-1.445 2.009 55.887 55.887 0 0 0-1.09 1.65c-.097.151-.192.305-.286.457a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496a54.546 54.546 0 0 0-.148 2.159A56.796 56.796 0 0 0 3.18 76.44c.13.375.277.74.415 1.111a56.72 56.72 0 0 0 8.827 15.631l.206.261a59.275 59.275 0 0 0 1.861 2.197l.05.055a56.419 56.419 0 0 0 1.953 2.066c.228.229.465.45.697.676.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.07.033.134.071.202.104l.004-.007a.03.03 0 0 0 .012.006l-.006.012c.228.11.46.205.688.312a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 1.934-.057c.093-.005.186-.008.28-.012.727-.04 1.457-.094 2.186-.16l.313-.032a54.834 54.834 0 0 0 2.09-.24c.046-.007.09-.011.134-.017a58.17 58.17 0 0 0 2.135-.34c.156-.028.31-.058.465-.087a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35l-.001-.002.013-.005.003.008c.066-.024.13-.051.196-.076a56.368 56.368 0 0 0 2.472-.937c.832-.34 1.654-.703 2.47-1.081.174-.082.342-.169.514-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305c.494-.466.982-.943 1.46-1.429l.227-.23a55.362 55.362 0 0 0 1.623-1.737z" fill="#C28A2F" />
-        </g>
-        <g fill="#F7691C" transform="rotate(-175 108.214 55.111)">
-          <path d="M183.063 131.066c12.148-10.5 19.834-26.018 19.834-43.33 0-29.724-22.657-54.158-51.652-56.99 10.044-8.683 23.139-13.934 37.46-13.934C220.348 16.812 246 42.45 246 74.076s-25.652 57.265-57.295 57.265c-1.904 0-3.786-.093-5.642-.275z" opacity=".399" />
-          <path d="M53.244 114.388a58.16 58.16 0 0 0 4.05.141c31.644 0 57.296-25.638 57.296-57.265a56.988 56.988 0 0 0-11.193-34.007c29.753 2.077 53.244 26.858 53.244 57.123 0 31.627-25.652 57.265-57.295 57.265-18.902 0-35.666-9.148-46.102-23.257z" opacity=".8" />
-          <ellipse cx="57.295" cy="57.264" rx="57.295" ry="57.264" />
-          <path d="M114.266 135.684c24.409-6.562 42.375-28.836 42.375-55.304 0-20.066-10.327-37.722-25.959-47.948a57.398 57.398 0 0 1 14.92-1.961c31.644 0 57.295 25.638 57.295 57.265 0 31.626-25.651 57.264-57.294 57.264-11.566 0-22.332-3.425-31.337-9.316z" opacity=".604" />
-        </g>
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgCirclesHeroMobileLandscape() {
-
-  return (
-    <svg
-      className="svg-circles-hero-mobile-landscape"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 768 382"
-      preserveAspectRatio="xMinYMid slice"
-    >
-      <g fill="none" fillRule="evenodd" transform="translate(-2)">
-        <path d="M728.406 348.214c-7.84-10.002-19.467-17.017-33.055-18.686a49.417 49.417 0 0 0-26.04 3.835c13.443 10.424 21.222 27.487 18.992 45.644-2.23 18.156-13.905 32.83-29.47 39.693a49.417 49.417 0 0 0 24.339 10.02c13.587 1.67 26.567-2.325 36.593-10.133 10.117-7.878 17.228-19.638 18.917-33.394 1.69-13.757-2.366-26.887-10.276-36.98" fill="#FF1921" opacity=".6" />
-        <path d="M798.554 357.427c-7.842-10.352-19.673-17.643-33.554-19.347-13.587-1.668-26.566 2.326-36.594 10.134 7.91 10.091 11.967 23.222 10.277 36.979-1.689 13.756-8.801 25.516-18.918 33.394 7.841 10.002 19.469 17.017 33.055 18.685 13.882 1.705 27.126-2.507 37.239-10.654 9.77-7.871 16.62-19.414 18.272-32.874 1.653-13.46-2.201-26.317-9.777-36.317" fill="#FF1921" opacity=".4" />
-        <g transform="rotate(7 -2256.059 4613.011)">
-          <path d="M50.763 51.739l.003-.006.005-.01.006.016-.002.005-.007.012-.005-.017zM88.5 85.825c.435-.485.86-.98 1.279-1.483l.125-.152c.41-.5.812-1.008 1.205-1.525l.144-.19a50.38 50.38 0 0 0 1.103-1.523c.063-.092.126-.186.19-.277.333-.487.66-.979.978-1.479.088-.136.173-.274.258-.412.286-.46.569-.925.844-1.398.11-.19.215-.382.323-.574.242-.43.482-.862.713-1.301.127-.242.248-.488.372-.733.156-.31.319-.613.47-.927.047-.098.087-.198.134-.296.127-.27.247-.543.37-.814.184-.41.371-.818.543-1.228.086-.204.163-.411.247-.615a51.01 51.01 0 0 0 2.35-7.327c.025-.108.054-.215.08-.325.108-.464.2-.932.295-1.398.054-.265.113-.528.163-.793.066-.36.12-.72.179-1.081.063-.382.129-.763.183-1.149.036-.256.063-.513.095-.768.062-.496.123-.992.17-1.49.013-.148.023-.297.035-.445.054-.61.098-1.22.13-1.834l.003-.102a50.968 50.968 0 0 0-2.79-19.265c-.12-.336-.25-.663-.374-.997l-.001-.003a50.875 50.875 0 0 0-6.339-11.931l-.08-.111a49.265 49.265 0 0 0-1.498-1.974c-.061-.079-.12-.156-.183-.233a50.926 50.926 0 0 0-1.67-1.97c-.015-.017-.03-.035-.045-.05a50.6 50.6 0 0 0-1.752-1.853c-.204-.206-.417-.404-.624-.607a57.506 57.506 0 0 0-1.142-1.08c-.292-.268-.59-.53-.889-.793-.317-.275-.639-.547-.964-.816-.332-.275-.665-.552-1.005-.821-.325-.254-.656-.497-.986-.745a51.5 51.5 0 0 0-1.84-1.313c-.34-.232-.68-.465-1.025-.688-.362-.233-.73-.459-1.098-.683-.18-.111-.357-.229-.54-.337a52.706 52.706 0 0 0-1.589-.9 51.34 51.34 0 0 0-2.061-1.07c-.06-.03-.119-.063-.18-.093l-.003.007c-.004-.003-.008-.004-.013-.006l.006-.01c-.203-.099-.41-.183-.614-.279a54.192 54.192 0 0 0-2.398-1.045 51.337 51.337 0 0 0-5.863-1.975c-.189-.052-.376-.113-.566-.162-.15-.039-.3-.069-.45-.106-.154-.038-.31-.072-.464-.108a51.185 51.185 0 0 0-2.633-.554 49.52 49.52 0 0 0-1.531-.256c-.315-.047-.63-.094-.945-.136a52.382 52.382 0 0 0-3.094-.308 55.11 55.11 0 0 0-1.339-.076c-.199-.008-.397-.018-.595-.023a47.568 47.568 0 0 0-1.81-.023c-.126 0-.252.004-.377.006-.578.006-1.156.024-1.735.05-.083.005-.167.008-.251.012-.652.035-1.305.083-1.958.143l-.285.028c-.624.061-1.248.133-1.873.216l-.12.017c-.638.087-1.276.19-1.913.304-.14.024-.28.051-.42.077-.525.098-1.052.204-1.577.319-.124.027-.246.05-.369.08-.561.126-1.123.267-1.683.414-.247.065-.494.134-.74.203a49.154 49.154 0 0 0-2.24.682c-.311.103-.623.204-.934.314h.002l-.012.005-.003-.005c-.058.02-.114.044-.174.064-.736.262-1.467.537-2.192.833-.008.002-.016.006-.025.01-.746.304-1.485.63-2.216.97-.156.072-.307.15-.462.224-.559.267-1.115.541-1.664.83-.203.104-.404.217-.606.325-.494.267-.983.54-1.467.824-.101.058-.206.112-.307.172-.123.071-.239.15-.361.224a52.575 52.575 0 0 0-1.847 1.171 48.414 48.414 0 0 0-1.77 1.232 49.913 49.913 0 0 0-2.927 2.303 48.754 48.754 0 0 0-1.684 1.494l-.292.272c-.444.419-.882.848-1.313 1.285l-.2.202a49.85 49.85 0 0 0-1.316 1.403c-.047.051-.094.102-.14.155-.436.487-.863.982-1.281 1.488l-.122.146A49.825 49.825 0 0 0 10.29 21a48.387 48.387 0 0 0-1.295 1.8c-.334.487-.66.98-.979 1.48l-.255.41c-.289.463-.571.928-.846 1.4-.109.191-.215.382-.323.574-.242.43-.482.863-.714 1.303-.126.24-.246.485-.369.729-.157.31-.321.614-.472.93-.047.098-.088.198-.135.296-.126.27-.245.54-.367.812-.185.41-.372.82-.546 1.23-.084.203-.161.408-.243.612a51.112 51.112 0 0 0-2.352 7.33c-.026.108-.056.216-.081.324-.108.466-.2.932-.295 1.4-.054.263-.113.527-.162.792-.067.36-.122.721-.18 1.082-.063.382-.13.763-.183 1.147-.035.257-.063.514-.094.77-.064.496-.123.99-.171 1.49-.014.146-.023.295-.035.444a48.92 48.92 0 0 0-.134 1.936 50.939 50.939 0 0 0 2.793 19.264c.117.337.249.664.373.997l.001.003a50.87 50.87 0 0 0 7.915 14.017l.185.233c.542.67 1.097 1.33 1.67 1.97l.044.05a50.6 50.6 0 0 0 1.751 1.853c.205.205.418.404.625.606.377.365.755.729 1.142 1.081a50.468 50.468 0 0 0 1.853 1.608c.332.276.664.553 1.005.822.324.254.656.498.985.744.602.45 1.216.888 1.842 1.313.34.233.679.465 1.023.688.363.234.73.46 1.1.684.18.11.357.228.539.336.32.19.644.375.967.557.206.116.414.23.622.343.68.372 1.365.73 2.061 1.07.062.03.12.064.18.093l.004-.006.011.005-.005.01c.204.1.412.185.616.28a53.783 53.783 0 0 0 2.396 1.045 52.559 52.559 0 0 0 4.684 1.64c.391.116.785.229 1.179.335.188.052.377.113.565.161.151.04.302.07.451.107.154.038.309.072.464.108a50.679 50.679 0 0 0 2.633.554c.509.093 1.018.178 1.528.255.317.048.633.095.95.137a51.773 51.773 0 0 0 2.261.246 56.85 56.85 0 0 0 2.17.137c.199.009.396.018.596.023.602.02 1.205.027 1.81.023.124 0 .251-.004.376-.005a49.622 49.622 0 0 0 1.734-.051l.252-.011a53.529 53.529 0 0 0 1.96-.143c.093-.008.187-.02.28-.029.625-.06 1.251-.13 1.876-.216l.12-.015a52.17 52.17 0 0 0 1.914-.305c.14-.024.279-.051.418-.077a50.26 50.26 0 0 0 1.579-.319c.122-.027.246-.051.369-.08a50.002 50.002 0 0 0 1.681-.415c.248-.064.494-.133.74-.202a48.387 48.387 0 0 0 2.237-.681c.312-.103.625-.206.937-.314l-.001-.002.012-.004.003.006c.059-.021.115-.046.175-.067.732-.26 1.46-.534 2.18-.827l.037-.014a53.684 53.684 0 0 0 2.216-.97c.155-.073.306-.151.46-.224a54.502 54.502 0 0 0 1.665-.828c.203-.107.405-.22.607-.329.493-.266.981-.539 1.465-.821.102-.06.207-.112.308-.172.123-.073.241-.152.362-.227a49.368 49.368 0 0 0 4.795-3.288c.173-.135.348-.27.52-.408.414-.329.822-.665 1.226-1.007.136-.116.273-.23.408-.348.432-.375.856-.757 1.276-1.146l.295-.275c.443-.417.88-.845 1.31-1.28.067-.07.136-.138.203-.207.446-.457.884-.924 1.313-1.4l.143-.158z" fill="#FF1921" />
-        </g>
-        <path d="M669.311 333.363a49.417 49.417 0 0 0-24.339-10.02 49.423 49.423 0 0 0-24.505 3.183 50.676 50.676 0 0 1 18.598 29.492c5.224 24.127-7.538 47.985-29.266 57.508a49.414 49.414 0 0 0 22.994 9.01 49.418 49.418 0 0 0 26.04-3.837c15.566-6.86 27.241-21.537 29.47-39.692 2.23-18.156-5.549-35.22-18.992-45.644" fill="#FF1921" opacity=".8" />
-        <g fill="#F7691C" transform="scale(-1 1) rotate(-3 6462.427 3155.055)">
-          <path d="M161.027 116.498c10.747-9.31 17.548-23.067 17.548-38.416 0-26.352-20.045-48.015-45.696-50.526 8.886-7.697 20.47-12.354 33.14-12.354 27.994 0 50.688 22.73 50.688 50.77 0 28.039-22.694 50.77-50.688 50.77-1.684 0-3.35-.083-4.992-.244z" opacity=".399" />
-          <path d="M46.179 101.712c1.184.082 2.379.124 3.583.124 27.994 0 50.688-22.73 50.688-50.769 0-11.29-3.68-21.72-9.902-30.15 26.322 1.841 47.104 23.812 47.104 50.644 0 28.04-22.693 50.77-50.688 50.77-16.722 0-31.553-8.111-40.785-20.62z" opacity=".8" />
-          <ellipse cx="50.755" cy="51.394" rx="50.688" ry="50.769" />
-          <path d="M100.164 120.592c21.594-5.818 37.488-25.565 37.488-49.03 0-17.791-9.136-33.445-22.965-42.51a50.677 50.677 0 0 1 13.2-1.74c27.994 0 50.688 22.73 50.688 50.77 0 28.04-22.694 50.77-50.688 50.77a50.393 50.393 0 0 1-27.723-8.26z" opacity=".604" />
-        </g>
-        <path d="M839.701 144.582c6.678 26.783-9.51 53.881-36.155 60.525-26.646 6.643-53.66-9.684-60.338-36.467-6.678-26.783 9.51-53.881 36.155-60.525 26.646-6.643 53.66 9.684 60.338 36.467" fill="#CE147C" opacity=".1" />
-        <path d="M674.05 134.373c-13.551 3.379-24.388 12.055-30.897 23.284 10.645 6.626 18.842 17.252 22.124 30.414 3.281 13.161 1.032 26.392-5.256 37.24 11.019 6.86 24.66 9.432 38.212 6.053 26.646-6.643 42.833-33.742 36.156-60.525-6.678-26.783-33.693-43.11-60.338-36.466" fill="#CE147C" opacity=".2" />
-        <path d="M576.42 175.672c-10.657-6.908-23.919-9.761-37.291-6.919a49.289 49.289 0 0 0-23.338 12.098c16.085 5.486 28.982 19.093 32.787 36.99 3.804 17.898-2.444 35.575-14.908 47.128a49.289 49.289 0 0 0 26.242 1.56c13.372-2.842 24.327-10.842 31.253-21.488 6.988-10.743 9.875-24.178 6.993-37.738-2.883-13.56-10.985-24.66-21.738-31.63" fill="#CE147C" opacity=".6" />
-        <path d="M643.152 157.657c-11.019-6.86-24.66-9.432-38.212-6.053-13.264 3.307-23.933 11.685-30.484 22.566 10.99 6.592 19.476 17.402 22.83 30.853 3.353 13.451.935 26.98-5.673 37.959 10.893 6.532 24.246 8.92 37.51 5.613 13.551-3.379 24.388-12.054 30.898-23.284 6.288-10.848 8.538-24.08 5.257-37.24-3.282-13.162-11.48-23.788-22.126-30.414" fill="#CE147C" opacity=".4" />
-        <path d="M461.076 237.401l.001-.006.003-.012.009.015v.005l-.004.014-.01-.016zm44.812 23.965c.304-.576.597-1.159.88-1.749l.085-.177c.277-.585.543-1.175.799-1.772l.093-.22a50.39 50.39 0 0 0 .7-1.744l.118-.315a52.874 52.874 0 0 0 .74-2.132c.166-.517.327-1.036.479-1.562.06-.211.116-.423.174-.635.13-.475.258-.953.377-1.435.064-.265.122-.533.182-.8.077-.339.161-.673.231-1.014.022-.107.037-.214.06-.32.057-.292.107-.586.16-.88.08-.441.162-.882.23-1.323.034-.218.058-.438.09-.656.363-2.576.53-5.142.503-7.678v-.335c-.008-.477-.032-.953-.053-1.428-.012-.27-.018-.54-.034-.81-.023-.364-.057-.727-.088-1.091-.031-.387-.06-.773-.1-1.16-.027-.257-.063-.513-.094-.768a50.02 50.02 0 0 0-.196-1.488c-.023-.146-.05-.293-.074-.44-.095-.604-.2-1.208-.318-1.81-.006-.035-.014-.068-.021-.101a51.055 51.055 0 0 0-7.366-18.022c-.196-.298-.402-.584-.603-.878l-.002-.002a50.913 50.913 0 0 0-9.029-10.049l-.104-.088a49.269 49.269 0 0 0-1.93-1.554c-.078-.06-.154-.121-.233-.181a50.917 50.917 0 0 0-2.096-1.51l-.055-.037a50.578 50.578 0 0 0-2.145-1.375c-.248-.15-.503-.291-.753-.438a57.466 57.466 0 0 0-1.367-.773c-.348-.19-.7-.372-1.053-.554-.375-.19-.752-.376-1.132-.559-.388-.187-.778-.376-1.173-.554-.376-.168-.755-.324-1.135-.485a51.43 51.43 0 0 0-2.101-.83c-.386-.143-.771-.287-1.16-.42-.407-.139-.817-.27-1.23-.398-.2-.064-.4-.135-.604-.196a52.612 52.612 0 0 0-1.757-.49 51.24 51.24 0 0 0-2.256-.54c-.066-.014-.13-.033-.197-.047l-.001.007-.013-.002.002-.012c-.22-.046-.442-.077-.662-.121a54.075 54.075 0 0 0-2.577-.436 51.213 51.213 0 0 0-6.16-.5c-.194-.005-.39-.019-.586-.02-.154-.002-.307.006-.463.005-.157 0-.316.005-.475.007a51.044 51.044 0 0 0-2.685.099c-.517.033-1.031.073-1.545.122-.317.03-.634.06-.95.096a52.23 52.23 0 0 0-3.072.448c-.439.078-.877.162-1.315.25-.195.04-.389.079-.582.122-.589.127-1.175.265-1.76.415-.121.03-.243.064-.364.096-.558.146-1.113.303-1.668.469l-.241.071c-.623.192-1.244.397-1.862.612l-.27.097c-.59.21-1.177.43-1.762.662l-.113.044c-.597.24-1.19.494-1.78.758-.13.057-.259.117-.388.176-.485.223-.97.452-1.45.69-.114.056-.227.11-.339.167a49.94 49.94 0 0 0-3.114 1.72c-.366.217-.73.44-1.09.668-.277.175-.554.348-.829.53h.002l-.01.007-.005-.004-.152.105c-.65.432-1.292.875-1.923 1.337l-.022.015c-.649.477-1.286.972-1.913 1.477-.133.109-.26.221-.393.33-.477.394-.949.795-1.412 1.207-.171.151-.338.308-.508.462a50.27 50.27 0 0 0-1.222 1.155c-.084.08-.173.158-.256.24-.102.1-.195.205-.296.305a52.49 52.49 0 0 0-1.506 1.583 48.344 48.344 0 0 0-1.416 1.624 49.85 49.85 0 0 0-2.28 2.941l-.309.435c-.328.468-.648.942-.96 1.422l-.218.334c-.329.514-.649 1.037-.96 1.565-.05.081-.097.163-.146.245a49.829 49.829 0 0 0-.935 1.679 50.186 50.186 0 0 0-.98 1.938l-.083.17a49.824 49.824 0 0 0-.892 1.992 48.396 48.396 0 0 0-.819 2.06c-.206.553-.403 1.111-.59 1.673l-.149.46a52.122 52.122 0 0 0-.48 1.563c-.06.211-.117.422-.175.634-.13.477-.258.955-.376 1.438-.064.264-.12.53-.18.796-.078.34-.164.674-.234 1.016-.022.107-.036.214-.058.321-.058.292-.107.584-.16.877-.08.442-.162.885-.23 1.325-.034.217-.059.435-.089.653a51.182 51.182 0 0 0-.505 7.682v.334c.008.479.031.954.053 1.43.012.269.018.539.035.807.022.367.056.73.087 1.094.031.387.06.773.1 1.158.028.258.063.514.095.77.059.497.121.99.195 1.486.023.147.05.293.074.44a49.01 49.01 0 0 0 .339 1.912 51.026 51.026 0 0 0 7.367 18.021c.195.299.402.584.603.878a50.91 50.91 0 0 0 11.063 11.693l.235.183c.688.52 1.385 1.025 2.095 1.508l.055.038c.705.478 1.42.936 2.145 1.375.248.15.502.291.752.438.454.263.908.524 1.368.773.348.189.7.371 1.053.554.375.192.752.376 1.131.559.39.187.778.375 1.174.554.375.168.756.325 1.135.484.692.292 1.392.568 2.102.83.385.144.77.287 1.158.42.407.14.818.27 1.23.398.202.064.402.135.604.196.357.108.715.208 1.072.307.228.063.457.123.686.183.748.196 1.499.378 2.256.54.067.014.131.033.197.047l.002-.007.012.003-.002.011c.222.047.443.08.665.123a53.666 53.666 0 0 0 2.573.435 52.432 52.432 0 0 0 4.935.46c.409.018.817.032 1.225.04.195.005.392.018.586.02.156.001.31-.006.463-.006.158 0 .316-.005.475-.007a50.539 50.539 0 0 0 2.686-.098c.516-.033 1.03-.073 1.542-.121a51.625 51.625 0 0 0 3.204-.405 56.695 56.695 0 0 0 2.137-.39c.193-.04.387-.08.582-.122a47.322 47.322 0 0 0 1.759-.415c.121-.03.243-.065.364-.096a49.476 49.476 0 0 0 1.668-.47l.24-.07a53.37 53.37 0 0 0 1.866-.613c.088-.03.176-.064.264-.095.59-.21 1.18-.43 1.765-.663l.113-.044c.596-.24 1.19-.494 1.78-.758.13-.058.258-.118.387-.176.486-.222.97-.452 1.453-.691l.338-.166c.513-.26 1.022-.531 1.529-.81a48.264 48.264 0 0 0 2.67-1.577c.278-.174.556-.35.833-.53l-.002-.002.01-.007.005.006c.052-.035.1-.072.153-.108a50.434 50.434 0 0 0 1.945-1.351c.649-.477 1.285-.97 1.913-1.477.132-.108.26-.22.392-.329.476-.394.949-.795 1.412-1.206.17-.152.34-.31.509-.465.413-.377.82-.76 1.22-1.151.085-.083.174-.16.257-.241.101-.1.197-.207.296-.308a49.287 49.287 0 0 0 2.924-3.207c.314-.376.622-.757.926-1.143.136-.173.273-.346.406-.522.321-.418.635-.843.944-1.273.103-.145.209-.29.31-.437.329-.467.647-.94.96-1.42l.22-.337c.328-.513.648-1.034.959-1.56l.147-.25c.321-.55.632-1.11.933-1.675l.1-.188z" fill="#CE147C" />
-        <path d="M513.217 180.248a49.289 49.289 0 0 0-26.287.275 49.292 49.292 0 0 0-21.28 12.497c11.168 2.936 21.417 9.685 28.615 19.895 14.226 20.176 11.815 47.126-4.437 64.377a49.284 49.284 0 0 0 24.655-.681 49.29 49.29 0 0 0 22.438-13.698c11.627-12.394 16.626-30.465 11.583-48.052-5.043-17.588-18.858-30.263-35.287-34.613" fill="#CE147C" opacity=".8" />
-        <path d="M624.615-39.182c-12.506-2.382-25.922-.02-37.315 7.664a49.764 49.764 0 0 0-17.219 20.03c17.03-.986 34.122 6.763 44.348 21.923 10.226 15.16 11.007 33.911 3.713 49.331a49.764 49.764 0 0 0 25.023-8.462c11.392-7.684 18.608-19.239 21.083-31.726 2.498-12.6.17-26.149-7.578-37.635-7.747-11.486-19.437-18.72-32.055-21.125" fill="#00B2FF" opacity=".6" />
-        <path d="M683.825-78.402c-12.736-2.645-26.488-.357-38.127 7.494-11.392 7.684-18.607 19.238-21.084 31.726 12.618 2.405 24.31 9.638 32.057 21.124 7.748 11.486 10.074 25.036 7.577 37.636 12.506 2.382 25.923.02 37.315-7.664 11.639-7.85 18.912-19.744 21.23-32.544 2.24-12.365-.144-25.578-7.725-36.817-7.58-11.24-18.938-18.399-31.243-20.955" fill="#00B2FF" opacity=".4" />
-        <g transform="rotate(-34 315.164 -726.927)">
-          <path d="M50.958 51.72l.003-.005.005-.01.005.015v.006l-.008.012-.005-.018zm37.88 34.075c.437-.485.864-.98 1.284-1.483l.126-.152c.412-.5.815-1.007 1.21-1.525l.144-.19a50.37 50.37 0 0 0 1.107-1.522c.064-.092.127-.185.192-.277a52.818 52.818 0 0 0 1.24-1.89c.287-.46.571-.925.847-1.398.11-.19.216-.382.324-.574.243-.43.484-.861.716-1.3.127-.242.25-.488.373-.732.157-.31.32-.614.472-.927.047-.098.088-.198.135-.296.127-.27.248-.543.37-.814.186-.41.374-.817.547-1.228.086-.204.163-.41.247-.614a50.84 50.84 0 0 0 2.358-7.325c.026-.108.056-.215.081-.324.109-.465.202-.932.296-1.398.054-.265.114-.528.164-.793.066-.36.122-.72.18-1.08.063-.383.13-.763.184-1.148.036-.257.063-.514.095-.769.063-.496.123-.991.17-1.49.014-.147.024-.296.036-.444.054-.61.099-1.22.13-1.833l.004-.103a50.753 50.753 0 0 0-2.802-19.257c-.119-.336-.25-.663-.374-.997-.002 0-.002-.001-.002-.003a50.786 50.786 0 0 0-6.363-11.927l-.08-.11a49.268 49.268 0 0 0-1.504-1.974c-.062-.079-.121-.156-.183-.233a50.963 50.963 0 0 0-1.678-1.97l-.044-.05a50.67 50.67 0 0 0-1.759-1.852c-.205-.205-.419-.404-.627-.606a57.627 57.627 0 0 0-1.146-1.08 55.6 55.6 0 0 0-.892-.792 56.098 56.098 0 0 0-.968-.816c-.333-.276-.667-.553-1.01-.821-.325-.254-.657-.497-.988-.744-.605-.45-1.22-.888-1.848-1.313-.342-.233-.682-.465-1.028-.688a48.39 48.39 0 0 0-1.103-.683c-.181-.111-.358-.228-.542-.337a52.965 52.965 0 0 0-1.595-.9c-.683-.37-1.37-.728-2.07-1.069-.06-.03-.118-.063-.18-.093l-.003.007-.012-.006.005-.01c-.204-.099-.412-.183-.616-.279a54.512 54.512 0 0 0-2.407-1.045 51.673 51.673 0 0 0-5.886-1.975c-.19-.051-.377-.112-.568-.16-.15-.04-.301-.07-.452-.107-.155-.038-.31-.072-.465-.108a51.567 51.567 0 0 0-2.644-.554 49.9 49.9 0 0 0-1.537-.255c-.316-.048-.632-.095-.948-.137a52.794 52.794 0 0 0-3.106-.308A55.551 55.551 0 0 0 53.1.84c-.2-.008-.398-.018-.597-.023a47.951 47.951 0 0 0-1.818-.023c-.125 0-.252.004-.378.005-.58.007-1.16.025-1.74.052l-.254.01a53.59 53.59 0 0 0-1.965.144l-.286.028a52.69 52.69 0 0 0-1.88.216l-.12.016c-.641.088-1.282.19-1.921.304-.14.024-.281.051-.422.077-.527.099-1.055.204-1.582.319-.125.027-.248.051-.37.08a50.442 50.442 0 0 0-3.457.914c-.409.123-.819.251-1.225.385-.313.103-.625.204-.937.313l.001.002c-.004 0-.008.002-.012.004l-.002-.006c-.059.02-.115.045-.175.065-.74.262-1.473.536-2.2.832l-.026.01c-.749.305-1.49.63-2.224.97-.157.072-.309.15-.464.223-.56.268-1.119.542-1.67.83-.205.105-.406.217-.609.325-.496.268-.986.54-1.473.824-.101.058-.207.112-.308.172-.123.071-.24.151-.362.224A52.813 52.813 0 0 0 22.77 9.3a48.608 48.608 0 0 0-1.776 1.233 50.087 50.087 0 0 0-2.939 2.301l-.408.347c-.434.374-.86.757-1.282 1.147l-.294.272c-.446.418-.885.848-1.317 1.284l-.202.203c-.448.457-.888.925-1.32 1.402a50.238 50.238 0 0 0-1.427 1.642l-.122.146a49.84 49.84 0 0 0-1.354 1.715c-.379.5-.749 1.01-1.11 1.526l-.19.275c-.336.486-.664.979-.983 1.479l-.257.41c-.289.462-.573.927-.848 1.4-.11.19-.217.38-.325.572-.243.43-.484.863-.716 1.303-.127.24-.247.485-.37.728-.159.311-.323.615-.475.93-.047.098-.088.198-.135.297-.127.269-.246.54-.369.812-.185.409-.373.818-.547 1.229-.085.202-.162.408-.245.612A50.94 50.94 0 0 0 1.4 39.893l-.081.324c-.108.466-.202.932-.296 1.4-.054.263-.114.526-.162.79-.068.361-.123.722-.181 1.083-.064.382-.13.763-.184 1.146-.035.257-.064.514-.095.77-.063.496-.123.99-.171 1.489-.014.147-.023.295-.036.444a48.702 48.702 0 0 0-.133 1.935 50.725 50.725 0 0 0 2.803 19.258c.117.336.25.663.374.997l.001.002a50.782 50.782 0 0 0 7.946 14.012l.185.234a53.2 53.2 0 0 0 1.676 1.969l.045.05a50.67 50.67 0 0 0 1.758 1.852c.205.205.419.404.627.606.378.365.758.728 1.146 1.08a50.609 50.609 0 0 0 1.86 1.607c.334.276.667.553 1.01.822.325.254.658.499.989.744.604.45 1.22.888 1.849 1.313.34.232.68.465 1.027.688.363.233.732.459 1.103.683.18.11.36.228.542.336.321.19.646.375.97.557.207.116.416.23.624.343.683.371 1.37.73 2.07 1.07.062.03.12.063.18.093l.005-.007a.027.027 0 0 0 .01.005l-.005.011c.206.099.414.184.62.28a54.101 54.101 0 0 0 2.403 1.044 52.903 52.903 0 0 0 4.702 1.64c.394.116.788.228 1.184.335.19.05.379.112.568.16.151.04.303.07.453.107.154.038.31.072.465.108a51.057 51.057 0 0 0 2.643.554 49.73 49.73 0 0 0 2.488.391 52.19 52.19 0 0 0 2.27.246 57.326 57.326 0 0 0 2.178.138c.199.008.398.018.598.023.604.019 1.21.027 1.816.023.126 0 .253-.004.378-.005.58-.007 1.16-.025 1.741-.052l.253-.01a53.957 53.957 0 0 0 1.968-.144c.093-.008.188-.019.28-.028.628-.06 1.256-.131 1.884-.216l.12-.015c.64-.09 1.281-.192 1.922-.305l.419-.077a50.637 50.637 0 0 0 1.585-.319c.123-.027.247-.051.37-.08a50.367 50.367 0 0 0 3.456-.914 55.5 55.5 0 0 0 1.22-.384c.314-.102.628-.205.941-.313l-.001-.003.012-.004.003.007c.06-.022.116-.046.176-.068a50.88 50.88 0 0 0 2.225-.84c.75-.305 1.49-.63 2.225-.97.155-.072.307-.15.462-.224.561-.267 1.12-.541 1.67-.828.205-.106.408-.218.61-.328.495-.266.986-.539 1.47-.821.103-.06.209-.112.31-.172.123-.072.242-.152.364-.226a49.591 49.591 0 0 0 4.813-3.288c.174-.135.35-.27.523-.408.415-.328.824-.664 1.23-1.006.136-.116.274-.231.409-.349.434-.374.86-.756 1.281-1.145l.296-.274c.445-.417.884-.846 1.315-1.28l.204-.207a49.72 49.72 0 0 0 1.461-1.557z" fill="#00B2FF" />
-        </g>
-        <path d="M570.082-11.487a49.764 49.764 0 0 0-25.023 8.461 49.776 49.776 0 0 0-16.483 18.529c11.57-.691 23.435 2.533 33.45 10 19.792 14.759 25.776 41.15 15.558 62.633a49.767 49.767 0 0 0 23.34-8.34 49.765 49.765 0 0 0 17.218-20.031c7.294-15.419 6.512-34.17-3.713-49.33-10.225-15.16-27.318-22.909-44.347-21.922" fill="#00B2FF" opacity=".8" />
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgCirclesHeroMobile() {
-
-  return (
-    <svg
-      className="svg-circles-hero-mobile"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 50 375 250"
-      preserveAspectRatio="xMinYMid"
-    >
-      <g fill="none" fillRule="evenodd">
-        <path d="M411.614 332.21c-3.946-12.08-12.473-22.65-24.67-28.864a49.417 49.417 0 0 0-25.782-5.303c9.067 14.393 10.54 33.088 2.236 49.387-8.305 16.299-24.295 26.095-41.269 27.22a49.417 49.417 0 0 0 19.444 17.74c12.198 6.216 25.76 6.901 37.853 2.994 12.2-3.943 22.905-12.561 29.197-24.91 6.293-12.35 6.973-26.076 2.991-38.264" fill="#00B2FF" opacity=".6" />
-        <path d="M474.38 364.86c-3.828-12.41-12.451-23.308-24.913-29.657-12.196-6.215-25.76-6.9-37.853-2.993 3.982 12.188 3.303 25.914-2.99 38.264-6.292 12.349-16.997 20.966-29.199 24.91 3.948 12.08 12.475 22.65 24.671 28.864 12.462 6.35 26.347 6.92 38.637 2.724 11.873-4.055 22.257-12.558 28.414-24.642 6.157-12.083 6.932-25.482 3.234-37.47" fill="#FF1921" opacity=".4" />
-        <g transform="rotate(27 -358.94 643.266)">
-          <path d="M50.763 51.739l.003-.006.005-.01.006.016-.002.005-.007.012-.005-.017zM88.5 85.825c.435-.485.86-.98 1.279-1.483l.125-.152c.41-.5.812-1.008 1.205-1.525l.144-.19a50.38 50.38 0 0 0 1.103-1.523c.063-.092.126-.186.19-.277.333-.487.66-.979.978-1.479.088-.136.173-.274.258-.412.286-.46.569-.925.844-1.398.11-.19.215-.382.323-.574.242-.43.482-.862.713-1.301.127-.242.248-.488.372-.733.156-.31.319-.613.47-.927.047-.098.087-.198.134-.296.127-.27.247-.543.37-.814.184-.41.371-.818.543-1.228.086-.204.163-.411.247-.615a51.01 51.01 0 0 0 2.35-7.327c.025-.108.054-.215.08-.325.108-.464.2-.932.295-1.398.054-.265.113-.528.163-.793.066-.36.12-.72.179-1.081.063-.382.129-.763.183-1.149.036-.256.063-.513.095-.768.062-.496.123-.992.17-1.49.013-.148.023-.297.035-.445.054-.61.098-1.22.13-1.834l.003-.102a50.968 50.968 0 0 0-2.79-19.265c-.12-.336-.25-.663-.374-.997l-.001-.003a50.875 50.875 0 0 0-6.339-11.931l-.08-.111a49.265 49.265 0 0 0-1.498-1.974c-.061-.079-.12-.156-.183-.233a50.926 50.926 0 0 0-1.67-1.97c-.015-.017-.03-.035-.045-.05a50.6 50.6 0 0 0-1.752-1.853c-.204-.206-.417-.404-.624-.607a57.506 57.506 0 0 0-1.142-1.08c-.292-.268-.59-.53-.889-.793-.317-.275-.639-.547-.964-.816-.332-.275-.665-.552-1.005-.821-.325-.254-.656-.497-.986-.745a51.5 51.5 0 0 0-1.84-1.313c-.34-.232-.68-.465-1.025-.688-.362-.233-.73-.459-1.098-.683-.18-.111-.357-.229-.54-.337a52.706 52.706 0 0 0-1.589-.9 51.34 51.34 0 0 0-2.061-1.07c-.06-.03-.119-.063-.18-.093l-.003.007c-.004-.003-.008-.004-.013-.006l.006-.01c-.203-.099-.41-.183-.614-.279a54.192 54.192 0 0 0-2.398-1.045 51.337 51.337 0 0 0-5.863-1.975c-.189-.052-.376-.113-.566-.162-.15-.039-.3-.069-.45-.106-.154-.038-.31-.072-.464-.108a51.185 51.185 0 0 0-2.633-.554 49.52 49.52 0 0 0-1.531-.256c-.315-.047-.63-.094-.945-.136a52.382 52.382 0 0 0-3.094-.308 55.11 55.11 0 0 0-1.339-.076c-.199-.008-.397-.018-.595-.023a47.568 47.568 0 0 0-1.81-.023c-.126 0-.252.004-.377.006-.578.006-1.156.024-1.735.05-.083.005-.167.008-.251.012-.652.035-1.305.083-1.958.143l-.285.028c-.624.061-1.248.133-1.873.216l-.12.017c-.638.087-1.276.19-1.913.304-.14.024-.28.051-.42.077-.525.098-1.052.204-1.577.319-.124.027-.246.05-.369.08-.561.126-1.123.267-1.683.414-.247.065-.494.134-.74.203a49.154 49.154 0 0 0-2.24.682c-.311.103-.623.204-.934.314h.002l-.012.005-.003-.005c-.058.02-.114.044-.174.064-.736.262-1.467.537-2.192.833-.008.002-.016.006-.025.01-.746.304-1.485.63-2.216.97-.156.072-.307.15-.462.224-.559.267-1.115.541-1.664.83-.203.104-.404.217-.606.325-.494.267-.983.54-1.467.824-.101.058-.206.112-.307.172-.123.071-.239.15-.361.224a52.575 52.575 0 0 0-1.847 1.171 48.414 48.414 0 0 0-1.77 1.232 49.913 49.913 0 0 0-2.927 2.303 48.754 48.754 0 0 0-1.684 1.494l-.292.272c-.444.419-.882.848-1.313 1.285l-.2.202a49.85 49.85 0 0 0-1.316 1.403c-.047.051-.094.102-.14.155-.436.487-.863.982-1.281 1.488l-.122.146A49.825 49.825 0 0 0 10.29 21a48.387 48.387 0 0 0-1.295 1.8c-.334.487-.66.98-.979 1.48l-.255.41c-.289.463-.571.928-.846 1.4-.109.191-.215.382-.323.574-.242.43-.482.863-.714 1.303-.126.24-.246.485-.369.729-.157.31-.321.614-.472.93-.047.098-.088.198-.135.296-.126.27-.245.54-.367.812-.185.41-.372.82-.546 1.23-.084.203-.161.408-.243.612a51.112 51.112 0 0 0-2.352 7.33c-.026.108-.056.216-.081.324-.108.466-.2.932-.295 1.4-.054.263-.113.527-.162.792-.067.36-.122.721-.18 1.082-.063.382-.13.763-.183 1.147-.035.257-.063.514-.094.77-.064.496-.123.99-.171 1.49-.014.146-.023.295-.035.444a48.92 48.92 0 0 0-.134 1.936 50.939 50.939 0 0 0 2.793 19.264c.117.337.249.664.373.997l.001.003a50.87 50.87 0 0 0 7.915 14.017l.185.233c.542.67 1.097 1.33 1.67 1.97l.044.05a50.6 50.6 0 0 0 1.751 1.853c.205.205.418.404.625.606.377.365.755.729 1.142 1.081a50.468 50.468 0 0 0 1.853 1.608c.332.276.664.553 1.005.822.324.254.656.498.985.744.602.45 1.216.888 1.842 1.313.34.233.679.465 1.023.688.363.234.73.46 1.1.684.18.11.357.228.539.336.32.19.644.375.967.557.206.116.414.23.622.343.68.372 1.365.73 2.061 1.07.062.03.12.064.18.093l.004-.006.011.005-.005.01c.204.1.412.185.616.28a53.783 53.783 0 0 0 2.396 1.045 52.559 52.559 0 0 0 4.684 1.64c.391.116.785.229 1.179.335.188.052.377.113.565.161.151.04.302.07.451.107.154.038.309.072.464.108a50.679 50.679 0 0 0 2.633.554c.509.093 1.018.178 1.528.255.317.048.633.095.95.137a51.773 51.773 0 0 0 2.261.246 56.85 56.85 0 0 0 2.17.137c.199.009.396.018.596.023.602.02 1.205.027 1.81.023.124 0 .251-.004.376-.005a49.622 49.622 0 0 0 1.734-.051l.252-.011a53.529 53.529 0 0 0 1.96-.143c.093-.008.187-.02.28-.029.625-.06 1.251-.13 1.876-.216l.12-.015a52.17 52.17 0 0 0 1.914-.305c.14-.024.279-.051.418-.077a50.26 50.26 0 0 0 1.579-.319c.122-.027.246-.051.369-.08a50.002 50.002 0 0 0 1.681-.415c.248-.064.494-.133.74-.202a48.387 48.387 0 0 0 2.237-.681c.312-.103.625-.206.937-.314l-.001-.002.012-.004.003.006c.059-.021.115-.046.175-.067.732-.26 1.46-.534 2.18-.827l.037-.014a53.684 53.684 0 0 0 2.216-.97c.155-.073.306-.151.46-.224a54.502 54.502 0 0 0 1.665-.828c.203-.107.405-.22.607-.329.493-.266.981-.539 1.465-.821.102-.06.207-.112.308-.172.123-.073.241-.152.362-.227a49.368 49.368 0 0 0 4.795-3.288c.173-.135.348-.27.52-.408.414-.329.822-.665 1.226-1.007.136-.116.273-.23.408-.348.432-.375.856-.757 1.276-1.146l.295-.275c.443-.417.88-.845 1.31-1.28.067-.07.136-.138.203-.207.446-.457.884-.924 1.313-1.4l.143-.158z" fill="#00B2FF" />
-        </g>
-        <path d="M361.162 298.043a49.417 49.417 0 0 0-19.444-17.74 49.423 49.423 0 0 0-24.115-5.39 50.676 50.676 0 0 1 7.39 34.074c-3.344 24.459-23.497 42.512-47.171 44.03a49.414 49.414 0 0 0 18.526 16.33 49.418 49.418 0 0 0 25.782 5.302c16.973-1.123 32.964-10.92 41.268-27.219 8.305-16.299 6.832-34.993-2.236-49.387" fill="#00B2FF" opacity=".8" />
-        <g fill="#F7691C" transform="scale(-1 1) rotate(12 -1307.849 -562.485)">
-          <path d="M161.027 116.498c10.747-9.31 17.548-23.067 17.548-38.416 0-26.352-20.045-48.015-45.696-50.526 8.886-7.697 20.47-12.354 33.14-12.354 27.994 0 50.688 22.73 50.688 50.77 0 28.039-22.694 50.77-50.688 50.77-1.684 0-3.35-.083-4.992-.244z" opacity=".399" />
-          <path d="M46.179 101.712c1.184.082 2.379.124 3.583.124 27.994 0 50.688-22.73 50.688-50.769 0-11.29-3.68-21.72-9.902-30.15 26.322 1.841 47.104 23.812 47.104 50.644 0 28.04-22.693 50.77-50.688 50.77-16.722 0-31.553-8.111-40.785-20.62z" opacity=".8" />
-          <ellipse cx="50.755" cy="51.394" rx="50.688" ry="50.769" />
-          <path d="M100.164 120.592c21.594-5.818 37.488-25.565 37.488-49.03 0-17.791-9.136-33.445-22.965-42.51a50.677 50.677 0 0 1 13.2-1.74c27.994 0 50.688 22.73 50.688 50.77 0 28.04-22.694 50.77-50.688 50.77a50.393 50.393 0 0 1-27.723-8.26z" opacity=".604" />
-        </g>
-        <path d="M393.234 63.858c-12.669.896-24.977 6.6-33.946 16.917a49.289 49.289 0 0 0-11.358 23.707c16.148-5.3 34.637-2.194 48.446 9.81 13.81 12.004 19.458 29.882 16.457 46.61a49.289 49.289 0 0 0 21.896-14.547c8.97-10.318 12.904-23.3 12.028-35.97-.884-12.785-6.664-25.252-17.127-34.347-10.462-9.095-23.613-13.084-36.396-12.18" fill="#CE147C" opacity=".6" />
-        <g transform="rotate(-23 471.104 -568.397)">
-          <path d="M50.693 51.75l.003-.005.005-.01.006.016-.002.005-.006.012-.006-.017zm37.684 34.095c.434-.486.859-.98 1.277-1.484l.125-.152c.41-.5.81-1.008 1.203-1.525l.144-.191a50.39 50.39 0 0 0 1.101-1.523l.191-.277c.332-.487.66-.979.976-1.479.088-.136.173-.274.257-.412.287-.46.569-.925.843-1.398.11-.191.215-.383.323-.575.242-.43.481-.862.713-1.301.126-.242.247-.488.37-.733.157-.31.32-.613.47-.927.047-.099.087-.199.134-.296.127-.27.246-.543.369-.815.184-.41.37-.817.543-1.228.086-.204.163-.411.246-.615a51.081 51.081 0 0 0 2.346-7.329c.026-.108.055-.215.08-.324.108-.465.201-.933.295-1.399.054-.265.113-.528.163-.793.066-.36.12-.72.179-1.081.063-.383.129-.764.183-1.15.036-.256.063-.513.095-.768.062-.496.122-.992.17-1.49.013-.148.022-.297.034-.445.054-.61.098-1.22.13-1.834l.003-.103a51.055 51.055 0 0 0-2.787-19.269c-.118-.336-.248-.663-.372-.997l-.001-.003a50.913 50.913 0 0 0-6.33-11.934l-.08-.111a49.269 49.269 0 0 0-1.496-1.975l-.183-.232a50.917 50.917 0 0 0-1.669-1.972l-.044-.05a50.578 50.578 0 0 0-1.75-1.853c-.203-.205-.416-.404-.623-.606a57.466 57.466 0 0 0-1.14-1.082c-.292-.267-.589-.53-.887-.792a55.888 55.888 0 0 0-.963-.816c-.332-.276-.664-.553-1.004-.821-.324-.255-.655-.498-.985-.745a51.43 51.43 0 0 0-1.838-1.314c-.34-.232-.679-.465-1.023-.688-.361-.233-.728-.46-1.097-.683-.18-.111-.356-.229-.539-.337a52.612 52.612 0 0 0-1.586-.9 51.24 51.24 0 0 0-2.059-1.07c-.06-.03-.118-.064-.18-.094l-.003.007c-.004-.003-.008-.004-.012-.005l.006-.011c-.203-.099-.41-.182-.614-.278a54.075 54.075 0 0 0-2.394-1.046 51.213 51.213 0 0 0-5.855-1.976c-.189-.052-.376-.112-.565-.161-.15-.04-.3-.069-.45-.107-.154-.038-.308-.071-.463-.108a51.044 51.044 0 0 0-2.63-.554 49.378 49.378 0 0 0-1.529-.255c-.314-.048-.629-.095-.943-.137a52.23 52.23 0 0 0-3.09-.308c-.445-.031-.89-.055-1.336-.076-.2-.008-.397-.017-.595-.023a47.427 47.427 0 0 0-1.808-.023c-.125 0-.252.004-.377.006-.576.006-1.153.024-1.731.051l-.252.01c-.65.036-1.303.085-1.955.144l-.285.028a52.125 52.125 0 0 0-1.99.233c-.637.088-1.274.19-1.91.304-.14.024-.28.051-.42.077a54.12 54.12 0 0 0-1.574.319c-.124.027-.246.051-.368.08a49.94 49.94 0 0 0-3.438.915c-.408.123-.815.251-1.22.385-.31.103-.62.204-.931.313v.002l-.011.004-.003-.006c-.058.02-.114.045-.173.065-.736.262-1.466.537-2.19.833l-.025.01c-.745.305-1.483.63-2.213.97-.156.072-.306.15-.46.224-.559.267-1.114.542-1.663.83-.203.105-.403.217-.605.325-.493.268-.981.54-1.465.825-.101.058-.206.112-.307.171-.122.072-.238.152-.36.225a52.49 52.49 0 0 0-1.845 1.171 48.344 48.344 0 0 0-1.766 1.233 49.85 49.85 0 0 0-2.923 2.303 48.713 48.713 0 0 0-1.682 1.495l-.292.271c-.444.419-.88.849-1.311 1.285l-.2.203c-.447.458-.884.926-1.314 1.403-.047.051-.094.102-.14.155a50.186 50.186 0 0 0-2.748 3.35 48.396 48.396 0 0 0-1.293 1.802c-.334.486-.66.98-.978 1.48-.086.136-.17.274-.255.41-.288.463-.57.928-.845 1.4l-.322.574c-.242.43-.482.863-.713 1.304-.126.24-.246.485-.368.728-.157.311-.322.615-.472.93-.047.099-.087.199-.135.297-.126.27-.244.54-.367.812-.184.41-.37.82-.544 1.23-.085.203-.161.408-.243.612a51.182 51.182 0 0 0-2.35 7.332c-.025.108-.054.216-.08.324-.108.467-.2.933-.294 1.4-.054.264-.113.528-.162.792-.067.361-.122.722-.18 1.083-.063.382-.129.763-.183 1.147-.035.257-.063.514-.094.77-.063.497-.122.991-.17 1.49-.014.147-.024.296-.036.444A49.01 49.01 0 0 0 .063 49.2l-.004.103a51.026 51.026 0 0 0 2.789 19.269c.117.336.249.663.372.997a50.91 50.91 0 0 0 7.906 14.023l.184.234c.542.67 1.096 1.33 1.667 1.97l.045.05a50.578 50.578 0 0 0 1.749 1.853c.204.205.417.404.624.607.376.364.754.728 1.14 1.08.292.268.589.53.887.793.317.277.639.547.963.816.332.276.664.553 1.004.822.324.254.655.498.984.744.601.45 1.214.888 1.84 1.314.338.232.677.465 1.021.688.362.233.73.46 1.098.684.18.11.357.228.539.336.32.19.642.374.965.557.206.116.414.23.621.343.68.372 1.364.73 2.059 1.07.061.03.12.064.18.094l.004-.007c.003.003.007.004.01.005l-.005.011c.205.099.412.184.616.28a53.667 53.667 0 0 0 2.392 1.045 52.432 52.432 0 0 0 4.678 1.64c.39.116.783.229 1.177.335.188.052.377.113.565.161.15.04.301.07.45.107.154.038.308.072.463.108a50.537 50.537 0 0 0 2.63.554 49.211 49.211 0 0 0 2.475.392 51.625 51.625 0 0 0 2.257.246 56.695 56.695 0 0 0 2.168.138c.197.008.395.017.594.023a47.32 47.32 0 0 0 1.807.023c.125 0 .251-.004.376-.006a49.479 49.479 0 0 0 1.732-.051l.251-.01a53.37 53.37 0 0 0 1.958-.144c.093-.008.187-.019.28-.028.624-.061 1.249-.131 1.873-.217l.12-.015c.637-.089 1.274-.191 1.911-.305.14-.024.278-.051.417-.077a50.117 50.117 0 0 0 1.577-.319c.122-.027.246-.051.368-.08.561-.127 1.12-.267 1.68-.415.247-.064.493-.133.74-.202a48.26 48.26 0 0 0 2.232-.681c.312-.103.624-.206.936-.314l-.001-.003.012-.004.003.007c.059-.022.115-.046.174-.067a50.434 50.434 0 0 0 2.215-.841 53.565 53.565 0 0 0 2.213-.97c.154-.073.305-.152.46-.225a54.391 54.391 0 0 0 1.661-.828c.203-.107.405-.22.607-.329.492-.266.98-.539 1.463-.821.102-.06.207-.112.307-.172.123-.073.241-.153.362-.227a49.287 49.287 0 0 0 3.613-2.404c.395-.29.786-.584 1.175-.885.173-.135.348-.27.52-.408.413-.329.82-.665 1.224-1.007.136-.116.273-.231.407-.349.432-.374.855-.757 1.275-1.146l.294-.274c.442-.418.88-.846 1.308-1.281l.203-.207a49.632 49.632 0 0 0 1.454-1.558z" fill="#CE147C" />
-        </g>
-        <path d="M368.983 102.113c-8.6-.877-17.529.471-25.92 4.384A49.292 49.292 0 0 0 324 122.17c11.49 1.152 22.668 6.215 31.375 15.173 17.207 17.702 19.041 44.697 5.689 64.278 8.078.55 16.397-.87 24.244-4.53a49.29 49.29 0 0 0 20.019-17.039c9.545-14.06 11.656-32.69 3.924-49.272-7.733-16.583-23.36-26.94-40.268-28.667" fill="#CE147C" opacity=".8" />
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgThankYouHeroDesktop() {
-
-  return (
-    <svg
-      className="svg-thank-you-hero-desktop"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 1400 308"
-      preserveAspectRatio="xMinYMid slice"
-    >
-      <g fill="none" fillRule="evenodd">
-        <g transform="scale(1 -1) rotate(-55 -173.131 -218.706)">
-          <path
-            d="M174.133 124.74C185.688 114.745 193 99.976 193 83.5c0-28.288-21.552-51.543-49.133-54.239C153.422 20.998 165.877 16 179.5 16c30.1 0 54.5 24.4 54.5 54.5S209.6 125 179.5 125c-1.811 0-3.602-.088-5.367-.26z"
-            fill="#1896D7"
-            opacity=".399"
-          />
-          <path
-            d="M50.647 108.866c1.273.089 2.558.134 3.853.134 30.1 0 54.5-24.4 54.5-54.5a54.256 54.256 0 0 0-10.647-32.366C126.655 24.111 149 47.696 149 76.5c0 30.1-24.4 54.5-54.5 54.5-17.98 0-33.926-8.707-43.853-22.134z"
-            fill="#CE147C"
-            opacity=".8"
-          />
-          <circle cx="54.5" cy="54.5" fill="#CE147C" r="54.5" />
-          <path
-            d="M108.692 129.134C131.91 122.889 149 101.69 149 76.5c0-19.098-9.823-35.901-24.692-45.634A54.571 54.571 0 0 1 138.5 29c30.1 0 54.5 24.4 54.5 54.5S168.6 138 138.5 138c-11.002 0-21.242-3.26-29.808-8.866z"
-            fill="#CE147C"
-            opacity=".604"
-          />
-        </g>
-        <g fill="#F7691C" transform="scale(1 -1) rotate(-22 -554.639 -1815.699)">
-          <path
-            d="M44.104 120.11c4.48-8.004 11.045-14.984 19.47-20.046 18.818-11.307 41.722-9.926 58.756 1.66 9.24-16.506 9.616-37.365-.824-54.74-15.502-25.8-48.985-34.148-74.785-18.646-25.8 15.502-34.148 48.985-18.646 74.785a54.297 54.297 0 0 0 16.03 16.986z"
-            fillOpacity=".4"
-          />
-          <path
-            d="M86.567 201.046c1.836-16.282 10.964-31.555 26.094-40.646a54.256 54.256 0 0 1 33.146-7.55 54.262 54.262 0 0 0-7.448-34.14c-15.502-25.8-48.985-34.149-74.785-18.646-25.8 15.502-34.148 48.985-18.646 74.785 9.275 15.437 24.987 24.626 41.639 26.197z"
-            fillOpacity=".6"
-          />
-          <circle cx="140.73" cy="207.116" fillOpacity=".8" r="54.5" transform="scale(1 -1) rotate(-59 -225.346 0)" />
-          <circle cx="224.643" cy="244.671" r="54.5" transform="scale(1 -1) rotate(-59 -207.811 0)" />
-        </g>
-        <g fill="#FF1A21" transform="rotate(179 230.917 167.735)">
-          <path
-            d="M56.344 127.9c2.87-8.712 7.982-16.816 15.287-23.393 16.314-14.69 39.06-17.704 57.993-9.582 5.92-17.965 2.31-38.513-11.254-53.577-20.14-22.368-54.6-24.174-76.969-4.033-22.368 20.14-24.174 54.6-4.033 76.969a54.297 54.297 0 0 0 18.976 13.615z"
-            opacity=".2"
-          />
-          <path
-            d="M108.34 199.367c.124-16.384 7.606-32.528 21.702-43.15a54.256 54.256 0 0 1 32.176-10.975 54.262 54.262 0 0 0-10.976-33.174c-18.115-24.038-52.286-28.84-76.325-10.726-24.038 18.114-28.841 52.286-10.727 76.324 10.838 14.383 27.424 21.88 44.15 21.701z"
-            opacity=".4"
-          />
-          <path
-            d="M216.938 198.218c-16.643 10.6-25.652 28.838-25.193 47.357-.405.258-.815.511-1.229.76-25.8 15.503-59.283 7.154-74.785-18.646-15.502-25.8-7.154-59.283 18.646-74.785 25.8-15.502 59.283-7.154 74.785 18.646 5.031 8.373 7.55 17.555 7.776 26.668z"
-            opacity=".6"
-          />
-          <circle cx="246.237" cy="244.173" opacity=".8" r="54.5" transform="scale(1 -1) rotate(-59 -185.338 0)" />
-          <circle cx="366.747" cy="273.281" r="54.5" transform="scale(1 -1) rotate(-59 -116.275 0)" />
-        </g>
-        <g transform="rotate(179 415.292 185.413)">
-          <circle cx="327.714" cy="75.779" fill="#BDDBEA" r="54.5" transform="scale(1 -1) rotate(-59 193.775 0)" />
-          <circle cx="292.38" cy="209.183" fill="#94CAE5" r="54.5" transform="scale(1 -1) rotate(-59 -77.349 0)" />
-          <circle cx="235.202" cy="276.195" fill="#7FC1E3" r="54.5" transform="scale(1 -1) rotate(-59 -252.971 0)" />
-          <circle cx="186.985" cy="292.117" fill="#6AB8E0" r="54.5" transform="scale(1 -1) rotate(-59 -329.33 0)" />
-          <circle cx="121.155" cy="285.798" fill="#41A7DB" r="54.5" transform="scale(1 -1) rotate(-59 -383.992 0)" />
-          <circle cx="74.984" cy="238.052" fill="#1896D7" r="54.5" transform="scale(1 -1) rotate(-59 -345.771 0)" />
-        </g>
-      </g>
-    </svg>
-  );
-}
-
-function SvgThankYouHeroMobile() {
-
-  return (
-    <svg
-      className="svg-thank-you-hero-mobile"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 375 277"
-      preserveAspectRatio="xMinYMid"
-    >
-      <g fill="#CE147C" fillRule="evenodd" transform="rotate(-148 126.06 183.951)">
-        <path
-          d="M56.344 127.9c2.87-8.712 7.982-16.816 15.287-23.393 16.314-14.69 39.06-17.704 57.993-9.582 5.92-17.965 2.31-38.513-11.254-53.577-20.14-22.368-54.6-24.174-76.969-4.033-22.368 20.14-24.174 54.6-4.033 76.969a54.297 54.297 0 0 0 18.976 13.615z"
-          opacity=".2"
-        />
-        <path
-          d="M108.34 199.367c.124-16.384 7.606-32.528 21.702-43.15a54.256 54.256 0 0 1 32.176-10.975 54.262 54.262 0 0 0-10.976-33.174c-18.115-24.038-52.286-28.84-76.325-10.726-24.038 18.114-28.841 52.286-10.727 76.324 10.838 14.383 27.424 21.88 44.15 21.701z"
-          opacity=".4"
-        />
-        <path
-          d="M216.938 198.218c-16.643 10.6-25.652 28.838-25.193 47.357-.405.258-.815.511-1.229.76-25.8 15.503-59.283 7.154-74.785-18.646-15.502-25.8-7.154-59.283 18.646-74.785 25.8-15.502 59.283-7.154 74.785 18.646 5.031 8.373 7.55 17.555 7.776 26.668z"
-          opacity=".6"
-        />
-        <circle cx="246.237" cy="244.173" opacity=".8" r="54.5" transform="scale(1 -1) rotate(-59 -185.338 0)" />
-        <circle cx="366.747" cy="273.281" r="54.5" transform="scale(1 -1) rotate(-59 -116.275 0)" />
-      </g>
-      <g transform="rotate(156 224.236 227.744)">
-        <ellipse cx="278.059" cy="173.895" fill="#94CAE5" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -29.299 0)" />
-        <ellipse cx="231.847" cy="254.693" fill="#7FC1E3" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -218.322 0)" />
-        <ellipse cx="190.741" cy="300.625" fill="#6AB8E0" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -340.611 0)" />
-        <ellipse cx="133.727" cy="326.437" fill="#41A7DB" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -443.248 0)" />
-        <ellipse cx="67.806" cy="339.043" fill="#1896D7" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -531.45 0)" />
-      </g>
-      <g transform="rotate(119 136.185 213.425)">
-        <path
-          d="M157.017 112.084c10.42-8.98 17.013-22.25 17.013-37.055 0-25.419-19.434-46.314-44.303-48.737a49.053 49.053 0 0 1 32.13-11.915c27.14 0 49.143 21.925 49.143 48.97 0 27.047-22.002 48.972-49.143 48.972a49.91 49.91 0 0 1-4.84-.235z"
-          fill="#F7691C"
-          opacity=".399"
-        />
-        <path
-          d="M45.669 97.822c1.148.08 2.306.12 3.474.12 27.141 0 49.143-21.925 49.143-48.971 0-10.89-3.567-20.95-9.6-29.082 25.52 1.776 45.669 22.968 45.669 48.85 0 27.046-22.002 48.971-49.143 48.971-16.213 0-30.592-7.823-39.543-19.888z"
-          fill="#F48345"
-        />
-        <ellipse cx="49.143" cy="48.971" fill="#F7691C" rx="49.143" ry="48.971" />
-        <path
-          d="M98.009 116.033c20.936-5.611 36.346-24.66 36.346-47.294 0-17.16-8.858-32.26-22.266-41.004a49.37 49.37 0 0 1 12.798-1.677c27.14 0 49.143 21.925 49.143 48.971 0 27.046-22.002 48.971-49.143 48.971-9.92 0-19.154-2.93-26.878-7.967z"
-          fill="#F09D70"
-        />
-      </g>
-    </svg>
-  );
-}
-
-function SvgScribble() {
-
-  return (
-    <svg
-      className="svg-scribble"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 640 218"
-      preserveAspectRatio="xMinYMid"
-      aria-labelledby="title"
-    >
-      <title>No one edits our editor</title>
-      <g fill="none" fillRule="evenodd" transform="translate(-290) translate(314 9)">
-        <path d="M558.726 55.313l-19.435 12.931 1.98-25.508-27.768 34.594 8.726-50.6-32.131 51.224 9.093-39.001-40.138 57.37 19.98-67.161-34.686 45.065 5.767-31.49-23.185 28.3 6.48-36.477-30.084 42.77L426.929 0l-45.234 77.954 13.42-39.001-29.648 39.001 20.18-62.264-32.397 61.64 12.217-54.721L338.73 83.62l7.342-58.18-27.868 55.662 6.865-46.542-32.961 42.77 17.768-71.81-47.112 78.1 5.787-40.884c-23.274 27.256-34.647 40.884-34.12 40.884.527 0 7.404-20.337 20.632-61.011l-27.77 48.429V42.736L204.284 83.62l16.662-58.18-38.88 55.662 7.144-38.366-30.946 47.803 19.835-65.1C150.594 62.97 136.84 82.362 136.84 83.62V55.313L108.285 94.94l18.248-79.249-41.26 89.944 23.012-83.025-42.054 67.93 13.497-67.93C45.417 72.776 27.694 97.726 26.561 97.458c-1.134-.268 8.056-26.549 27.568-78.842L7.522 77.33 26.56 8.653 0 38.953l4.162-26.407" stroke="#121212" strokeLinecap="round" strokeLinejoin="round" strokeWidth="8" />
-        <path d="M611.067 151.56l-11.02 12.32-5.94-22.847-20.355 22.846 3.398-31.775-34.88 41.1 14.343-52.583-31.688 52.583 21.568-83.946-34.626 83.105 13.058-73.777-28.577 82.257 7.847-78.44-29.786 75.044 7.338-62.749-35.23 57.665 18.992-96.815-50.354 105.295 6.184-55.12c-24.874 36.746-37.03 55.12-36.467 55.12.563 0 7.914-27.42 22.051-82.257l-29.68 65.293v-38.157l-24.592 55.12 17.808-78.44-41.555 75.045 7.635-51.725-33.075 64.449 21.2-87.769c-31.688 56.817-47.532 86.073-47.532 87.769 0 1.696 3.31-27.56 9.931-87.769l-37.017 93.701 4.548-70.38-13.608 31.882L267.828 85.3l-48.834 87.904 18.441-70.802-43.616 81.935 18.756-76.916-39.052 65.783 10.826-65.783-29.442 44.139-6.897-19.456-16.346 14.56" stroke="#121212" strokeLinecap="round" strokeLinejoin="round" strokeWidth="8" />
-        <g fill="#FFF" transform="translate(80.08 24.8)">
-          <circle cx="51.876" cy="51.84" r="50.88" />
-          <circle cx="163.88" cy="51.88" r="50.88" />
-          <circle cx="261.88" cy="51.88" r="50.88" />
-          <circle cx="334.88" cy="50.88" r="50.88" />
-          <circle cx="213.88" cy="71.88" r="50.88" />
-          <circle cx="123.88" cy="61.88" r="50.88" />
-          <circle cx="221.88" cy="99.88" r="50.88" />
-          <circle cx="313.88" cy="98.88" r="50.88" />
-          <circle cx="153.88" cy="111.88" r="50.88" />
-          <circle cx="256.88" cy="108.88" r="50.88" />
-          <circle cx="375.88" cy="112.88" r="50.88" />
-        </g>
-        <text fontSize="56" fontWeight="bold">
-          <tspan fill="#121212" x="110" y="93">No one edits</tspan>
-        </text>
-        <text fontSize="56" fontWeight="bold">
-          <tspan fill="#121212" x="207" y="150">our editor</tspan>
-        </text>
-        <circle cx="513" cy="179" fill="#1896D7" r="30" />
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgAdvertisingGraphMobile() {
-
-  return (
-    <svg
-      className="svg-advertising-graph-mobile"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 349 196"
-      preserveAspectRatio="xMinYMid"
-      aria-labelledby="title"
-    >
-      <title>Advertising revenues are falling</title>
-      <g fill="none" fillRule="evenodd" transform="translate(-19 -17)">
-        <path d="M11.246 50.876l92.33 72.031 45.891-6.29 67.933 48.89 87.43-14.197 48.213 36.305" stroke="#121212" strokeLinecap="round" strokeWidth="8" />
-        <circle cx="338" cy="183" fill="#F7691C" r="30" transform="matrix(-1 0 0 1 676 0)" />
-        <text fill="#121212" fontSize="36" fontWeight="bold">
-          <tspan x="80" y="49">Advertising </tspan>
-        </text>
-        <text fill="#121212" fontSize="36" fontWeight="bold">
-          <tspan x="118" y="84">revenues are</tspan>
-        </text>
-        <text fill="#121212" fontSize="36" fontWeight="bold">
-          <tspan x="189" y="120">falling</tspan>
-        </text>
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgAdvertisingGraphDesktop() {
-
-  return (
-    <svg
-      className="svg-advertising-graph-desktop"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 618 235"
-      preserveAspectRatio="xMinYMid"
-      aria-labelledby="title"
-    >
-      <title>Advertising revenues are falling</title>
-      <g fill="none" fillRule="evenodd" transform="translate(4 0)">
-        <path d="M.36 57.373L75 105.277l80 5.567 80 48.234 80-7.413 80 19.976 80-12.563 80 33.112" stroke="#121212" strokeLinecap="round" strokeWidth="8" />
-        <circle cx="583" cy="204" fill="#FF7F0F" r="30" transform="matrix(-1 0 0 1 1166 0)" />
-        <text fontSize="56" fontWeight="bold">
-          <tspan fill="#121212" x="35" y="60">Advertising revenues</tspan>
-        </text>
-        <text fontSize="56" fontWeight="bold">
-          <tspan fill="#121212" x="270" y="119">are falling</tspan>
-        </text>
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgPaywallMobile() {
-
-  return (
-    <svg
-      className="svg-paywall-mobile"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 375 362"
-      preserveAspectRatio="xMinYMid"
-      aria-labelledby="title"
-    >
-      <title>We haven&#39;t put up a paywall</title>
-      <g fill="none" fillRule="evenodd" transform="translate(-233)">
-        <path fill="#121212" d="M812.571 150.476c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095c16.622 0 30.096 13.474 30.096 30.095s-13.474 30.095-30.096 30.095zm0-60.19c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm0 180.571c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095c16.622 0 30.096 13.474 30.096 30.095s-13.474 30.095-30.096 30.095zm0 60.19c-16.62 0-30.095-13.473-30.095-30.095 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm0-120.38c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm-52.165-90.286c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.096 13.475 30.096 30.096c0 16.62-13.475 30.095-30.096 30.095zm0-60.19c-16.62 0-30.095-13.475-30.095-30.096C730.311 13.475 743.785 0 760.406 0s30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 300.952c-16.62 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.096 13.474 30.096 30.096c0 16.62-13.475 30.095-30.096 30.095zm0-180.572c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 60.19c-16.62 0-30.095-13.473-30.095-30.094 0-16.621 13.474-30.096 30.095-30.096s30.096 13.475 30.096 30.096c0 16.62-13.475 30.095-30.096 30.095zm0 60.191c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zM708.241 90.286c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm0 180.571c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 60.19c-16.62 0-30.095-13.473-30.095-30.095 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zM603.911 90.287c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm52.165-210.667c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096C625.981 13.475 639.455 0 656.076 0s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 300.952c-16.621 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096 0-16.621 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zM551.746 120.38c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096C521.65 13.475 535.125 0 551.746 0s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 300.952c-16.621 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zM499.581 90.286c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zM447.416 120.38c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096C417.32 13.475 430.795 0 447.416 0s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 300.952c-16.621 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zM395.25 90.286c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zM343.086 120.38c-16.621 0-30.096-13.474-30.096-30.095s13.475-30.096 30.096-30.096c16.62 0 30.095 13.475 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zm0 240.762c-16.621 0-30.096-13.474-30.096-30.095 0-16.622 13.475-30.096 30.096-30.096 16.62 0 30.095 13.474 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zM134.426 120.38c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.096 30.095-30.096c16.622 0 30.096 13.475 30.096 30.096 0 16.62-13.474 30.095-30.096 30.095zm0 240.762c-16.622 0-30.096-13.474-30.096-30.095 0-16.622 13.474-30.096 30.095-30.096 16.622 0 30.096 13.474 30.096 30.096 0 16.62-13.474 30.095-30.096 30.095zm0-60.19c-16.622 0-30.096-13.475-30.096-30.096 0-16.621 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095s-13.474 30.095-30.096 30.095zM290.92 90.285c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.096-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.096-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zM82.26 150.476c-16.622 0-30.096-13.474-30.096-30.095S65.64 90.286 82.26 90.286s30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0-60.19c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm0 180.571c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 60.19c-16.622 0-30.096-13.473-30.096-30.095 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm0-120.38c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm156.495-90.286c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.096 30.096-30.096c16.62 0 30.095 13.475 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zm0 60.19c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095zm0 60.19c-16.622 0-30.096-13.473-30.096-30.094 0-16.621 13.474-30.096 30.096-30.096 16.62 0 30.095 13.475 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zm0 60.191c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095zM603.91 150.476c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 60.19c-16.621 0-30.095-13.473-30.095-30.095 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm0 60.191c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm-365.155 90.286c-16.622 0-30.096-13.474-30.096-30.095 0-16.622 13.474-30.096 30.096-30.096 16.62 0 30.095 13.474 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zM186.59 90.286c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm0 240.762c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zM30.095 120.38C13.475 120.381 0 106.907 0 90.286S13.474 60.19 30.095 60.19 60.19 73.665 60.19 90.286c0 16.62-13.474 30.095-30.095 30.095zm0-60.19C13.475 60.19 0 46.715 0 30.094 0 13.475 13.474 0 30.095 0S60.19 13.474 60.19 30.095 46.716 60.19 30.095 60.19zm0 300.952C13.475 361.143 0 347.669 0 331.048c0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-180.572C13.475 180.571 0 167.097 0 150.476s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 60.19C13.475 240.762 0 227.289 0 210.668c0-16.621 13.474-30.096 30.095-30.096s30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0 60.191C13.475 300.952 0 287.478 0 270.857s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm260.826-30.095c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095zm52.165 30.095c-16.621 0-30.096-13.474-30.096-30.095s13.475-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095z" id="a" />
-        <circle cx="552" cy="271" fill="#C28A2F" r="30" transform="matrix(-1 0 0 1 1104 0)" />
-        <text fill="#121212" fontSize="36" fontWeight="bold">
-          <tspan x="280" y="170">We haven’t put</tspan>
-        </text>
-        <text fill="#121212" fontSize="36" fontWeight="bold">
-          <tspan x="327" y="205">up a paywall</tspan>
-        </text>
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgPaywallDesktop() {
-
-  return (
-    <svg
-      className="svg-paywall-desktop"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 1400 360"
-      width="1400"
-      height="360"
-      preserveAspectRatio="xMinYMid"
-      aria-labelledby="title"
-    >
-      <title>We haven&#39;t put up a paywall</title>
-      <g fill="none" fillRule="evenodd">
-        <path fill="#121212" d="M179.961 190.691A29.93 29.93 0 0 1 178 180c0-3.766.694-7.37 1.961-10.691C174.458 175.846 166.214 180 157 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 178 120c0-3.766.694-7.37 1.961-10.691C174.458 115.846 166.214 120 157 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 178 60c0-3.766.694-7.37 1.961-10.691C174.458 55.846 166.214 60 157 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 3.766-.694 7.37-1.961 10.691C190.542 34.154 198.786 30 208 30c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 187 90c0 3.766-.694 7.37-1.961 10.691C190.542 94.154 198.786 90 208 90c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 187 150c0 3.766-.694 7.37-1.961 10.691C190.542 154.154 198.786 150 208 150c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 187 210c0 3.766-.694 7.37-1.961 10.691C190.542 214.154 198.786 210 208 210c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 187 270c0 3.766-.694 7.37-1.961 10.691C190.542 274.154 198.786 270 208 270c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 187 330c0 16.569-13.431 30-30 30-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 178 300c0-3.766.694-7.37 1.961-10.691C174.458 295.846 166.214 300 157 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 178 240c0-3.766.694-7.37 1.961-10.691C174.458 235.846 166.214 240 157 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691zM1196 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-150c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-104-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm260-150c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-104-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM988 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM936 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM832 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM780 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM728 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM624 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM364 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM520 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM312 150c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm156-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM416 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM260 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM53 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30C23 13.431 36.431 0 53 0c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-150c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm1299-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM1 150c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30z" id="b" />
-        <circle cx="832" cy="241" fill="#C28A2F" r="30" transform="matrix(-1 0 0 1 1664 0)" />
-        <text fill="#121212" fontSize="56" fontWeight="bold">
-          <tspan x="380" y="170">We haven’t put up</tspan>
-        </text>
-        <text fill="#121212" fontSize="56" fontWeight="bold">
-          <tspan x="457" y="229">a paywall</tspan>
-        </text>
-      </g>
-    </svg>
-  );
-
-}
-
-function SvgPaywallWide() {
-
-  return (
-    <svg
-      className="svg-paywall-wide"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 890 360"
-      width="890"
-      height="360"
-      preserveAspectRatio="xMinYMid"
-      aria-labelledby="title"
-    >
-      <title>We haven&#39;t put up a paywall</title>
-      <g fill="none" fillRule="evenodd">
-        <path fill="#121212" d="M52.961 190.691A29.93 29.93 0 0 1 51 180c0-3.766.694-7.37 1.961-10.691C47.458 175.846 39.214 180 30 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 120c0-3.766.694-7.37 1.961-10.691C47.458 115.846 39.214 120 30 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 60c0-3.766.694-7.37 1.961-10.691C47.458 55.846 39.214 60 30 60 13.431 60 0 46.569 0 30 0 13.431 13.431 0 30 0c16.569 0 30 13.431 30 30 0 3.766-.694 7.37-1.961 10.691C63.542 34.154 71.786 30 81 30c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 90c0 3.766-.694 7.37-1.961 10.691C63.542 94.154 71.786 90 81 90c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 150c0 3.766-.694 7.37-1.961 10.691C63.542 154.154 71.786 150 81 150c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 210c0 3.766-.694 7.37-1.961 10.691C63.542 214.154 71.786 210 81 210c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 270c0 3.766-.694 7.37-1.961 10.691C63.542 274.154 71.786 270 81 270c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 330c0 16.569-13.431 30-30 30-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 300c0-3.766.694-7.37 1.961-10.691C47.458 295.846 39.214 300 30 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 240c0-3.766.694-7.37 1.961-10.691C47.458 235.846 39.214 240 30 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691zm160.078 120C218.542 304.154 226.786 300 236 300c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-16.569 0-30-13.431-30-30 0-3.766.694-7.37 1.961-10.691C202.458 325.846 194.214 330 185 330c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 3.766-.694 7.37-1.961 10.691zm-5.078-210A29.93 29.93 0 0 1 206 90c0-3.766.694-7.37 1.961-10.691C202.458 85.846 194.214 90 185 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 3.766-.694 7.37-1.961 10.691C218.542 64.154 226.786 60 236 60c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 215 120c0 16.569-13.431 30-30 30-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691zM808 150c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM756 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM704 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM600 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM548 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM496 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM392 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM288 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-103-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30z" />
-        <circle cx="600" cy="241" fill="#C28A2F" r="30" transform="matrix(-1 0 0 1 1200 0)" />
-        <text fill="#121212" fontSize="56" fontWeight="bold">
-          <tspan x="225" y="173">We haven’t put up</tspan>
-        </text>
-        <text fill="#121212" fontSize="56" fontWeight="bold">
-          <tspan x="270" y="230">a paywall</tspan>
-        </text>
-      </g>
-    </svg>
-  );
-
-}
-
-
-// ----- Exports ----- //
-
-export {
-  SvgAmexLogo,
-  SvgArrowRightStraight,
-  SvgCreditCard,
-  SvgExclamation,
-  SvgExclamationAlternate,
-  SvgFacebook,
-  SvgGlobe,
-  SvgGuardianLogo,
-  SvgLock,
-  SvgPaypalLogo,
-  SvgMastercardLogo,
-  SvgTwitter,
-  SvgUser,
-  SvgVideo,
-  SvgPaypalPLogo,
-  SvgVisaLogo,
-  SvgChevronUp,
-  SvgCross,
-  SvgMovingCircle,
-  SvgDirectDebitSymbol,
-  SvgDirectDebitSymbolAndText,
-  SvgCirclesHeroDesktop,
-  SvgCirclesHeroMobileLandscape,
-  SvgCirclesHeroMobile,
-  SvgThankYouHeroDesktop,
-  SvgThankYouHeroMobile,
-  SvgScribble,
-  SvgAdvertisingGraphMobile,
-  SvgAdvertisingGraphDesktop,
-  SvgPaywallMobile,
-  SvgPaywallDesktop,
-  SvgPaywallWide,
-};
