+++ assets/components/svgs/advertisingGraphDesktop.jsx
@@ -1,0 +1,31 @@
+// @flow
+
+import React from 'react';
+
+function SvgAdvertisingGraphDesktop() {
+
+  return (
+    <svg
+      className="svg-advertising-graph-desktop"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 618 235"
+      preserveAspectRatio="xMinYMid"
+      aria-labelledby="title"
+    >
+      <title>Advertising revenues are falling</title>
+      <g fill="none" fillRule="evenodd" transform="translate(4 0)">
+        <path d="M.36 57.373L75 105.277l80 5.567 80 48.234 80-7.413 80 19.976 80-12.563 80 33.112" stroke="#121212" strokeLinecap="round" strokeWidth="8" />
+        <circle cx="583" cy="204" fill="#FF7F0F" r="30" transform="matrix(-1 0 0 1 1166 0)" />
+        <text fontSize="56" fontWeight="bold">
+          <tspan fill="#121212" x="35" y="60">Advertising revenues</tspan>
+        </text>
+        <text fontSize="56" fontWeight="bold">
+          <tspan fill="#121212" x="270" y="119">are falling</tspan>
+        </text>
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgAdvertisingGraphDesktop };
