+++ assets/components/svgs/advertisingGraphMobile.jsx
@@ -1,0 +1,33 @@
+// @flow
+
+import React from 'react';
+
+function SvgAdvertisingGraphMobile() {
+
+  return (
+    <svg
+      className="svg-advertising-graph-mobile"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 349 196"
+      preserveAspectRatio="xMinYMid"
+      aria-labelledby="title"
+    >
+      <title>Advertising revenues are falling</title>
+      <g fill="none" fillRule="evenodd" transform="translate(-19 -17)">
+        <path d="M11.246 50.876l92.33 72.031 45.891-6.29 67.933 48.89 87.43-14.197 48.213 36.305" stroke="#121212" strokeLinecap="round" strokeWidth="8" />
+        <circle cx="338" cy="183" fill="#F7691C" r="30" transform="matrix(-1 0 0 1 676 0)" />
+        <text fill="#121212" fontSize="36" fontWeight="bold">
+          <tspan x="80" y="49">Advertising </tspan>
+        </text>
+        <text fill="#121212" fontSize="36" fontWeight="bold">
+          <tspan x="118" y="84">revenues are</tspan>
+        </text>
+        <text fill="#121212" fontSize="36" fontWeight="bold">
+          <tspan x="189" y="120">falling</tspan>
+        </text>
+      </g>
+    </svg>
+  );
+}
+
+export { SvgAdvertisingGraphMobile };
