+++ assets/components/svgs/chevronUp.jsx
@@ -1,0 +1,16 @@
+// @flow
+
+import React from 'react';
+
+// A chevron symbol pointing up, used on buttons.
+function SvgChevronUp() {
+
+  return (
+    <svg className="svg-chevron-up" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30" preserveAspectRatio="xMinYMid">
+      <path d="M25.8999253 18.4499643 15.4499609 8 14.4499643 8 4 18.4499643 4.97499667 19.3999611 14.9499626 10.9999898 24.9249286 19.3999611Z" />
+    </svg>
+  );
+
+}
+
+export { SvgChevronUp };
