+++ assets/components/svgs/circlesHeroDesktop.jsx
@@ -1,0 +1,56 @@
+// @flow
+
+import React from 'react';
+
+function SvgCirclesHeroDesktop() {
+
+  return (
+    <svg
+      className="svg-circles-hero-desktop"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 1400 406"
+      preserveAspectRatio="xMinYMid slice"
+    >
+      <g fill="none" fillRule="evenodd">
+        <path d="M18.512 320.347c6.889-12.381 18.47-22.136 33.143-26.343A55.1 55.1 0 0 1 81 293.685c-12.986 13.825-18.576 33.973-12.954 53.579s21.039 33.73 39.378 38.572a55.1 55.1 0 0 1-25.056 15.28c-14.673 4.208-29.663 2.073-42.067-4.776-12.515-6.912-22.398-18.622-26.657-33.477-4.26-14.855-2.084-30.022 4.867-42.516" fill="#00B2FF" opacity=".6" />
+        <path d="M-57.133 342.729c6.83-12.768 18.588-22.86 33.578-27.159 14.671-4.207 29.662-2.072 42.067 4.777-6.951 12.494-9.128 27.661-4.868 42.516 4.26 14.855 14.143 26.565 26.658 33.476-6.89 12.383-18.472 22.137-33.143 26.344-14.99 4.298-30.31 1.97-42.867-5.239-12.132-6.963-21.69-18.48-25.857-33.015-4.168-14.535-2.166-29.366 4.432-41.7" fill="#00B2FF" opacity=".4" />
+        <g transform="scale(-1 1) rotate(16 -980.416 -636.061)">
+          <path d="M56.6 57.688l.003-.006.006-.012.006.018-.001.006-.008.014-.006-.02zm42.075 38.006c.485-.54.96-1.092 1.426-1.654l.14-.169c.458-.557.905-1.124 1.343-1.7l.161-.213c.419-.556.829-1.122 1.23-1.698.07-.102.14-.206.213-.309.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044.003-.04.003-.077.004-.115a56.828 56.828 0 0 0-3.112-21.48c-.132-.374-.277-.739-.415-1.111-.002 0-.002-.001-.002-.003a56.725 56.725 0 0 0-7.067-13.304l-.089-.123a54.93 54.93 0 0 0-1.67-2.201l-.205-.26a56.782 56.782 0 0 0-1.863-2.197l-.05-.056a56.419 56.419 0 0 0-1.952-2.066c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.067-.033-.132-.07-.201-.104l-.003.008c-.005-.003-.009-.005-.014-.006l.006-.012c-.226-.11-.457-.204-.684-.31a60.424 60.424 0 0 0-2.674-1.167 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006-.644.008-1.288.027-1.933.058l-.281.012a59.28 59.28 0 0 0-2.183.16l-.318.031a58.288 58.288 0 0 0-2.222.26c-.711.097-1.423.211-2.133.338-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35l.002.002-.014.004-.003-.006c-.064.023-.127.05-.193.072-.822.293-1.637.598-2.444.928-.01.003-.018.008-.029.01a58.98 58.98 0 0 0-2.47 1.083c-.175.08-.343.168-.516.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303c-.495.467-.983.946-1.464 1.433l-.223.226c-.499.51-.987 1.032-1.467 1.564-.052.057-.105.114-.156.173a55.969 55.969 0 0 0-1.43 1.658l-.134.163a55.554 55.554 0 0 0-1.504 1.913 53.951 53.951 0 0 0-1.445 2.009 55.887 55.887 0 0 0-1.09 1.65c-.097.151-.192.305-.286.457a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496a54.546 54.546 0 0 0-.148 2.159A56.796 56.796 0 0 0 3.18 76.44c.13.375.277.74.415 1.111a56.72 56.72 0 0 0 8.827 15.631l.206.261a59.275 59.275 0 0 0 1.861 2.197l.05.055a56.419 56.419 0 0 0 1.953 2.066c.228.229.465.45.697.676.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.07.033.134.071.202.104l.004-.007a.03.03 0 0 0 .012.006l-.006.012c.228.11.46.205.688.312a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 1.934-.057c.093-.005.186-.008.28-.012.727-.04 1.457-.094 2.186-.16l.313-.032a54.834 54.834 0 0 0 2.09-.24c.046-.007.09-.011.134-.017a58.17 58.17 0 0 0 2.135-.34c.156-.028.31-.058.465-.087a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35l-.001-.002.013-.005.003.008c.066-.024.13-.051.196-.076a56.368 56.368 0 0 0 2.472-.937c.832-.34 1.654-.703 2.47-1.081.174-.082.342-.169.514-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305c.494-.466.982-.943 1.46-1.429l.227-.23a55.362 55.362 0 0 0 1.623-1.737z" fill="#00B2FF" />
+        </g>
+        <path d="M81 293.686a55.1 55.1 0 0 1 25.056-15.28 55.106 55.106 0 0 1 27.542-.77 56.503 56.503 0 0 0-15.337 35.722c-1.545 27.482 16.671 51.53 42.26 58.228a55.096 55.096 0 0 1-23.75 13.932 55.1 55.1 0 0 1-29.347.317c-18.338-4.84-33.756-18.966-39.378-38.57-5.622-19.607-.032-39.755 12.955-53.58" fill="#00B2FF" opacity=".8" />
+        <path d="M1311.685 287.418c-2.147 30.695-28.682 53.845-59.27 51.706-30.586-2.14-53.642-28.757-51.495-59.452 2.146-30.695 28.682-53.845 59.269-51.706 30.586 2.139 53.642 28.756 51.496 59.452" fill="#FF1921" opacity=".1" />
+        <path d="M1446.65 304.256c-2.147 30.696-28.683 53.845-59.27 51.707-30.586-2.14-53.642-28.757-51.496-59.452 2.147-30.695 28.683-53.845 59.27-51.706 30.586 2.139 53.642 28.756 51.495 59.451" fill="#FF1921" opacity=".05" />
+        <path d="M1139.038 216.524c-15.556-1.088-30.056 4.375-40.841 14.037 9.018 10.693 14.058 24.783 13.004 39.867-1.055 15.084-8.007 28.336-18.426 37.67 9.335 11.068 22.934 18.496 38.49 19.584 30.587 2.139 57.123-21.011 59.27-51.706 2.146-30.695-20.91-57.313-51.497-59.452" fill="#FF1921" opacity=".2" />
+        <path d="M1019.552 224.396c-9.313-10.68-22.67-17.812-37.896-18.877a55.1 55.1 0 0 0-28.771 5.79c15.576 10.822 25.233 29.367 23.81 49.714-1.422 20.346-13.566 37.367-30.498 45.916a55.1 55.1 0 0 0 27.686 9.738c15.226 1.064 29.445-4.14 40.154-13.42 10.805-9.362 18.037-22.87 19.115-38.287 1.078-15.415-4.204-29.8-13.6-40.574" fill="#FF1921" opacity=".6" />
+        <path d="M1098.196 230.56c-9.335-11.068-22.934-18.496-38.49-19.584-15.226-1.064-29.445 4.14-40.155 13.42 9.397 10.775 14.68 25.158 13.602 40.574-1.078 15.416-8.312 28.925-19.116 38.288 9.314 10.678 22.67 17.812 37.896 18.876 15.556 1.088 30.056-4.375 40.841-14.037 10.42-9.333 17.373-22.585 18.428-37.669 1.054-15.084-3.987-29.174-13.006-39.867" fill="#FF1921" opacity=".4" />
+        <g transform="rotate(4 -2339.192 11743.357)">
+          <path d="M100.101 94.04a59.115 59.115 0 0 0 1.483-1.87c.58-.768.99-1.334 1.39-1.91l.214-.309c.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044a56.995 56.995 0 0 0-3.108-21.594c-.132-.375-.277-.74-.415-1.112a56.746 56.746 0 0 0-8.829-15.631 57.603 57.603 0 0 0-2.067-2.457 56.73 56.73 0 0 0-2.002-2.122c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.438-.224-.67-.318-.896-.425a60.424 60.424 0 0 0-2.674-1.166 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006a55.331 55.331 0 0 0-9.07.859c-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35-1.03.365-1.845.67-2.652 1a57.84 57.84 0 0 0-2.5 1.093c-.174.08-.342.168-.515.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303a57.085 57.085 0 0 0-4.74 5.054 57.078 57.078 0 0 0-4.174 5.734c-.096.152-.19.306-.285.458a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496-.06.68-.109 1.362-.143 2.044A56.963 56.963 0 0 0 3.18 76.44c.13.376.277.74.415 1.112a56.742 56.742 0 0 0 8.827 15.632 59.92 59.92 0 0 0 2.067 2.457 56.73 56.73 0 0 0 2.003 2.12c.228.23.465.451.697.677.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.44.225.672.32.9.427a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 9.072-.859c.156-.027.31-.057.465-.086a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35a60.552 60.552 0 0 0 2.642-.997 58.305 58.305 0 0 0 2.512-1.096c.173-.082.34-.169.513-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305a56.786 56.786 0 0 0 3.151-3.22 60.168 60.168 0 0 0 1.585-1.83z" fill="#FF1921" />
+        </g>
+        <path d="M952.885 211.309a55.1 55.1 0 0 0-27.685-9.738 55.106 55.106 0 0 0-27.1 4.975 56.503 56.503 0 0 1 22.43 31.752c7.224 26.56-5.594 53.87-29.232 65.742a55.096 55.096 0 0 0 26.129 8.69 55.1 55.1 0 0 0 28.771-5.792c16.931-8.548 29.075-25.57 30.498-45.915 1.423-20.347-8.234-38.892-23.811-49.714" fill="#FF1921" opacity=".8" />
+        <path d="M1351.417-24.959c6.921 29.982-11.687 59.878-41.562 66.776-29.876 6.897-59.706-11.817-66.628-41.798-6.922-29.982 11.686-59.878 41.562-66.776 29.875-6.897 59.706 11.817 66.628 41.798" fill="#CE147C" opacity=".1" />
+        <path d="M1166.71-39.496c-15.194 3.508-27.464 12.972-34.953 25.364 11.751 7.59 20.69 19.59 24.092 34.324 3.402 14.733.628 29.438-6.607 41.41 12.164 7.856 27.34 10.983 42.535 7.475 29.875-6.897 48.483-36.794 41.561-66.775-6.921-29.982-36.752-48.695-66.628-41.798" fill="#CE147C" opacity=".2" />
+        <path d="M1054.746 2.966c-12.028-7.49-26.886-10.406-41.759-6.972a55.1 55.1 0 0 0-25.821 13.948c18.06 5.796 32.717 20.707 37.305 40.58 4.589 19.874-2.048 39.702-15.74 52.828a55.1 55.1 0 0 0 29.322 1.217c14.873-3.433 26.949-12.568 34.477-24.572 7.595-12.113 10.561-27.146 7.085-42.203-3.476-15.058-12.732-27.269-24.869-34.826" fill="#CE147C" opacity=".6" />
+        <path d="M1131.757-14.132c-12.164-7.855-27.34-10.983-42.535-7.475-14.871 3.434-26.947 12.568-34.476 24.573 12.136 7.557 21.394 19.768 24.87 34.825 3.476 15.058.509 30.091-7.087 42.204 12.03 7.489 26.888 10.405 41.76 6.972 15.194-3.508 27.463-12.972 34.953-25.365 7.234-11.972 10.01-26.678 6.608-41.41-3.401-14.734-12.342-26.735-24.093-34.324" fill="#CE147C" opacity=".4" />
+        <g transform="rotate(-13 564.614 -3706.003)">
+          <path d="M100.101 94.04a59.115 59.115 0 0 0 1.483-1.87c.58-.768.99-1.334 1.39-1.91l.214-.309c.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044a56.995 56.995 0 0 0-3.108-21.594c-.132-.375-.277-.74-.415-1.112a56.746 56.746 0 0 0-8.829-15.631 57.603 57.603 0 0 0-2.067-2.457 56.73 56.73 0 0 0-2.002-2.122c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.438-.224-.67-.318-.896-.425a60.424 60.424 0 0 0-2.674-1.166 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006a55.331 55.331 0 0 0-9.07.859c-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35-1.03.365-1.845.67-2.652 1a57.84 57.84 0 0 0-2.5 1.093c-.174.08-.342.168-.515.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303a57.085 57.085 0 0 0-4.74 5.054 57.078 57.078 0 0 0-4.174 5.734c-.096.152-.19.306-.285.458a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496-.06.68-.109 1.362-.143 2.044A56.963 56.963 0 0 0 3.18 76.44c.13.376.277.74.415 1.112a56.742 56.742 0 0 0 8.827 15.632 59.92 59.92 0 0 0 2.067 2.457 56.73 56.73 0 0 0 2.003 2.12c.228.23.465.451.697.677.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.44.225.672.32.9.427a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 9.072-.859c.156-.027.31-.057.465-.086a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35a60.552 60.552 0 0 0 2.642-.997 58.305 58.305 0 0 0 2.512-1.096c.173-.082.34-.169.513-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305a56.786 56.786 0 0 0 3.151-3.22 60.168 60.168 0 0 0 1.585-1.83z" fill="#CE147C" />
+        </g>
+        <path d="M987.166 9.943a55.1 55.1 0 0 0-29.322-1.218 55.106 55.106 0 0 0-24.46 12.68 56.503 56.503 0 0 1 30.731 23.808c14.675 23.287 10.401 53.15-8.733 71.415a55.096 55.096 0 0 0 27.528.67 55.1 55.1 0 0 0 25.82-13.95c13.693-13.124 20.33-32.953 15.742-52.825-4.588-19.874-19.245-34.785-37.306-40.58" fill="#CE147C" opacity=".8" />
+        <path d="M970.018 391.233c-15.195-3.508-30.371-.38-42.535 7.475 7.235 11.972 10.009 26.678 6.608 41.41-3.402 14.734-12.341 26.735-24.092 34.324 7.489 12.393 19.758 21.856 34.953 25.364 29.875 6.898 59.706-11.816 66.628-41.797 6.922-29.982-11.686-59.879-41.562-66.776" fill="#AB8958" opacity=".4" />
+        <path d="M850.62 380.064c-7.528-12.004-19.603-21.14-34.476-24.573a55.1 55.1 0 0 0-29.323 1.218c13.692 13.126 20.33 32.953 15.741 52.827-4.588 19.873-19.245 34.785-37.305 40.58a55.1 55.1 0 0 0 25.82 13.949c14.873 3.433 29.732.517 41.76-6.973 12.136-7.557 21.392-19.768 24.869-34.825 3.476-15.058.51-30.09-7.086-42.203" fill="#C28A2F" opacity=".8" />
+        <path d="M927.332 398.456c-7.489-12.393-19.758-21.857-34.953-25.365-14.871-3.433-29.73-.516-41.759 6.973 7.595 12.112 10.563 27.145 7.087 42.203-3.477 15.057-12.734 27.268-24.87 34.825 7.529 12.005 19.605 21.14 34.476 24.573 15.195 3.508 30.371.38 42.535-7.475 11.75-7.59 20.692-19.59 24.093-34.323s.626-29.44-6.609-41.411" fill="#C28A2F" opacity=".6" />
+        <g transform="rotate(13 -1081.625 3262.394)">
+          <path d="M56.6 57.688l.003-.006.006-.012.006.018-.001.006-.008.014-.006-.02zm42.075 38.006c.485-.54.96-1.092 1.426-1.654l.14-.169c.458-.557.905-1.124 1.343-1.7l.161-.213c.419-.556.829-1.122 1.23-1.698.07-.102.14-.206.213-.309.37-.542.735-1.09 1.09-1.648.097-.152.192-.306.286-.46.32-.513.635-1.031.941-1.559.124-.212.24-.426.36-.64.27-.479.538-.961.796-1.45.141-.27.277-.544.415-.817.174-.345.355-.684.524-1.034.052-.11.097-.221.15-.33.14-.3.274-.605.411-.908.206-.456.414-.911.606-1.37.097-.227.182-.457.275-.685a56.876 56.876 0 0 0 2.62-8.17c.028-.12.061-.239.09-.361.12-.518.223-1.04.329-1.56.06-.295.126-.588.181-.884.074-.4.135-.803.2-1.205.07-.426.144-.85.204-1.28.04-.287.07-.573.107-.857.069-.553.136-1.106.189-1.662.015-.164.025-.33.039-.496.06-.68.11-1.36.144-2.044.003-.04.003-.077.004-.115a56.828 56.828 0 0 0-3.112-21.48c-.132-.374-.277-.739-.415-1.111-.002 0-.002-.001-.002-.003a56.725 56.725 0 0 0-7.067-13.304l-.089-.123a54.93 54.93 0 0 0-1.67-2.201l-.205-.26a56.782 56.782 0 0 0-1.863-2.197l-.05-.056a56.419 56.419 0 0 0-1.952-2.066c-.229-.229-.466-.45-.697-.676-.42-.407-.842-.81-1.273-1.205a61.84 61.84 0 0 0-.99-.883c-.355-.307-.714-.61-1.076-.91-.37-.307-.741-.616-1.12-.916-.363-.283-.732-.554-1.1-.83a57.422 57.422 0 0 0-2.052-1.464c-.38-.26-.758-.518-1.142-.767a53.71 53.71 0 0 0-1.225-.762c-.201-.124-.398-.255-.602-.375a58.767 58.767 0 0 0-1.772-1.004 57.245 57.245 0 0 0-2.298-1.193c-.067-.033-.132-.07-.201-.104l-.003.008c-.005-.003-.009-.005-.014-.006l.006-.012c-.226-.11-.457-.204-.684-.31a60.424 60.424 0 0 0-2.674-1.167 57.24 57.24 0 0 0-6.538-2.202c-.21-.058-.418-.125-.63-.18-.166-.043-.335-.076-.503-.119-.17-.042-.343-.08-.516-.12a57.07 57.07 0 0 0-2.15-.47 55.213 55.213 0 0 0-2.493-.432 63.363 63.363 0 0 0-1.054-.153 58.405 58.405 0 0 0-2.513-.272 61.447 61.447 0 0 0-2.429-.155c-.222-.01-.443-.02-.663-.026a53.038 53.038 0 0 0-2.02-.026c-.139 0-.28.005-.42.006-.644.008-1.288.027-1.933.058l-.281.012a59.28 59.28 0 0 0-2.183.16l-.318.031a58.288 58.288 0 0 0-2.222.26c-.711.097-1.423.211-2.133.338-.156.027-.312.057-.468.086-.586.11-1.173.227-1.758.355-.138.03-.275.058-.411.09-.626.14-1.252.298-1.877.462-.276.072-.55.149-.826.226a54.806 54.806 0 0 0-2.498.76c-.346.115-.693.228-1.04.35l.002.002-.014.004-.003-.006c-.064.023-.127.05-.193.072-.822.293-1.637.598-2.444.928-.01.003-.018.008-.029.01a58.98 58.98 0 0 0-2.47 1.083c-.175.08-.343.168-.516.25-.623.298-1.243.604-1.855.925-.227.117-.45.242-.676.363-.55.298-1.096.602-1.636.919-.113.064-.23.125-.342.191-.137.08-.266.169-.403.25a58.621 58.621 0 0 0-2.06 1.306 53.98 53.98 0 0 0-1.972 1.374 55.652 55.652 0 0 0-1.894 1.442 57.18 57.18 0 0 0-3.247 2.791l-.326.303c-.495.467-.983.946-1.464 1.433l-.223.226c-.499.51-.987 1.032-1.467 1.564-.052.057-.105.114-.156.173a55.969 55.969 0 0 0-1.43 1.658l-.134.163a55.554 55.554 0 0 0-1.504 1.913 53.951 53.951 0 0 0-1.445 2.009 55.887 55.887 0 0 0-1.09 1.65c-.097.151-.192.305-.286.457a58.083 58.083 0 0 0-.943 1.56c-.121.213-.24.426-.36.64-.27.48-.538.962-.796 1.453-.14.269-.274.541-.411.812-.176.347-.359.686-.527 1.037-.053.11-.098.221-.15.331-.141.3-.273.603-.41.906-.206.456-.414.913-.608 1.37-.094.227-.18.456-.272.683a56.989 56.989 0 0 0-2.622 8.173c-.029.12-.062.24-.09.361-.12.52-.224 1.04-.329 1.561-.06.294-.126.588-.18.883-.075.402-.137.804-.201 1.207-.07.426-.144.85-.204 1.279-.04.286-.071.572-.106.858-.07.553-.136 1.105-.19 1.66-.015.165-.026.33-.04.496a54.546 54.546 0 0 0-.148 2.159A56.796 56.796 0 0 0 3.18 76.44c.13.375.277.74.415 1.111a56.72 56.72 0 0 0 8.827 15.631l.206.261a59.275 59.275 0 0 0 1.861 2.197l.05.055a56.419 56.419 0 0 0 1.953 2.066c.228.229.465.45.697.676.42.407.842.812 1.273 1.205a56.271 56.271 0 0 0 2.066 1.793c.37.308.74.616 1.12.916.362.283.732.556 1.1.83.67.502 1.355.99 2.053 1.465.378.259.757.518 1.14.766.405.261.815.513 1.226.763.201.123.4.254.602.375.357.212.717.417 1.078.62.23.13.462.257.693.383.758.414 1.522.814 2.298 1.193.07.033.134.071.202.104l.004-.007a.03.03 0 0 0 .012.006l-.006.012c.228.11.46.205.688.312a59.971 59.971 0 0 0 2.67 1.164 58.6 58.6 0 0 0 5.223 1.829c.437.13.875.255 1.315.374.21.057.42.125.63.179.168.044.337.077.503.119.171.042.344.08.517.12a56.508 56.508 0 0 0 2.936.618c.567.104 1.135.199 1.704.285.353.053.705.105 1.06.152a57.742 57.742 0 0 0 2.52.274 63.403 63.403 0 0 0 2.42.154c.22.009.441.02.663.025.671.021 1.344.03 2.018.026.14 0 .28-.005.42-.006a55.33 55.33 0 0 0 1.934-.057c.093-.005.186-.008.28-.012.727-.04 1.457-.094 2.186-.16l.313-.032a54.834 54.834 0 0 0 2.09-.24c.046-.007.09-.011.134-.017a58.17 58.17 0 0 0 2.135-.34c.156-.028.31-.058.465-.087a56.04 56.04 0 0 0 1.761-.355c.137-.03.275-.057.412-.089a55.752 55.752 0 0 0 3.838-1.02c.453-.137.905-.28 1.356-.428.348-.114.696-.229 1.044-.35l-.001-.002.013-.005.003.008c.066-.024.13-.051.196-.076a56.368 56.368 0 0 0 2.472-.937c.832-.34 1.654-.703 2.47-1.081.174-.082.342-.169.514-.25a60.768 60.768 0 0 0 1.856-.924c.226-.119.452-.244.677-.366.55-.297 1.094-.601 1.633-.916.114-.066.231-.125.344-.191.136-.082.269-.17.404-.253a55.042 55.042 0 0 0 4.033-2.68c.442-.323.878-.651 1.312-.987.194-.15.39-.301.581-.455.461-.366.916-.741 1.366-1.122.152-.13.305-.258.455-.39.482-.416.955-.843 1.423-1.277l.329-.305c.494-.466.982-.943 1.46-1.429l.227-.23a55.362 55.362 0 0 0 1.623-1.737z" fill="#C28A2F" />
+        </g>
+        <g fill="#F7691C" transform="rotate(-175 108.214 55.111)">
+          <path d="M183.063 131.066c12.148-10.5 19.834-26.018 19.834-43.33 0-29.724-22.657-54.158-51.652-56.99 10.044-8.683 23.139-13.934 37.46-13.934C220.348 16.812 246 42.45 246 74.076s-25.652 57.265-57.295 57.265c-1.904 0-3.786-.093-5.642-.275z" opacity=".399" />
+          <path d="M53.244 114.388a58.16 58.16 0 0 0 4.05.141c31.644 0 57.296-25.638 57.296-57.265a56.988 56.988 0 0 0-11.193-34.007c29.753 2.077 53.244 26.858 53.244 57.123 0 31.627-25.652 57.265-57.295 57.265-18.902 0-35.666-9.148-46.102-23.257z" opacity=".8" />
+          <ellipse cx="57.295" cy="57.264" rx="57.295" ry="57.264" />
+          <path d="M114.266 135.684c24.409-6.562 42.375-28.836 42.375-55.304 0-20.066-10.327-37.722-25.959-47.948a57.398 57.398 0 0 1 14.92-1.961c31.644 0 57.295 25.638 57.295 57.265 0 31.626-25.651 57.264-57.294 57.264-11.566 0-22.332-3.425-31.337-9.316z" opacity=".604" />
+        </g>
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgCirclesHeroDesktop };
