+++ assets/components/svgs/circlesHeroMobile.jsx
@@ -1,0 +1,38 @@
+// @flow
+
+import React from 'react';
+
+function SvgCirclesHeroMobile() {
+
+  return (
+    <svg
+      className="svg-circles-hero-mobile"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 50 375 250"
+      preserveAspectRatio="xMinYMid"
+    >
+      <g fill="none" fillRule="evenodd">
+        <path d="M411.614 332.21c-3.946-12.08-12.473-22.65-24.67-28.864a49.417 49.417 0 0 0-25.782-5.303c9.067 14.393 10.54 33.088 2.236 49.387-8.305 16.299-24.295 26.095-41.269 27.22a49.417 49.417 0 0 0 19.444 17.74c12.198 6.216 25.76 6.901 37.853 2.994 12.2-3.943 22.905-12.561 29.197-24.91 6.293-12.35 6.973-26.076 2.991-38.264" fill="#00B2FF" opacity=".6" />
+        <path d="M474.38 364.86c-3.828-12.41-12.451-23.308-24.913-29.657-12.196-6.215-25.76-6.9-37.853-2.993 3.982 12.188 3.303 25.914-2.99 38.264-6.292 12.349-16.997 20.966-29.199 24.91 3.948 12.08 12.475 22.65 24.671 28.864 12.462 6.35 26.347 6.92 38.637 2.724 11.873-4.055 22.257-12.558 28.414-24.642 6.157-12.083 6.932-25.482 3.234-37.47" fill="#FF1921" opacity=".4" />
+        <g transform="rotate(27 -358.94 643.266)">
+          <path d="M50.763 51.739l.003-.006.005-.01.006.016-.002.005-.007.012-.005-.017zM88.5 85.825c.435-.485.86-.98 1.279-1.483l.125-.152c.41-.5.812-1.008 1.205-1.525l.144-.19a50.38 50.38 0 0 0 1.103-1.523c.063-.092.126-.186.19-.277.333-.487.66-.979.978-1.479.088-.136.173-.274.258-.412.286-.46.569-.925.844-1.398.11-.19.215-.382.323-.574.242-.43.482-.862.713-1.301.127-.242.248-.488.372-.733.156-.31.319-.613.47-.927.047-.098.087-.198.134-.296.127-.27.247-.543.37-.814.184-.41.371-.818.543-1.228.086-.204.163-.411.247-.615a51.01 51.01 0 0 0 2.35-7.327c.025-.108.054-.215.08-.325.108-.464.2-.932.295-1.398.054-.265.113-.528.163-.793.066-.36.12-.72.179-1.081.063-.382.129-.763.183-1.149.036-.256.063-.513.095-.768.062-.496.123-.992.17-1.49.013-.148.023-.297.035-.445.054-.61.098-1.22.13-1.834l.003-.102a50.968 50.968 0 0 0-2.79-19.265c-.12-.336-.25-.663-.374-.997l-.001-.003a50.875 50.875 0 0 0-6.339-11.931l-.08-.111a49.265 49.265 0 0 0-1.498-1.974c-.061-.079-.12-.156-.183-.233a50.926 50.926 0 0 0-1.67-1.97c-.015-.017-.03-.035-.045-.05a50.6 50.6 0 0 0-1.752-1.853c-.204-.206-.417-.404-.624-.607a57.506 57.506 0 0 0-1.142-1.08c-.292-.268-.59-.53-.889-.793-.317-.275-.639-.547-.964-.816-.332-.275-.665-.552-1.005-.821-.325-.254-.656-.497-.986-.745a51.5 51.5 0 0 0-1.84-1.313c-.34-.232-.68-.465-1.025-.688-.362-.233-.73-.459-1.098-.683-.18-.111-.357-.229-.54-.337a52.706 52.706 0 0 0-1.589-.9 51.34 51.34 0 0 0-2.061-1.07c-.06-.03-.119-.063-.18-.093l-.003.007c-.004-.003-.008-.004-.013-.006l.006-.01c-.203-.099-.41-.183-.614-.279a54.192 54.192 0 0 0-2.398-1.045 51.337 51.337 0 0 0-5.863-1.975c-.189-.052-.376-.113-.566-.162-.15-.039-.3-.069-.45-.106-.154-.038-.31-.072-.464-.108a51.185 51.185 0 0 0-2.633-.554 49.52 49.52 0 0 0-1.531-.256c-.315-.047-.63-.094-.945-.136a52.382 52.382 0 0 0-3.094-.308 55.11 55.11 0 0 0-1.339-.076c-.199-.008-.397-.018-.595-.023a47.568 47.568 0 0 0-1.81-.023c-.126 0-.252.004-.377.006-.578.006-1.156.024-1.735.05-.083.005-.167.008-.251.012-.652.035-1.305.083-1.958.143l-.285.028c-.624.061-1.248.133-1.873.216l-.12.017c-.638.087-1.276.19-1.913.304-.14.024-.28.051-.42.077-.525.098-1.052.204-1.577.319-.124.027-.246.05-.369.08-.561.126-1.123.267-1.683.414-.247.065-.494.134-.74.203a49.154 49.154 0 0 0-2.24.682c-.311.103-.623.204-.934.314h.002l-.012.005-.003-.005c-.058.02-.114.044-.174.064-.736.262-1.467.537-2.192.833-.008.002-.016.006-.025.01-.746.304-1.485.63-2.216.97-.156.072-.307.15-.462.224-.559.267-1.115.541-1.664.83-.203.104-.404.217-.606.325-.494.267-.983.54-1.467.824-.101.058-.206.112-.307.172-.123.071-.239.15-.361.224a52.575 52.575 0 0 0-1.847 1.171 48.414 48.414 0 0 0-1.77 1.232 49.913 49.913 0 0 0-2.927 2.303 48.754 48.754 0 0 0-1.684 1.494l-.292.272c-.444.419-.882.848-1.313 1.285l-.2.202a49.85 49.85 0 0 0-1.316 1.403c-.047.051-.094.102-.14.155-.436.487-.863.982-1.281 1.488l-.122.146A49.825 49.825 0 0 0 10.29 21a48.387 48.387 0 0 0-1.295 1.8c-.334.487-.66.98-.979 1.48l-.255.41c-.289.463-.571.928-.846 1.4-.109.191-.215.382-.323.574-.242.43-.482.863-.714 1.303-.126.24-.246.485-.369.729-.157.31-.321.614-.472.93-.047.098-.088.198-.135.296-.126.27-.245.54-.367.812-.185.41-.372.82-.546 1.23-.084.203-.161.408-.243.612a51.112 51.112 0 0 0-2.352 7.33c-.026.108-.056.216-.081.324-.108.466-.2.932-.295 1.4-.054.263-.113.527-.162.792-.067.36-.122.721-.18 1.082-.063.382-.13.763-.183 1.147-.035.257-.063.514-.094.77-.064.496-.123.99-.171 1.49-.014.146-.023.295-.035.444a48.92 48.92 0 0 0-.134 1.936 50.939 50.939 0 0 0 2.793 19.264c.117.337.249.664.373.997l.001.003a50.87 50.87 0 0 0 7.915 14.017l.185.233c.542.67 1.097 1.33 1.67 1.97l.044.05a50.6 50.6 0 0 0 1.751 1.853c.205.205.418.404.625.606.377.365.755.729 1.142 1.081a50.468 50.468 0 0 0 1.853 1.608c.332.276.664.553 1.005.822.324.254.656.498.985.744.602.45 1.216.888 1.842 1.313.34.233.679.465 1.023.688.363.234.73.46 1.1.684.18.11.357.228.539.336.32.19.644.375.967.557.206.116.414.23.622.343.68.372 1.365.73 2.061 1.07.062.03.12.064.18.093l.004-.006.011.005-.005.01c.204.1.412.185.616.28a53.783 53.783 0 0 0 2.396 1.045 52.559 52.559 0 0 0 4.684 1.64c.391.116.785.229 1.179.335.188.052.377.113.565.161.151.04.302.07.451.107.154.038.309.072.464.108a50.679 50.679 0 0 0 2.633.554c.509.093 1.018.178 1.528.255.317.048.633.095.95.137a51.773 51.773 0 0 0 2.261.246 56.85 56.85 0 0 0 2.17.137c.199.009.396.018.596.023.602.02 1.205.027 1.81.023.124 0 .251-.004.376-.005a49.622 49.622 0 0 0 1.734-.051l.252-.011a53.529 53.529 0 0 0 1.96-.143c.093-.008.187-.02.28-.029.625-.06 1.251-.13 1.876-.216l.12-.015a52.17 52.17 0 0 0 1.914-.305c.14-.024.279-.051.418-.077a50.26 50.26 0 0 0 1.579-.319c.122-.027.246-.051.369-.08a50.002 50.002 0 0 0 1.681-.415c.248-.064.494-.133.74-.202a48.387 48.387 0 0 0 2.237-.681c.312-.103.625-.206.937-.314l-.001-.002.012-.004.003.006c.059-.021.115-.046.175-.067.732-.26 1.46-.534 2.18-.827l.037-.014a53.684 53.684 0 0 0 2.216-.97c.155-.073.306-.151.46-.224a54.502 54.502 0 0 0 1.665-.828c.203-.107.405-.22.607-.329.493-.266.981-.539 1.465-.821.102-.06.207-.112.308-.172.123-.073.241-.152.362-.227a49.368 49.368 0 0 0 4.795-3.288c.173-.135.348-.27.52-.408.414-.329.822-.665 1.226-1.007.136-.116.273-.23.408-.348.432-.375.856-.757 1.276-1.146l.295-.275c.443-.417.88-.845 1.31-1.28.067-.07.136-.138.203-.207.446-.457.884-.924 1.313-1.4l.143-.158z" fill="#00B2FF" />
+        </g>
+        <path d="M361.162 298.043a49.417 49.417 0 0 0-19.444-17.74 49.423 49.423 0 0 0-24.115-5.39 50.676 50.676 0 0 1 7.39 34.074c-3.344 24.459-23.497 42.512-47.171 44.03a49.414 49.414 0 0 0 18.526 16.33 49.418 49.418 0 0 0 25.782 5.302c16.973-1.123 32.964-10.92 41.268-27.219 8.305-16.299 6.832-34.993-2.236-49.387" fill="#00B2FF" opacity=".8" />
+        <g fill="#F7691C" transform="scale(-1 1) rotate(12 -1307.849 -562.485)">
+          <path d="M161.027 116.498c10.747-9.31 17.548-23.067 17.548-38.416 0-26.352-20.045-48.015-45.696-50.526 8.886-7.697 20.47-12.354 33.14-12.354 27.994 0 50.688 22.73 50.688 50.77 0 28.039-22.694 50.77-50.688 50.77-1.684 0-3.35-.083-4.992-.244z" opacity=".399" />
+          <path d="M46.179 101.712c1.184.082 2.379.124 3.583.124 27.994 0 50.688-22.73 50.688-50.769 0-11.29-3.68-21.72-9.902-30.15 26.322 1.841 47.104 23.812 47.104 50.644 0 28.04-22.693 50.77-50.688 50.77-16.722 0-31.553-8.111-40.785-20.62z" opacity=".8" />
+          <ellipse cx="50.755" cy="51.394" rx="50.688" ry="50.769" />
+          <path d="M100.164 120.592c21.594-5.818 37.488-25.565 37.488-49.03 0-17.791-9.136-33.445-22.965-42.51a50.677 50.677 0 0 1 13.2-1.74c27.994 0 50.688 22.73 50.688 50.77 0 28.04-22.694 50.77-50.688 50.77a50.393 50.393 0 0 1-27.723-8.26z" opacity=".604" />
+        </g>
+        <path d="M393.234 63.858c-12.669.896-24.977 6.6-33.946 16.917a49.289 49.289 0 0 0-11.358 23.707c16.148-5.3 34.637-2.194 48.446 9.81 13.81 12.004 19.458 29.882 16.457 46.61a49.289 49.289 0 0 0 21.896-14.547c8.97-10.318 12.904-23.3 12.028-35.97-.884-12.785-6.664-25.252-17.127-34.347-10.462-9.095-23.613-13.084-36.396-12.18" fill="#CE147C" opacity=".6" />
+        <g transform="rotate(-23 471.104 -568.397)">
+          <path d="M50.693 51.75l.003-.005.005-.01.006.016-.002.005-.006.012-.006-.017zm37.684 34.095c.434-.486.859-.98 1.277-1.484l.125-.152c.41-.5.81-1.008 1.203-1.525l.144-.191a50.39 50.39 0 0 0 1.101-1.523l.191-.277c.332-.487.66-.979.976-1.479.088-.136.173-.274.257-.412.287-.46.569-.925.843-1.398.11-.191.215-.383.323-.575.242-.43.481-.862.713-1.301.126-.242.247-.488.37-.733.157-.31.32-.613.47-.927.047-.099.087-.199.134-.296.127-.27.246-.543.369-.815.184-.41.37-.817.543-1.228.086-.204.163-.411.246-.615a51.081 51.081 0 0 0 2.346-7.329c.026-.108.055-.215.08-.324.108-.465.201-.933.295-1.399.054-.265.113-.528.163-.793.066-.36.12-.72.179-1.081.063-.383.129-.764.183-1.15.036-.256.063-.513.095-.768.062-.496.122-.992.17-1.49.013-.148.022-.297.034-.445.054-.61.098-1.22.13-1.834l.003-.103a51.055 51.055 0 0 0-2.787-19.269c-.118-.336-.248-.663-.372-.997l-.001-.003a50.913 50.913 0 0 0-6.33-11.934l-.08-.111a49.269 49.269 0 0 0-1.496-1.975l-.183-.232a50.917 50.917 0 0 0-1.669-1.972l-.044-.05a50.578 50.578 0 0 0-1.75-1.853c-.203-.205-.416-.404-.623-.606a57.466 57.466 0 0 0-1.14-1.082c-.292-.267-.589-.53-.887-.792a55.888 55.888 0 0 0-.963-.816c-.332-.276-.664-.553-1.004-.821-.324-.255-.655-.498-.985-.745a51.43 51.43 0 0 0-1.838-1.314c-.34-.232-.679-.465-1.023-.688-.361-.233-.728-.46-1.097-.683-.18-.111-.356-.229-.539-.337a52.612 52.612 0 0 0-1.586-.9 51.24 51.24 0 0 0-2.059-1.07c-.06-.03-.118-.064-.18-.094l-.003.007c-.004-.003-.008-.004-.012-.005l.006-.011c-.203-.099-.41-.182-.614-.278a54.075 54.075 0 0 0-2.394-1.046 51.213 51.213 0 0 0-5.855-1.976c-.189-.052-.376-.112-.565-.161-.15-.04-.3-.069-.45-.107-.154-.038-.308-.071-.463-.108a51.044 51.044 0 0 0-2.63-.554 49.378 49.378 0 0 0-1.529-.255c-.314-.048-.629-.095-.943-.137a52.23 52.23 0 0 0-3.09-.308c-.445-.031-.89-.055-1.336-.076-.2-.008-.397-.017-.595-.023a47.427 47.427 0 0 0-1.808-.023c-.125 0-.252.004-.377.006-.576.006-1.153.024-1.731.051l-.252.01c-.65.036-1.303.085-1.955.144l-.285.028a52.125 52.125 0 0 0-1.99.233c-.637.088-1.274.19-1.91.304-.14.024-.28.051-.42.077a54.12 54.12 0 0 0-1.574.319c-.124.027-.246.051-.368.08a49.94 49.94 0 0 0-3.438.915c-.408.123-.815.251-1.22.385-.31.103-.62.204-.931.313v.002l-.011.004-.003-.006c-.058.02-.114.045-.173.065-.736.262-1.466.537-2.19.833l-.025.01c-.745.305-1.483.63-2.213.97-.156.072-.306.15-.46.224-.559.267-1.114.542-1.663.83-.203.105-.403.217-.605.325-.493.268-.981.54-1.465.825-.101.058-.206.112-.307.171-.122.072-.238.152-.36.225a52.49 52.49 0 0 0-1.845 1.171 48.344 48.344 0 0 0-1.766 1.233 49.85 49.85 0 0 0-2.923 2.303 48.713 48.713 0 0 0-1.682 1.495l-.292.271c-.444.419-.88.849-1.311 1.285l-.2.203c-.447.458-.884.926-1.314 1.403-.047.051-.094.102-.14.155a50.186 50.186 0 0 0-2.748 3.35 48.396 48.396 0 0 0-1.293 1.802c-.334.486-.66.98-.978 1.48-.086.136-.17.274-.255.41-.288.463-.57.928-.845 1.4l-.322.574c-.242.43-.482.863-.713 1.304-.126.24-.246.485-.368.728-.157.311-.322.615-.472.93-.047.099-.087.199-.135.297-.126.27-.244.54-.367.812-.184.41-.37.82-.544 1.23-.085.203-.161.408-.243.612a51.182 51.182 0 0 0-2.35 7.332c-.025.108-.054.216-.08.324-.108.467-.2.933-.294 1.4-.054.264-.113.528-.162.792-.067.361-.122.722-.18 1.083-.063.382-.129.763-.183 1.147-.035.257-.063.514-.094.77-.063.497-.122.991-.17 1.49-.014.147-.024.296-.036.444A49.01 49.01 0 0 0 .063 49.2l-.004.103a51.026 51.026 0 0 0 2.789 19.269c.117.336.249.663.372.997a50.91 50.91 0 0 0 7.906 14.023l.184.234c.542.67 1.096 1.33 1.667 1.97l.045.05a50.578 50.578 0 0 0 1.749 1.853c.204.205.417.404.624.607.376.364.754.728 1.14 1.08.292.268.589.53.887.793.317.277.639.547.963.816.332.276.664.553 1.004.822.324.254.655.498.984.744.601.45 1.214.888 1.84 1.314.338.232.677.465 1.021.688.362.233.73.46 1.098.684.18.11.357.228.539.336.32.19.642.374.965.557.206.116.414.23.621.343.68.372 1.364.73 2.059 1.07.061.03.12.064.18.094l.004-.007c.003.003.007.004.01.005l-.005.011c.205.099.412.184.616.28a53.667 53.667 0 0 0 2.392 1.045 52.432 52.432 0 0 0 4.678 1.64c.39.116.783.229 1.177.335.188.052.377.113.565.161.15.04.301.07.45.107.154.038.308.072.463.108a50.537 50.537 0 0 0 2.63.554 49.211 49.211 0 0 0 2.475.392 51.625 51.625 0 0 0 2.257.246 56.695 56.695 0 0 0 2.168.138c.197.008.395.017.594.023a47.32 47.32 0 0 0 1.807.023c.125 0 .251-.004.376-.006a49.479 49.479 0 0 0 1.732-.051l.251-.01a53.37 53.37 0 0 0 1.958-.144c.093-.008.187-.019.28-.028.624-.061 1.249-.131 1.873-.217l.12-.015c.637-.089 1.274-.191 1.911-.305.14-.024.278-.051.417-.077a50.117 50.117 0 0 0 1.577-.319c.122-.027.246-.051.368-.08.561-.127 1.12-.267 1.68-.415.247-.064.493-.133.74-.202a48.26 48.26 0 0 0 2.232-.681c.312-.103.624-.206.936-.314l-.001-.003.012-.004.003.007c.059-.022.115-.046.174-.067a50.434 50.434 0 0 0 2.215-.841 53.565 53.565 0 0 0 2.213-.97c.154-.073.305-.152.46-.225a54.391 54.391 0 0 0 1.661-.828c.203-.107.405-.22.607-.329.492-.266.98-.539 1.463-.821.102-.06.207-.112.307-.172.123-.073.241-.153.362-.227a49.287 49.287 0 0 0 3.613-2.404c.395-.29.786-.584 1.175-.885.173-.135.348-.27.52-.408.413-.329.82-.665 1.224-1.007.136-.116.273-.231.407-.349.432-.374.855-.757 1.275-1.146l.294-.274c.442-.418.88-.846 1.308-1.281l.203-.207a49.632 49.632 0 0 0 1.454-1.558z" fill="#CE147C" />
+        </g>
+        <path d="M368.983 102.113c-8.6-.877-17.529.471-25.92 4.384A49.292 49.292 0 0 0 324 122.17c11.49 1.152 22.668 6.215 31.375 15.173 17.207 17.702 19.041 44.697 5.689 64.278 8.078.55 16.397-.87 24.244-4.53a49.29 49.29 0 0 0 20.019-17.039c9.545-14.06 11.656-32.69 3.924-49.272-7.733-16.583-23.36-26.94-40.268-28.667" fill="#CE147C" opacity=".8" />
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgCirclesHeroMobile };
