+++ assets/components/svgs/circlesHeroMobileLandscape.jsx
@@ -1,0 +1,45 @@
+// @flow
+
+import React from 'react';
+
+function SvgCirclesHeroMobileLandscape() {
+
+  return (
+    <svg
+      className="svg-circles-hero-mobile-landscape"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 768 382"
+      preserveAspectRatio="xMinYMid slice"
+    >
+      <g fill="none" fillRule="evenodd" transform="translate(-2)">
+        <path d="M728.406 348.214c-7.84-10.002-19.467-17.017-33.055-18.686a49.417 49.417 0 0 0-26.04 3.835c13.443 10.424 21.222 27.487 18.992 45.644-2.23 18.156-13.905 32.83-29.47 39.693a49.417 49.417 0 0 0 24.339 10.02c13.587 1.67 26.567-2.325 36.593-10.133 10.117-7.878 17.228-19.638 18.917-33.394 1.69-13.757-2.366-26.887-10.276-36.98" fill="#FF1921" opacity=".6" />
+        <path d="M798.554 357.427c-7.842-10.352-19.673-17.643-33.554-19.347-13.587-1.668-26.566 2.326-36.594 10.134 7.91 10.091 11.967 23.222 10.277 36.979-1.689 13.756-8.801 25.516-18.918 33.394 7.841 10.002 19.469 17.017 33.055 18.685 13.882 1.705 27.126-2.507 37.239-10.654 9.77-7.871 16.62-19.414 18.272-32.874 1.653-13.46-2.201-26.317-9.777-36.317" fill="#FF1921" opacity=".4" />
+        <g transform="rotate(7 -2256.059 4613.011)">
+          <path d="M50.763 51.739l.003-.006.005-.01.006.016-.002.005-.007.012-.005-.017zM88.5 85.825c.435-.485.86-.98 1.279-1.483l.125-.152c.41-.5.812-1.008 1.205-1.525l.144-.19a50.38 50.38 0 0 0 1.103-1.523c.063-.092.126-.186.19-.277.333-.487.66-.979.978-1.479.088-.136.173-.274.258-.412.286-.46.569-.925.844-1.398.11-.19.215-.382.323-.574.242-.43.482-.862.713-1.301.127-.242.248-.488.372-.733.156-.31.319-.613.47-.927.047-.098.087-.198.134-.296.127-.27.247-.543.37-.814.184-.41.371-.818.543-1.228.086-.204.163-.411.247-.615a51.01 51.01 0 0 0 2.35-7.327c.025-.108.054-.215.08-.325.108-.464.2-.932.295-1.398.054-.265.113-.528.163-.793.066-.36.12-.72.179-1.081.063-.382.129-.763.183-1.149.036-.256.063-.513.095-.768.062-.496.123-.992.17-1.49.013-.148.023-.297.035-.445.054-.61.098-1.22.13-1.834l.003-.102a50.968 50.968 0 0 0-2.79-19.265c-.12-.336-.25-.663-.374-.997l-.001-.003a50.875 50.875 0 0 0-6.339-11.931l-.08-.111a49.265 49.265 0 0 0-1.498-1.974c-.061-.079-.12-.156-.183-.233a50.926 50.926 0 0 0-1.67-1.97c-.015-.017-.03-.035-.045-.05a50.6 50.6 0 0 0-1.752-1.853c-.204-.206-.417-.404-.624-.607a57.506 57.506 0 0 0-1.142-1.08c-.292-.268-.59-.53-.889-.793-.317-.275-.639-.547-.964-.816-.332-.275-.665-.552-1.005-.821-.325-.254-.656-.497-.986-.745a51.5 51.5 0 0 0-1.84-1.313c-.34-.232-.68-.465-1.025-.688-.362-.233-.73-.459-1.098-.683-.18-.111-.357-.229-.54-.337a52.706 52.706 0 0 0-1.589-.9 51.34 51.34 0 0 0-2.061-1.07c-.06-.03-.119-.063-.18-.093l-.003.007c-.004-.003-.008-.004-.013-.006l.006-.01c-.203-.099-.41-.183-.614-.279a54.192 54.192 0 0 0-2.398-1.045 51.337 51.337 0 0 0-5.863-1.975c-.189-.052-.376-.113-.566-.162-.15-.039-.3-.069-.45-.106-.154-.038-.31-.072-.464-.108a51.185 51.185 0 0 0-2.633-.554 49.52 49.52 0 0 0-1.531-.256c-.315-.047-.63-.094-.945-.136a52.382 52.382 0 0 0-3.094-.308 55.11 55.11 0 0 0-1.339-.076c-.199-.008-.397-.018-.595-.023a47.568 47.568 0 0 0-1.81-.023c-.126 0-.252.004-.377.006-.578.006-1.156.024-1.735.05-.083.005-.167.008-.251.012-.652.035-1.305.083-1.958.143l-.285.028c-.624.061-1.248.133-1.873.216l-.12.017c-.638.087-1.276.19-1.913.304-.14.024-.28.051-.42.077-.525.098-1.052.204-1.577.319-.124.027-.246.05-.369.08-.561.126-1.123.267-1.683.414-.247.065-.494.134-.74.203a49.154 49.154 0 0 0-2.24.682c-.311.103-.623.204-.934.314h.002l-.012.005-.003-.005c-.058.02-.114.044-.174.064-.736.262-1.467.537-2.192.833-.008.002-.016.006-.025.01-.746.304-1.485.63-2.216.97-.156.072-.307.15-.462.224-.559.267-1.115.541-1.664.83-.203.104-.404.217-.606.325-.494.267-.983.54-1.467.824-.101.058-.206.112-.307.172-.123.071-.239.15-.361.224a52.575 52.575 0 0 0-1.847 1.171 48.414 48.414 0 0 0-1.77 1.232 49.913 49.913 0 0 0-2.927 2.303 48.754 48.754 0 0 0-1.684 1.494l-.292.272c-.444.419-.882.848-1.313 1.285l-.2.202a49.85 49.85 0 0 0-1.316 1.403c-.047.051-.094.102-.14.155-.436.487-.863.982-1.281 1.488l-.122.146A49.825 49.825 0 0 0 10.29 21a48.387 48.387 0 0 0-1.295 1.8c-.334.487-.66.98-.979 1.48l-.255.41c-.289.463-.571.928-.846 1.4-.109.191-.215.382-.323.574-.242.43-.482.863-.714 1.303-.126.24-.246.485-.369.729-.157.31-.321.614-.472.93-.047.098-.088.198-.135.296-.126.27-.245.54-.367.812-.185.41-.372.82-.546 1.23-.084.203-.161.408-.243.612a51.112 51.112 0 0 0-2.352 7.33c-.026.108-.056.216-.081.324-.108.466-.2.932-.295 1.4-.054.263-.113.527-.162.792-.067.36-.122.721-.18 1.082-.063.382-.13.763-.183 1.147-.035.257-.063.514-.094.77-.064.496-.123.99-.171 1.49-.014.146-.023.295-.035.444a48.92 48.92 0 0 0-.134 1.936 50.939 50.939 0 0 0 2.793 19.264c.117.337.249.664.373.997l.001.003a50.87 50.87 0 0 0 7.915 14.017l.185.233c.542.67 1.097 1.33 1.67 1.97l.044.05a50.6 50.6 0 0 0 1.751 1.853c.205.205.418.404.625.606.377.365.755.729 1.142 1.081a50.468 50.468 0 0 0 1.853 1.608c.332.276.664.553 1.005.822.324.254.656.498.985.744.602.45 1.216.888 1.842 1.313.34.233.679.465 1.023.688.363.234.73.46 1.1.684.18.11.357.228.539.336.32.19.644.375.967.557.206.116.414.23.622.343.68.372 1.365.73 2.061 1.07.062.03.12.064.18.093l.004-.006.011.005-.005.01c.204.1.412.185.616.28a53.783 53.783 0 0 0 2.396 1.045 52.559 52.559 0 0 0 4.684 1.64c.391.116.785.229 1.179.335.188.052.377.113.565.161.151.04.302.07.451.107.154.038.309.072.464.108a50.679 50.679 0 0 0 2.633.554c.509.093 1.018.178 1.528.255.317.048.633.095.95.137a51.773 51.773 0 0 0 2.261.246 56.85 56.85 0 0 0 2.17.137c.199.009.396.018.596.023.602.02 1.205.027 1.81.023.124 0 .251-.004.376-.005a49.622 49.622 0 0 0 1.734-.051l.252-.011a53.529 53.529 0 0 0 1.96-.143c.093-.008.187-.02.28-.029.625-.06 1.251-.13 1.876-.216l.12-.015a52.17 52.17 0 0 0 1.914-.305c.14-.024.279-.051.418-.077a50.26 50.26 0 0 0 1.579-.319c.122-.027.246-.051.369-.08a50.002 50.002 0 0 0 1.681-.415c.248-.064.494-.133.74-.202a48.387 48.387 0 0 0 2.237-.681c.312-.103.625-.206.937-.314l-.001-.002.012-.004.003.006c.059-.021.115-.046.175-.067.732-.26 1.46-.534 2.18-.827l.037-.014a53.684 53.684 0 0 0 2.216-.97c.155-.073.306-.151.46-.224a54.502 54.502 0 0 0 1.665-.828c.203-.107.405-.22.607-.329.493-.266.981-.539 1.465-.821.102-.06.207-.112.308-.172.123-.073.241-.152.362-.227a49.368 49.368 0 0 0 4.795-3.288c.173-.135.348-.27.52-.408.414-.329.822-.665 1.226-1.007.136-.116.273-.23.408-.348.432-.375.856-.757 1.276-1.146l.295-.275c.443-.417.88-.845 1.31-1.28.067-.07.136-.138.203-.207.446-.457.884-.924 1.313-1.4l.143-.158z" fill="#FF1921" />
+        </g>
+        <path d="M669.311 333.363a49.417 49.417 0 0 0-24.339-10.02 49.423 49.423 0 0 0-24.505 3.183 50.676 50.676 0 0 1 18.598 29.492c5.224 24.127-7.538 47.985-29.266 57.508a49.414 49.414 0 0 0 22.994 9.01 49.418 49.418 0 0 0 26.04-3.837c15.566-6.86 27.241-21.537 29.47-39.692 2.23-18.156-5.549-35.22-18.992-45.644" fill="#FF1921" opacity=".8" />
+        <g fill="#F7691C" transform="scale(-1 1) rotate(-3 6462.427 3155.055)">
+          <path d="M161.027 116.498c10.747-9.31 17.548-23.067 17.548-38.416 0-26.352-20.045-48.015-45.696-50.526 8.886-7.697 20.47-12.354 33.14-12.354 27.994 0 50.688 22.73 50.688 50.77 0 28.039-22.694 50.77-50.688 50.77-1.684 0-3.35-.083-4.992-.244z" opacity=".399" />
+          <path d="M46.179 101.712c1.184.082 2.379.124 3.583.124 27.994 0 50.688-22.73 50.688-50.769 0-11.29-3.68-21.72-9.902-30.15 26.322 1.841 47.104 23.812 47.104 50.644 0 28.04-22.693 50.77-50.688 50.77-16.722 0-31.553-8.111-40.785-20.62z" opacity=".8" />
+          <ellipse cx="50.755" cy="51.394" rx="50.688" ry="50.769" />
+          <path d="M100.164 120.592c21.594-5.818 37.488-25.565 37.488-49.03 0-17.791-9.136-33.445-22.965-42.51a50.677 50.677 0 0 1 13.2-1.74c27.994 0 50.688 22.73 50.688 50.77 0 28.04-22.694 50.77-50.688 50.77a50.393 50.393 0 0 1-27.723-8.26z" opacity=".604" />
+        </g>
+        <path d="M839.701 144.582c6.678 26.783-9.51 53.881-36.155 60.525-26.646 6.643-53.66-9.684-60.338-36.467-6.678-26.783 9.51-53.881 36.155-60.525 26.646-6.643 53.66 9.684 60.338 36.467" fill="#CE147C" opacity=".1" />
+        <path d="M674.05 134.373c-13.551 3.379-24.388 12.055-30.897 23.284 10.645 6.626 18.842 17.252 22.124 30.414 3.281 13.161 1.032 26.392-5.256 37.24 11.019 6.86 24.66 9.432 38.212 6.053 26.646-6.643 42.833-33.742 36.156-60.525-6.678-26.783-33.693-43.11-60.338-36.466" fill="#CE147C" opacity=".2" />
+        <path d="M576.42 175.672c-10.657-6.908-23.919-9.761-37.291-6.919a49.289 49.289 0 0 0-23.338 12.098c16.085 5.486 28.982 19.093 32.787 36.99 3.804 17.898-2.444 35.575-14.908 47.128a49.289 49.289 0 0 0 26.242 1.56c13.372-2.842 24.327-10.842 31.253-21.488 6.988-10.743 9.875-24.178 6.993-37.738-2.883-13.56-10.985-24.66-21.738-31.63" fill="#CE147C" opacity=".6" />
+        <path d="M643.152 157.657c-11.019-6.86-24.66-9.432-38.212-6.053-13.264 3.307-23.933 11.685-30.484 22.566 10.99 6.592 19.476 17.402 22.83 30.853 3.353 13.451.935 26.98-5.673 37.959 10.893 6.532 24.246 8.92 37.51 5.613 13.551-3.379 24.388-12.054 30.898-23.284 6.288-10.848 8.538-24.08 5.257-37.24-3.282-13.162-11.48-23.788-22.126-30.414" fill="#CE147C" opacity=".4" />
+        <path d="M461.076 237.401l.001-.006.003-.012.009.015v.005l-.004.014-.01-.016zm44.812 23.965c.304-.576.597-1.159.88-1.749l.085-.177c.277-.585.543-1.175.799-1.772l.093-.22a50.39 50.39 0 0 0 .7-1.744l.118-.315a52.874 52.874 0 0 0 .74-2.132c.166-.517.327-1.036.479-1.562.06-.211.116-.423.174-.635.13-.475.258-.953.377-1.435.064-.265.122-.533.182-.8.077-.339.161-.673.231-1.014.022-.107.037-.214.06-.32.057-.292.107-.586.16-.88.08-.441.162-.882.23-1.323.034-.218.058-.438.09-.656.363-2.576.53-5.142.503-7.678v-.335c-.008-.477-.032-.953-.053-1.428-.012-.27-.018-.54-.034-.81-.023-.364-.057-.727-.088-1.091-.031-.387-.06-.773-.1-1.16-.027-.257-.063-.513-.094-.768a50.02 50.02 0 0 0-.196-1.488c-.023-.146-.05-.293-.074-.44-.095-.604-.2-1.208-.318-1.81-.006-.035-.014-.068-.021-.101a51.055 51.055 0 0 0-7.366-18.022c-.196-.298-.402-.584-.603-.878l-.002-.002a50.913 50.913 0 0 0-9.029-10.049l-.104-.088a49.269 49.269 0 0 0-1.93-1.554c-.078-.06-.154-.121-.233-.181a50.917 50.917 0 0 0-2.096-1.51l-.055-.037a50.578 50.578 0 0 0-2.145-1.375c-.248-.15-.503-.291-.753-.438a57.466 57.466 0 0 0-1.367-.773c-.348-.19-.7-.372-1.053-.554-.375-.19-.752-.376-1.132-.559-.388-.187-.778-.376-1.173-.554-.376-.168-.755-.324-1.135-.485a51.43 51.43 0 0 0-2.101-.83c-.386-.143-.771-.287-1.16-.42-.407-.139-.817-.27-1.23-.398-.2-.064-.4-.135-.604-.196a52.612 52.612 0 0 0-1.757-.49 51.24 51.24 0 0 0-2.256-.54c-.066-.014-.13-.033-.197-.047l-.001.007-.013-.002.002-.012c-.22-.046-.442-.077-.662-.121a54.075 54.075 0 0 0-2.577-.436 51.213 51.213 0 0 0-6.16-.5c-.194-.005-.39-.019-.586-.02-.154-.002-.307.006-.463.005-.157 0-.316.005-.475.007a51.044 51.044 0 0 0-2.685.099c-.517.033-1.031.073-1.545.122-.317.03-.634.06-.95.096a52.23 52.23 0 0 0-3.072.448c-.439.078-.877.162-1.315.25-.195.04-.389.079-.582.122-.589.127-1.175.265-1.76.415-.121.03-.243.064-.364.096-.558.146-1.113.303-1.668.469l-.241.071c-.623.192-1.244.397-1.862.612l-.27.097c-.59.21-1.177.43-1.762.662l-.113.044c-.597.24-1.19.494-1.78.758-.13.057-.259.117-.388.176-.485.223-.97.452-1.45.69-.114.056-.227.11-.339.167a49.94 49.94 0 0 0-3.114 1.72c-.366.217-.73.44-1.09.668-.277.175-.554.348-.829.53h.002l-.01.007-.005-.004-.152.105c-.65.432-1.292.875-1.923 1.337l-.022.015c-.649.477-1.286.972-1.913 1.477-.133.109-.26.221-.393.33-.477.394-.949.795-1.412 1.207-.171.151-.338.308-.508.462a50.27 50.27 0 0 0-1.222 1.155c-.084.08-.173.158-.256.24-.102.1-.195.205-.296.305a52.49 52.49 0 0 0-1.506 1.583 48.344 48.344 0 0 0-1.416 1.624 49.85 49.85 0 0 0-2.28 2.941l-.309.435c-.328.468-.648.942-.96 1.422l-.218.334c-.329.514-.649 1.037-.96 1.565-.05.081-.097.163-.146.245a49.829 49.829 0 0 0-.935 1.679 50.186 50.186 0 0 0-.98 1.938l-.083.17a49.824 49.824 0 0 0-.892 1.992 48.396 48.396 0 0 0-.819 2.06c-.206.553-.403 1.111-.59 1.673l-.149.46a52.122 52.122 0 0 0-.48 1.563c-.06.211-.117.422-.175.634-.13.477-.258.955-.376 1.438-.064.264-.12.53-.18.796-.078.34-.164.674-.234 1.016-.022.107-.036.214-.058.321-.058.292-.107.584-.16.877-.08.442-.162.885-.23 1.325-.034.217-.059.435-.089.653a51.182 51.182 0 0 0-.505 7.682v.334c.008.479.031.954.053 1.43.012.269.018.539.035.807.022.367.056.73.087 1.094.031.387.06.773.1 1.158.028.258.063.514.095.77.059.497.121.99.195 1.486.023.147.05.293.074.44a49.01 49.01 0 0 0 .339 1.912 51.026 51.026 0 0 0 7.367 18.021c.195.299.402.584.603.878a50.91 50.91 0 0 0 11.063 11.693l.235.183c.688.52 1.385 1.025 2.095 1.508l.055.038c.705.478 1.42.936 2.145 1.375.248.15.502.291.752.438.454.263.908.524 1.368.773.348.189.7.371 1.053.554.375.192.752.376 1.131.559.39.187.778.375 1.174.554.375.168.756.325 1.135.484.692.292 1.392.568 2.102.83.385.144.77.287 1.158.42.407.14.818.27 1.23.398.202.064.402.135.604.196.357.108.715.208 1.072.307.228.063.457.123.686.183.748.196 1.499.378 2.256.54.067.014.131.033.197.047l.002-.007.012.003-.002.011c.222.047.443.08.665.123a53.666 53.666 0 0 0 2.573.435 52.432 52.432 0 0 0 4.935.46c.409.018.817.032 1.225.04.195.005.392.018.586.02.156.001.31-.006.463-.006.158 0 .316-.005.475-.007a50.539 50.539 0 0 0 2.686-.098c.516-.033 1.03-.073 1.542-.121a51.625 51.625 0 0 0 3.204-.405 56.695 56.695 0 0 0 2.137-.39c.193-.04.387-.08.582-.122a47.322 47.322 0 0 0 1.759-.415c.121-.03.243-.065.364-.096a49.476 49.476 0 0 0 1.668-.47l.24-.07a53.37 53.37 0 0 0 1.866-.613c.088-.03.176-.064.264-.095.59-.21 1.18-.43 1.765-.663l.113-.044c.596-.24 1.19-.494 1.78-.758.13-.058.258-.118.387-.176.486-.222.97-.452 1.453-.691l.338-.166c.513-.26 1.022-.531 1.529-.81a48.264 48.264 0 0 0 2.67-1.577c.278-.174.556-.35.833-.53l-.002-.002.01-.007.005.006c.052-.035.1-.072.153-.108a50.434 50.434 0 0 0 1.945-1.351c.649-.477 1.285-.97 1.913-1.477.132-.108.26-.22.392-.329.476-.394.949-.795 1.412-1.206.17-.152.34-.31.509-.465.413-.377.82-.76 1.22-1.151.085-.083.174-.16.257-.241.101-.1.197-.207.296-.308a49.287 49.287 0 0 0 2.924-3.207c.314-.376.622-.757.926-1.143.136-.173.273-.346.406-.522.321-.418.635-.843.944-1.273.103-.145.209-.29.31-.437.329-.467.647-.94.96-1.42l.22-.337c.328-.513.648-1.034.959-1.56l.147-.25c.321-.55.632-1.11.933-1.675l.1-.188z" fill="#CE147C" />
+        <path d="M513.217 180.248a49.289 49.289 0 0 0-26.287.275 49.292 49.292 0 0 0-21.28 12.497c11.168 2.936 21.417 9.685 28.615 19.895 14.226 20.176 11.815 47.126-4.437 64.377a49.284 49.284 0 0 0 24.655-.681 49.29 49.29 0 0 0 22.438-13.698c11.627-12.394 16.626-30.465 11.583-48.052-5.043-17.588-18.858-30.263-35.287-34.613" fill="#CE147C" opacity=".8" />
+        <path d="M624.615-39.182c-12.506-2.382-25.922-.02-37.315 7.664a49.764 49.764 0 0 0-17.219 20.03c17.03-.986 34.122 6.763 44.348 21.923 10.226 15.16 11.007 33.911 3.713 49.331a49.764 49.764 0 0 0 25.023-8.462c11.392-7.684 18.608-19.239 21.083-31.726 2.498-12.6.17-26.149-7.578-37.635-7.747-11.486-19.437-18.72-32.055-21.125" fill="#00B2FF" opacity=".6" />
+        <path d="M683.825-78.402c-12.736-2.645-26.488-.357-38.127 7.494-11.392 7.684-18.607 19.238-21.084 31.726 12.618 2.405 24.31 9.638 32.057 21.124 7.748 11.486 10.074 25.036 7.577 37.636 12.506 2.382 25.923.02 37.315-7.664 11.639-7.85 18.912-19.744 21.23-32.544 2.24-12.365-.144-25.578-7.725-36.817-7.58-11.24-18.938-18.399-31.243-20.955" fill="#00B2FF" opacity=".4" />
+        <g transform="rotate(-34 315.164 -726.927)">
+          <path d="M50.958 51.72l.003-.005.005-.01.005.015v.006l-.008.012-.005-.018zm37.88 34.075c.437-.485.864-.98 1.284-1.483l.126-.152c.412-.5.815-1.007 1.21-1.525l.144-.19a50.37 50.37 0 0 0 1.107-1.522c.064-.092.127-.185.192-.277a52.818 52.818 0 0 0 1.24-1.89c.287-.46.571-.925.847-1.398.11-.19.216-.382.324-.574.243-.43.484-.861.716-1.3.127-.242.25-.488.373-.732.157-.31.32-.614.472-.927.047-.098.088-.198.135-.296.127-.27.248-.543.37-.814.186-.41.374-.817.547-1.228.086-.204.163-.41.247-.614a50.84 50.84 0 0 0 2.358-7.325c.026-.108.056-.215.081-.324.109-.465.202-.932.296-1.398.054-.265.114-.528.164-.793.066-.36.122-.72.18-1.08.063-.383.13-.763.184-1.148.036-.257.063-.514.095-.769.063-.496.123-.991.17-1.49.014-.147.024-.296.036-.444.054-.61.099-1.22.13-1.833l.004-.103a50.753 50.753 0 0 0-2.802-19.257c-.119-.336-.25-.663-.374-.997-.002 0-.002-.001-.002-.003a50.786 50.786 0 0 0-6.363-11.927l-.08-.11a49.268 49.268 0 0 0-1.504-1.974c-.062-.079-.121-.156-.183-.233a50.963 50.963 0 0 0-1.678-1.97l-.044-.05a50.67 50.67 0 0 0-1.759-1.852c-.205-.205-.419-.404-.627-.606a57.627 57.627 0 0 0-1.146-1.08 55.6 55.6 0 0 0-.892-.792 56.098 56.098 0 0 0-.968-.816c-.333-.276-.667-.553-1.01-.821-.325-.254-.657-.497-.988-.744-.605-.45-1.22-.888-1.848-1.313-.342-.233-.682-.465-1.028-.688a48.39 48.39 0 0 0-1.103-.683c-.181-.111-.358-.228-.542-.337a52.965 52.965 0 0 0-1.595-.9c-.683-.37-1.37-.728-2.07-1.069-.06-.03-.118-.063-.18-.093l-.003.007-.012-.006.005-.01c-.204-.099-.412-.183-.616-.279a54.512 54.512 0 0 0-2.407-1.045 51.673 51.673 0 0 0-5.886-1.975c-.19-.051-.377-.112-.568-.16-.15-.04-.301-.07-.452-.107-.155-.038-.31-.072-.465-.108a51.567 51.567 0 0 0-2.644-.554 49.9 49.9 0 0 0-1.537-.255c-.316-.048-.632-.095-.948-.137a52.794 52.794 0 0 0-3.106-.308A55.551 55.551 0 0 0 53.1.84c-.2-.008-.398-.018-.597-.023a47.951 47.951 0 0 0-1.818-.023c-.125 0-.252.004-.378.005-.58.007-1.16.025-1.74.052l-.254.01a53.59 53.59 0 0 0-1.965.144l-.286.028a52.69 52.69 0 0 0-1.88.216l-.12.016c-.641.088-1.282.19-1.921.304-.14.024-.281.051-.422.077-.527.099-1.055.204-1.582.319-.125.027-.248.051-.37.08a50.442 50.442 0 0 0-3.457.914c-.409.123-.819.251-1.225.385-.313.103-.625.204-.937.313l.001.002c-.004 0-.008.002-.012.004l-.002-.006c-.059.02-.115.045-.175.065-.74.262-1.473.536-2.2.832l-.026.01c-.749.305-1.49.63-2.224.97-.157.072-.309.15-.464.223-.56.268-1.119.542-1.67.83-.205.105-.406.217-.609.325-.496.268-.986.54-1.473.824-.101.058-.207.112-.308.172-.123.071-.24.151-.362.224A52.813 52.813 0 0 0 22.77 9.3a48.608 48.608 0 0 0-1.776 1.233 50.087 50.087 0 0 0-2.939 2.301l-.408.347c-.434.374-.86.757-1.282 1.147l-.294.272c-.446.418-.885.848-1.317 1.284l-.202.203c-.448.457-.888.925-1.32 1.402a50.238 50.238 0 0 0-1.427 1.642l-.122.146a49.84 49.84 0 0 0-1.354 1.715c-.379.5-.749 1.01-1.11 1.526l-.19.275c-.336.486-.664.979-.983 1.479l-.257.41c-.289.462-.573.927-.848 1.4-.11.19-.217.38-.325.572-.243.43-.484.863-.716 1.303-.127.24-.247.485-.37.728-.159.311-.323.615-.475.93-.047.098-.088.198-.135.297-.127.269-.246.54-.369.812-.185.409-.373.818-.547 1.229-.085.202-.162.408-.245.612A50.94 50.94 0 0 0 1.4 39.893l-.081.324c-.108.466-.202.932-.296 1.4-.054.263-.114.526-.162.79-.068.361-.123.722-.181 1.083-.064.382-.13.763-.184 1.146-.035.257-.064.514-.095.77-.063.496-.123.99-.171 1.489-.014.147-.023.295-.036.444a48.702 48.702 0 0 0-.133 1.935 50.725 50.725 0 0 0 2.803 19.258c.117.336.25.663.374.997l.001.002a50.782 50.782 0 0 0 7.946 14.012l.185.234a53.2 53.2 0 0 0 1.676 1.969l.045.05a50.67 50.67 0 0 0 1.758 1.852c.205.205.419.404.627.606.378.365.758.728 1.146 1.08a50.609 50.609 0 0 0 1.86 1.607c.334.276.667.553 1.01.822.325.254.658.499.989.744.604.45 1.22.888 1.849 1.313.34.232.68.465 1.027.688.363.233.732.459 1.103.683.18.11.36.228.542.336.321.19.646.375.97.557.207.116.416.23.624.343.683.371 1.37.73 2.07 1.07.062.03.12.063.18.093l.005-.007a.027.027 0 0 0 .01.005l-.005.011c.206.099.414.184.62.28a54.101 54.101 0 0 0 2.403 1.044 52.903 52.903 0 0 0 4.702 1.64c.394.116.788.228 1.184.335.19.05.379.112.568.16.151.04.303.07.453.107.154.038.31.072.465.108a51.057 51.057 0 0 0 2.643.554 49.73 49.73 0 0 0 2.488.391 52.19 52.19 0 0 0 2.27.246 57.326 57.326 0 0 0 2.178.138c.199.008.398.018.598.023.604.019 1.21.027 1.816.023.126 0 .253-.004.378-.005.58-.007 1.16-.025 1.741-.052l.253-.01a53.957 53.957 0 0 0 1.968-.144c.093-.008.188-.019.28-.028.628-.06 1.256-.131 1.884-.216l.12-.015c.64-.09 1.281-.192 1.922-.305l.419-.077a50.637 50.637 0 0 0 1.585-.319c.123-.027.247-.051.37-.08a50.367 50.367 0 0 0 3.456-.914 55.5 55.5 0 0 0 1.22-.384c.314-.102.628-.205.941-.313l-.001-.003.012-.004.003.007c.06-.022.116-.046.176-.068a50.88 50.88 0 0 0 2.225-.84c.75-.305 1.49-.63 2.225-.97.155-.072.307-.15.462-.224.561-.267 1.12-.541 1.67-.828.205-.106.408-.218.61-.328.495-.266.986-.539 1.47-.821.103-.06.209-.112.31-.172.123-.072.242-.152.364-.226a49.591 49.591 0 0 0 4.813-3.288c.174-.135.35-.27.523-.408.415-.328.824-.664 1.23-1.006.136-.116.274-.231.409-.349.434-.374.86-.756 1.281-1.145l.296-.274c.445-.417.884-.846 1.315-1.28l.204-.207a49.72 49.72 0 0 0 1.461-1.557z" fill="#00B2FF" />
+        </g>
+        <path d="M570.082-11.487a49.764 49.764 0 0 0-25.023 8.461 49.776 49.776 0 0 0-16.483 18.529c11.57-.691 23.435 2.533 33.45 10 19.792 14.759 25.776 41.15 15.558 62.633a49.767 49.767 0 0 0 23.34-8.34 49.765 49.765 0 0 0 17.218-20.031c7.294-15.419 6.512-34.17-3.713-49.33-10.225-15.16-27.318-22.909-44.347-21.922" fill="#00B2FF" opacity=".8" />
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgCirclesHeroMobileLandscape };
