+++ assets/components/svgs/creditCard.jsx
@@ -1,0 +1,21 @@
+// @flow
+
+import React from 'react';
+
+// The generic credit card logo used on checkouts.
+function SvgCreditCard() {
+
+  return (
+    <svg
+      className="svg-credit-card"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 44 28.1"
+      preserveAspectRatio="xMinYMid"
+    >
+      <path d="M0 10.1v15l3 3h38.1l2.9-3.2V10.1M44 6.1V3l-2.9-3H3L0 3v3.1" />
+    </svg>
+  );
+
+}
+
+export { SvgCreditCard };
