+++ assets/components/svgs/cross.jsx
@@ -1,0 +1,18 @@
+// @flow
+
+import React from 'react';
+
+function SvgCross() {
+  return (
+    <svg
+      className="svg-cross"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="-2 -2 38 38"
+      preserveAspectRatio="xMinYMid"
+    >
+      <path d="M16.525 19.87L30.865 33 33 30.87 19.9 16.5 33 2.13 30.866 0l-14.34 13.13L2.133.05 0 2.18 13.15 16.5 0 30.82l2.134 2.13" fillRule="evenodd" />
+    </svg>
+  );
+}
+
+export { SvgCross };
