+++ assets/components/svgs/directDebitSymbol.jsx
@@ -1,0 +1,21 @@
+// @flow
+
+import React from 'react';
+
+function SvgDirectDebitSymbol() {
+  return (
+    <svg
+      className="svg-direct-debit-symbol"
+      width="35"
+      height="26"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M18.254 9.655v13.73s-7.444-1.436-7.444-6.893c0-4.353 5.51-6.272 7.444-6.837zm-14.618 6.55c0 3.2 5.371 7.878 14.119 8 .345 0 10.282-.41 10.129-11.817-.129-9.582-6.834-11.26-8.978-11.552V.252c8.548.58 15.27 6.078 15.27 12.75 0 7.06-7.52 12.804-16.766 12.804C8.163 25.806.68 20.06.68 13.003.681 5.945 8.163.2 17.41.2c.283 0 .565.005.844.016V9.4c-4.474-.346-14.618 1.897-14.618 6.805z"
+        fill="#212121"
+      />
+    </svg>
+  );
+}
+
+export { SvgDirectDebitSymbol };
