+++ assets/components/svgs/directDebitSymbolAndText.jsx
@@ -1,0 +1,21 @@
+// @flow
+
+import React from 'react';
+
+function SvgDirectDebitSymbolAndText() {
+  return (
+    <svg className="svg-direct-debit-symbol-and-text" width="86" height="28" xmlns="http://www.w3.org/2000/svg">
+      <g>
+        <g
+          stroke="null"
+          clipRule="evenodd"
+          fillRule="evenodd"
+        >
+          <path d="M40.13 6.054h4.382s4.34-.088 4.295 3.856c-.043 2.936-4.207 3.375-4.207 3.375h-4.47V6.054zM53.24 20.21c.124.573.538 1.205 1.922 1.27 1.84.044 2.893-.658 2.893-.658v1.359s-2.016.482-2.98.482c-1.008 0-3.989-.22-3.989-2.98 0-2.235 2.455-2.936 3.857-2.936 2.454.087 3.637.964 3.594 3.462 0 .036-4.039.01-5.296 0zm.006-1.097h3.363s0-1.27-1.622-1.27c-1.22 0-1.62.657-1.741 1.27zm10.638-1.373c.486-.527 1.303-.95 2.674-.95 2.717 0 3.155 2.279 3.111 2.893-.044.613-.263 2.936-3.418 2.936-1.191 0-1.92-.406-2.366-.911v.736h-2.148v-7.889h2.148v3.185h-.001zm1.84 3.652c1.272-.044 1.841-1.14 1.841-1.665 0-.527-.307-1.622-1.928-1.622-1.259 0-1.641.95-1.753 1.415v.572c.11.473.495 1.3 1.84 1.3zm18.54-.219l.043 1.227s-1.008.22-1.665.22c-.658 0-2.367-.395-2.367-2.06v-2.543H78.48V16.88h1.796V15.52h2.192v1.36h1.796v1.138h-1.796v2.192s0 1.096.876 1.096.92-.132.92-.132zM77.601 7.237h2.541v6.092h2.148V7.237h2.586V5.966h-7.275v1.27zm-.877.307l.263-1.272s-1.095-.482-3.024-.482c-1.884 0-4.909 1.052-4.909 3.9 0 2.806 3.2 3.77 5.172 3.77 1.973 0 2.63-.306 2.63-.306l-.219-1.403s-.876.263-2.367.263c-1.446-.044-2.936-1.183-2.936-2.454.044-1.14 1.359-2.41 3.068-2.41 1.314 0 2.322.394 2.322.394zM57.353 6.01s3.595-.176 3.595 1.928c0 1.797-2.06 1.929-2.06 1.929s.57.088.832.482c.22.35 1.666 2.98 1.666 2.98h-2.192s-.35-.438-.526-.745c-.13-.306-.526-.964-.526-.964s-.307-1.095-1.358-1.095c-.027 0-.456-.017-.79-.03v2.835h-2.103V6.01h3.462zm-1.358 3.287h1.14s1.577-.088 1.665-1.052c.044-.921-1.534-1.052-1.534-1.052h-1.27v2.104zm8.81 1.14h3.505v-1.36h-3.506V7.412h3.813V6.009h-5.96v7.32h6.004v-1.315h-3.856v-1.578zm8.545 5.609h2.148v-1.49H73.35v1.49zm0 6.399h2.148v-5.566H73.35v5.566zm-23.403-9.116h2.104v-7.32h-2.104v7.32zM40.13 15.17h4.382s4.34-.088 4.295 3.856c-.043 2.937-4.207 3.375-4.207 3.375h-4.47V15.17zm2.147 5.873h1.402s2.893-.176 2.893-2.41c0-2.236-3.068-2.455-3.068-2.455h-1.227v4.865zm0-9.116h1.402s2.893-.22 2.893-2.411c0-2.28-3.068-2.455-3.068-2.455h-1.227v4.866zM20.451 10.41v14.664s-8.503-1.533-8.503-7.362c0-4.65 6.295-6.698 8.503-7.301zm0-.273c-5.11-.37-16.698 2.026-16.698 7.268 0 3.418 6.136 8.415 16.128 8.546.395 0 11.746-.439 11.57-12.622C31.306 3.096 23.647 1.303 21.197.99V.367c9.764.62 17.443 6.492 17.443 13.62 0 7.538-8.59 13.673-19.152 13.673C8.925 27.66.378 21.524.378 13.986S8.924.313 19.487.313c.323 0 .644.005.964.017v9.807z" />
+        </g>
+      </g>
+    </svg>
+  );
+}
+
+export { SvgDirectDebitSymbolAndText };
