+++ assets/components/svgs/exclamation.jsx
@@ -1,0 +1,21 @@
+// @flow
+
+import React from 'react';
+
+// Exclamation mark, used on error messages.
+function SvgExclamation() {
+
+  return (
+    <svg
+      className="svg-exclamation"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 2.2 8.9"
+      preserveAspectRatio="xMinYMid"
+    >
+      <path d="M1.1 8.9C.5 8.9 0 8.4 0 7.8c0-.6.5-1.1 1.1-1.1.6 0 1.1.5 1.1 1.1 0 .6-.5 1.1-1.1 1.1zM.6 0h1.1l.4.4-.4 5.2H.6L.2.4.6 0z" />
+    </svg>
+  );
+
+}
+
+export { SvgExclamation };
