+++ assets/components/svgs/facebook.jsx
@@ -1,0 +1,16 @@
+// @flow
+
+import React from 'react';
+
+// The Facebook logo, in monochrome.
+function SvgFacebook() {
+
+  return (
+    <svg className="svg-facebook" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" preserveAspectRatio="xMinYMid">
+      <path d="M10.897 8.371h-2.229v5.943h-2.154v-5.943h-1.485v-2.154h1.485v-1.783c0-1.412.817-2.748 2.971-2.748.892 0 1.485.074 1.485.074v2.229h-1.337c-.743 0-.892.371-.892.966v1.337h2.229l-.074 2.08z" />
+    </svg>
+  );
+
+}
+
+export { SvgFacebook };
