+++ assets/components/svgs/globe.jsx
@@ -1,0 +1,30 @@
+// @flow
+
+import React from 'react';
+
+// The Globe
+function SvgGlobe() {
+
+  return (
+    <svg className="svg-globe" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 96 96" preserveAspectRatio="xMinYMid">
+      <path d="M23.7,22.4 L45.8,22.4 L45.8,0.5 C40.7,0.7 35.6,1.8 30.8,3.7 C29.3,6 28,8.5 27,11.2 C25.6,14.8 24.5,18.6 23.7,22.4" />
+      <path d="M22,9.2 C22.2,8.7 22.4,8.2 22.6,7.7 C16.6,11.4 11.6,16.4 7.8,22.4 L18.2,22.4 C19.1,17.9 20.4,13.5 22,9.2" />
+      <path d="M15.8,50.7 L0.5,50.7 C0.8,56.8 2.3,62.7 4.9,68.2 L17.2,68.2 C16.3,62.4 15.9,56.5 15.8,50.7" />
+      <path d="M17.2,27.8 L4.9,27.8 C2.3,33.3 0.8,39.2 0.5,45.3 L15.8,45.3 C15.9,39.4 16.4,33.6 17.2,27.8" />
+      <path d="M78.8,22.4 L88.2,22.4 C84.7,16.9 80.1,12.2 74.7,8.5 C74.8,8.7 74.9,9 75,9.2 C76.6,13.5 77.9,17.9 78.8,22.4" />
+      <path d="M81.2,45.3 L95.5,45.3 C95.2,39.2 93.7,33.3 91.1,27.8 L79.7,27.8 C80.6,33.6 81.1,39.4 81.2,45.3" />
+      <path d="M73.3,22.4 C72.5,18.6 71.4,14.8 70,11.2 C69.1,8.7 67.9,6.4 66.5,4.1 C61.6,2.1 56.4,0.9 51.2,0.6 L51.2,22.5 L73.3,22.5 L73.3,22.4 Z" />
+      <path d="M51.2,68.2 L74.3,68.2 C75.2,62.4 75.7,56.6 75.8,50.7 L51.2,50.7 L51.2,68.2 Z" />
+      <path d="M79.7,68.2 L91.1,68.2 C93.7,62.7 95.2,56.8 95.5,50.7 L81.2,50.7 C81.1,56.5 80.6,62.4 79.7,68.2" />
+      <path d="M73.3,73.5 L51.2,73.5 L51.2,95.4 C56.5,95.1 61.6,93.9 66.5,91.9 C67.9,89.7 69.1,87.3 70,84.8 C71.4,81.1 72.5,77.4 73.3,73.5" />
+      <path d="M23.7,73.5 C24.5,77.3 25.6,81.1 27,84.8 C28,87.5 29.3,90 30.8,92.4 C35.6,94.3 40.7,95.3 45.8,95.6 L45.8,73.6 L23.7,73.5 Z" />
+      <path d="M45.8,27.8 L22.6,27.8 C21.7,33.6 21.2,39.4 21.1,45.3 L45.8,45.3 L45.8,27.8 Z" />
+      <path d="M75,86.7 C74.9,87 74.8,87.2 74.7,87.4 C80.1,83.7 84.7,79 88.2,73.5 L78.8,73.5 C77.9,78 76.6,82.4 75,86.7" />
+      <path d="M75.8,45.3 C75.7,39.4 75.2,33.6 74.3,27.8 L51.2,27.8 L51.2,45.3 L75.8,45.3 Z" />
+      <path d="M18.2,73.5 L7.8,73.5 C11.6,79.4 16.6,84.5 22.6,88.2 C22.4,87.7 22.1,87.2 22,86.7 C20.4,82.4 19.1,78 18.2,73.5" />
+      <path d="M21.1,50.7 C21.2,56.6 21.7,62.4 22.6,68.2 L45.7,68.2 L45.7,50.7 L21.1,50.7 Z" />
+    </svg>
+  );
+}
+
+export { SvgGlobe };
