+++ assets/components/svgs/guardianLogo.jsx
@@ -1,0 +1,16 @@
+// @flow
+
+import React from 'react';
+
+// The main guardian logo.
+function SvgGuardianLogo() {
+
+  return (
+    <svg className="svg-guardian-logo" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 324.999 104.42" preserveAspectRatio="xMinYMid">
+      <path fill="#121212" d="M72.381 55.621l5.539-2.883V9.17h-4.188L63.493 22.755h-1.151l.646-15.138h44.387l.645 15.138h-1.222L96.772 9.17h-4.295v43.474l5.576 2.943v1.469H72.381zM113.322 53.693V5.418l-4.303-1.721v-.899L124.587 0h1.642v22.842l.43-.365c3.442-3.012 8.389-4.948 13.337-4.948 6.813 0 9.824 3.845 9.824 11.017v25.147l3.639 1.982.002 1.463h-20.348v-1.455l3.657-1.989v-25.23c0-3.944-1.722-5.521-4.948-5.521-2.151 0-4.006.674-5.378 1.776v29.043l3.585 1.98v1.396h-20.361V55.75l3.654-2.057zM165.513 38.869c.422 7.979 3.996 14.154 12.469 14.154 4.094 0 7.002-1.898 9.742-3.343v1.571c-2.117 2.894-7.484 6.963-14.968 6.963-13.131 0-19.838-7.3-19.838-19.938 0-12.354 7.343-20.048 19.203-20.048 11.155 0 16.944 5.576 16.944 20.261v.379h-23.552zm-.213-1.838l11.551-.706c0-9.885-1.693-16.445-5.082-16.445-3.601-.001-6.469 7.62-6.469 17.151M0 76.391c0-21.012 13.907-28.493 29.401-28.493 6.575 0 12.773 1.058 16.249 2.494l.303 14.663h-1.478l-9.096-14.188c-1.558-.669-3.035-.931-5.756-.931-8.239 0-12.447 9.515-12.315 25.118.157 18.669 3.401 27.142 10.959 27.142 1.965 0 3.477-.303 4.535-.756V81.454l-4.988-2.858V76.92h24.111v1.785l-4.915 2.749v19.728c-4.08 1.588-10.958 3.099-18.214 3.099C11.186 104.28 0 96.118 0 76.391M50.87 66.585v-1.212l16.219-2.859 1.777.148v31.951c0 3.852 1.852 5.035 4.962 5.035 1.999 0 3.81-.761 5.258-2.482V68.51l-4.444-1.925v-1.259l16.219-2.859 1.63.147v36.624l4.369 1.835v1.165l-15.997 1.967-1.629-.148v-4.812h-.445c-2.962 2.739-7.109 5.109-12.145 5.109-7.776 0-11.331-4.592-11.331-11.553V68.51l-4.443-1.925zM154.449 62.446l1.333.147-.004 11.828h.37c1.732-8.672 5.554-11.907 10.22-11.907.741 0 1.555.073 1.999.298v12.1c-.741-.222-2.073-.296-3.332-.296-3.703 0-6.425.669-8.813 1.761l.004 23.378 3.687 2.046.001 1.518h-21.027v-1.503l3.787-2.056V67.862l-4.444-1.325v-1.083l16.219-3.008zM196.163 63.443V50.967l-4.443-1.556v-.999l16.367-3.009 1.555.223v53.596l4.518 1.64v1.381l-16.145 2.178-1.258-.148v-4.435h-.371c-2.369 2.368-5.629 4.517-10.737 4.517-8.813 0-15.257-6.739-15.257-20.513 0-14.518 7.479-21.657 18.811-21.657 3.257-.001 5.701.592 6.96 1.258m-.023 34.315V65.72c-1.036-.666-1.786-1.489-4.473-1.396-4.377.155-7.08 6.764-7.08 18.539 0 10.589 1.943 16.509 7.776 16.314 1.629-.054 2.963-.637 3.777-1.419M232.846 62.409l1.407.149v37.197l3.697 2.046.002 1.517h-21.027v-1.504l3.775-2.055V68.436l-4.518-1.777v-1.241l16.664-3.009zm1.482-10.043c0 3.926-3.333 6.888-7.185 6.888-3.998 0-7.109-2.962-7.109-6.888 0-3.925 3.111-6.962 7.109-6.962 3.852 0 7.185 3.037 7.185 6.962M283.532 99.759V68.092l-4.443-1.555v-1.536l16.145-3.008 1.629.147v4.74h.444c3.481-3.11 8.666-5.11 13.774-5.11 7.037 0 10.146 3.332 10.146 10.738v27.188l3.771 2.103v1.517h-21.026v-1.503l3.776-2.055V73.25c0-4.074-1.776-5.702-5.109-5.702-2.148 0-3.919.544-5.555 1.767v30.433l3.703 2.053v1.517h-21.035v-1.503l3.78-2.056zM259.661 79.857v-5.311c0-8.003-1.745-10.622-6.693-10.622-.58 0-1.09.073-1.673.146l-8.803 11.931h-1.235V65.017c3.782-1.164 8.51-2.547 14.767-2.547 10.768 0 17.024 2.983 17.024 12.004v25.899l3.855 1.019v1.018c-1.528.946-4.584 1.819-7.93 1.819-5.311 0-7.857-1.746-9.021-4.656h-.363c-2.256 3.055-5.457 4.802-10.476 4.802-6.402 0-10.767-4.001-10.767-10.913 0-6.692 4.146-10.331 12.586-11.931l8.729-1.674zm0 17.896V81.75l-2.691.217c-4.221.364-5.748 3.057-5.748 9.021 0 6.475 2.111 8.147 5.094 8.147 1.673 0 2.619-.509 3.345-1.382M119.175 79.857v-5.311c0-8.003-1.745-10.622-6.693-10.622-.581 0-1.09.073-1.673.146l-8.803 11.931h-1.236V65.017c3.783-1.164 8.511-2.547 14.767-2.547 10.768 0 17.024 2.983 17.024 12.004v25.899l3.856 1.019v1.018c-1.528.946-4.584 1.819-7.93 1.819-5.31 0-7.857-1.746-9.021-4.656h-.364c-2.255 3.055-5.456 4.802-10.475 4.802-6.402 0-10.767-4.001-10.767-10.913 0-6.692 4.146-10.331 12.586-11.931l8.729-1.674zm0 17.896V81.75l-2.691.217c-4.221.364-5.748 3.057-5.748 9.021 0 6.475 2.111 8.147 5.093 8.147 1.672 0 2.619-.509 3.346-1.382" />
+    </svg>
+  );
+
+}
+
+export { SvgGuardianLogo };
