+++ assets/components/svgs/lock.jsx
@@ -1,0 +1,16 @@
+// @flow
+
+import React from 'react';
+
+// A padlock icon used to signify that checkouts are secure.
+function SvgLock() {
+
+  return (
+    <svg className="svg-lock" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 9 10.57" preserveAspectRatio="xMinYMid">
+      <path d="M8.17 3.68C8.17 1.65 6.52 0 4.5 0S.83 1.65.83 3.68H0V9.3l1.28 1.28h6.45L9 9.3V3.68h-.83zM4.5 1.57c1.12 0 2.1.9 2.1 2.1H2.4c0-1.12.98-2.1 2.1-2.1z" />
+    </svg>
+  );
+
+}
+
+export { SvgLock };
