+++ assets/components/svgs/mastercardLogo.jsx
@@ -1,0 +1,26 @@
+// @flow
+
+import React from 'react';
+
+// The mastercard logo.
+function SvgMastercardLogo() {
+
+  return (
+    <svg
+      className="svg-mastercard-logo"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 36 24"
+      preserveAspectRatio="xMinYMid"
+    >
+      <g fill="none" fillRule="evenodd">
+        <rect fill="#333" width="35.25" height="23.25" rx="2.25" />
+        <path d="M10.748 20.047V18.9c.017-.193-.053-.384-.193-.523-.14-.138-.333-.212-.533-.2-.265-.017-.52.108-.662.325-.13-.21-.37-.336-.623-.325-.225-.013-.44.088-.568.267v-.223h-.39v1.83h.383V19c-.014-.13.03-.257.122-.35.092-.095.222-.144.355-.137.272 0 .426.172.426.482v1.054h.382V19c-.013-.13.032-.257.124-.35.092-.094.22-.144.354-.137.28 0 .43.172.43.482v1.054l.392-.003zm6.145-1.826h-.687v-.554h-.384v.555h-.395v.333h.395v.867c0 .425.154.677.64.677.18 0 .358-.048.51-.14l-.117-.327c-.11.064-.235.1-.363.104-.197 0-.288-.123-.288-.306v-.878h.687l.003-.33zm3.502-.045c-.208-.005-.402.097-.513.267v-.22h-.375v1.83h.38v-1.034c0-.302.15-.497.417-.497.088 0 .175.015.256.047l.127-.346c-.094-.032-.193-.05-.292-.05v.005zm-5.295.19c-.232-.133-.5-.2-.77-.19-.476 0-.768.218-.768.574 0 .297.213.477.635.534l.197.027c.225.032.347.108.347.21 0 .143-.164.235-.445.235-.225.007-.446-.06-.627-.192l-.197.287c.242.16.527.242.818.238.544 0 .846-.245.846-.593 0-.35-.23-.48-.64-.538l-.197-.027c-.177-.023-.335-.076-.335-.19 0-.116.15-.227.375-.227.208.003.41.057.59.157l.173-.304zm5.72.766c-.014.258.085.51.273.69.188.184.446.28.712.268.246.012.488-.066.678-.218l-.197-.287c-.142.104-.315.162-.493.164-.33-.03-.58-.3-.58-.62 0-.322.25-.59.58-.62.178 0 .35.058.493.163l.197-.287c-.19-.152-.432-.23-.678-.218-.266-.013-.524.084-.712.266-.188.183-.287.434-.274.692v.01zm-2.656-.955c-.552 0-.932.382-.932.956-.017.255.078.504.26.686.184.182.438.28.7.272.28.01.557-.082.773-.257l-.197-.27c-.156.12-.35.188-.548.19-.287.01-.534-.192-.57-.468h1.41v-.153c0-.575-.36-.957-.89-.957l-.006.002zm0 .34c.13-.004.256.043.35.13.093.087.146.207.147.333h-1.01c.015-.266.246-.472.52-.463h-.006zm-5.11.618v-.912h-.377v.222c-.15-.177-.376-.277-.612-.267-.547 0-.99.43-.99.96 0 .532.443.962.99.962.236.01.463-.09.612-.268v.22h.378v-.917zm-1.527 0c-.013-.16.043-.318.155-.436.113-.118.27-.184.436-.184.33.03.582.3.582.62 0 .322-.252.59-.582.622-.165 0-.323-.066-.436-.184-.112-.117-.168-.275-.155-.435v-.002zm14.57-.956c-.207-.005-.4.097-.512.267v-.223h-.375v1.83h.38V19.02c0-.302.15-.497.417-.497.088 0 .175.015.256.047l.127-.345c-.094-.032-.193-.05-.292-.05v.005zm3.042 1.61c.024 0 .05.005.073.014.023.01.044.023.062.04.017.015.03.035.04.056.02.045.02.097 0 .142-.01.02-.023.04-.04.057-.018.017-.04.03-.062.04-.023.01-.05.014-.074.014-.078-.002-.147-.045-.18-.112-.02-.045-.02-.096 0-.142.01-.02.025-.04.042-.057.018-.017.04-.03.062-.04.023-.01.048-.017.075-.018v.005zm0 .324c.018 0 .038-.003.056-.01.017-.008.032-.02.045-.032.052-.054.052-.137 0-.19-.012-.014-.027-.024-.044-.03-.018-.01-.038-.013-.057-.013-.02 0-.04.004-.06.012-.016.007-.032.017-.046.03-.05.054-.05.137 0 .19.014.015.03.025.047.032.02.005.04.008.06.007v.004zm.01-.227c.02 0 .037.004.052.015.012.01.018.025.017.04 0 .013-.005.026-.014.035-.012.01-.026.016-.042.017l.057.063h-.044l-.053-.063h-.018v.063h-.037v-.168l.082-.002zm-.043.032v.046h.043c.008.004.016.004.024 0 .003-.004.003-.01 0-.016.003-.005.003-.012 0-.017-.008-.002-.016-.002-.024 0l-.043-.012zm-4.463-.778v-.916h-.378v.222c-.15-.177-.376-.277-.612-.267-.547 0-.99.43-.99.96 0 .532.443.962.99.962.236.01.463-.09.612-.268v.22h.378v-.913zm-1.528 0c-.013-.16.043-.318.156-.436.112-.117.27-.183.436-.183.33.03.58.3.58.62 0 .322-.25.59-.58.622-.166 0-.324-.068-.436-.185-.113-.118-.17-.276-.156-.436v-.003zm5.338 0v-1.655h-.378v.957c-.15-.18-.376-.28-.612-.27-.547 0-.99.43-.99.962 0 .53.443.96.99.96.236.01.463-.09.612-.266v.22h.378v-.91zm-1.528 0c-.013-.16.043-.318.156-.436.112-.117.27-.183.436-.183.315.047.547.31.545.62-.002.31-.238.57-.553.614-.165 0-.323-.065-.436-.182-.112-.118-.17-.275-.156-.435l.008.004z" fill="#FFF" fillRule="nonzero" />
+        <path fill="#FF5F00" fillRule="nonzero" d="M14.44 4.15h6.27v11.047h-6.27z" />
+        <path d="M14.893 9.672c0-2.163.988-4.205 2.683-5.54-2.88-2.27-7.015-1.94-9.5.76-2.484 2.7-2.484 6.862 0 9.562 2.485 2.7 6.62 3.03 9.5.76-1.696-1.336-2.685-3.38-2.683-5.542z" fill="#EB001B" fillRule="nonzero" />
+        <path d="M28.285 13.94v-.222h.084v-.045h-.214v.045h.084v.22h.045zm.414 0v-.267h-.066l-.075.19-.076-.19h-.065v.266h.044v-.203l.07.174h.048l.07-.173v.202h.047zM28.94 9.674c0 2.698-1.536 5.16-3.955 6.338-2.418 1.18-5.295.87-7.41-.8 1.695-1.335 2.684-3.377 2.684-5.54 0-2.16-.99-4.204-2.684-5.54 2.114-1.667 4.99-1.978 7.41-.798 2.418 1.18 3.953 3.64 3.953 6.338v.002z" fill="#F79E1B" fillRule="nonzero" />
+      </g>
+    </svg>
+  );
+}
+
+export { SvgMastercardLogo };
