+++ assets/components/svgs/payPalLogo.jsx
@@ -1,0 +1,26 @@
+// @flow
+
+import React from 'react';
+
+// The standard PayPal logo.
+function SvgPaypalLogo() {
+
+  return (
+    <svg
+      className="svg-paypal-logo"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 37 24"
+      preserveAspectRatio="xMinYMid"
+    >
+      <g fillRule="nonzero" fill="none">
+        <path d="M34.823.7H2.177C.975.7 0 1.685 0 2.9v18.83c0 1.215.975 2.2 2.177 2.2h32.646c1.202 0 2.177-.985 2.177-2.2V2.9c0-1.215-.975-2.2-2.177-2.2z" fill="#F8F8F8" />
+        <path d="M24.775 9.563c-.31-.368-.86-.563-1.59-.563h-2.183c-.15 0-.276.11-.3.26l-.882 5.653c-.017.11.068.212.18.212h1.12c.104 0 .193-.077.21-.18l.25-1.604c.023-.148.15-.258.3-.258h.69c1.437 0 2.267-.703 2.483-2.096.098-.61.004-1.088-.278-1.423zm-1.338 1.502c-.12.79-.718.79-1.297.79h-.33l.233-1.476c.013-.09.09-.156.18-.156h.15c.394 0 .766 0 .958.227.115.136.15.337.107.615zM30.176 10.85H29.1c-.09 0-.17.07-.183.16l-.048.302-.076-.11c-.233-.34-.752-.452-1.27-.452-1.19 0-2.203.903-2.4 2.17-.103.63.042 1.235.4 1.656.328.388.796.55 1.354.55.957 0 1.488-.618 1.488-.618l-.048.3c-.018.114.07.217.185.217h.97c.152 0 .283-.113.307-.265l.58-3.692c.02-.113-.07-.217-.184-.217zm-1.5 2.1c-.102.617-.59 1.03-1.212 1.03-.312 0-.562-.1-.723-.29-.157-.19-.217-.458-.167-.758.097-.61.592-1.038 1.205-1.038.305 0 .553.102.717.294.165.194.23.464.182.763zM32.79 9h-1.174c-.103 0-.192.066-.208.155l-1.042 5.758c-.02.11.08.212.21.212h1.046c.174 0 .322-.11.348-.26l1.027-5.653c.02-.11-.08-.212-.208-.212z" fill="#179BD7" />
+        <path d="M8.957 9.563C8.647 9.195 8.097 9 7.367 9H5.184c-.15 0-.276.11-.3.26l-.882 5.653c-.017.11.068.212.18.212h1.042c.15 0 .276-.11.3-.26l.238-1.524c.023-.148.15-.258.3-.258h.69c1.437 0 2.267-.703 2.483-2.096.098-.61.004-1.088-.278-1.423zm-1.34 1.502c-.118.79-.716.79-1.295.79h-.33l.232-1.476c.014-.09.09-.156.18-.156h.15c.394 0 .766 0 .958.227.115.136.15.337.106.615zM14.36 10.85h-1.077c-.09 0-.17.07-.184.16l-.05.302-.074-.11c-.233-.34-.752-.452-1.27-.452-1.19 0-2.203.903-2.4 2.17-.104.63.042 1.235.4 1.656.327.388.796.55 1.354.55.957 0 1.488-.618 1.488-.618l-.048.3c-.018.114.07.217.185.217h.97c.152 0 .282-.113.307-.265l.58-3.692c.02-.113-.07-.217-.183-.217zm-1.5 2.1c-.104.617-.592 1.03-1.214 1.03-.312 0-.56-.1-.722-.29-.16-.19-.22-.458-.168-.758.096-.61.592-1.038 1.205-1.038.306 0 .554.102.718.294.164.194.23.464.182.763z" fill="#253B80" />
+        <path d="M19.628 10.75h-1.096c-.105 0-.203.048-.263.128l-1.512 2.052-.64-1.972c-.042-.123-.165-.208-.305-.208h-1.077c-.13 0-.222.118-.18.232l1.207 3.265-1.135 1.477c-.09.116 0 .276.155.276h1.094c.104 0 .202-.047.26-.125l3.647-4.85c.087-.116-.002-.275-.156-.275z" fill="#253B80" />
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgPaypalLogo };
