+++ assets/components/svgs/payWallMobile.jsx
@@ -1,0 +1,31 @@
+// @flow
+
+import React from 'react';
+
+function SvgPaywallMobile() {
+
+  return (
+    <svg
+      className="svg-paywall-mobile"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 375 362"
+      preserveAspectRatio="xMinYMid"
+      aria-labelledby="title"
+    >
+      <title>We haven&#39;t put up a paywall</title>
+      <g fill="none" fillRule="evenodd" transform="translate(-233)">
+        <path fill="#121212" d="M812.571 150.476c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095c16.622 0 30.096 13.474 30.096 30.095s-13.474 30.095-30.096 30.095zm0-60.19c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm0 180.571c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095c16.622 0 30.096 13.474 30.096 30.095s-13.474 30.095-30.096 30.095zm0 60.19c-16.62 0-30.095-13.473-30.095-30.095 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm0-120.38c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm-52.165-90.286c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.096 13.475 30.096 30.096c0 16.62-13.475 30.095-30.096 30.095zm0-60.19c-16.62 0-30.095-13.475-30.095-30.096C730.311 13.475 743.785 0 760.406 0s30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 300.952c-16.62 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.096 13.474 30.096 30.096c0 16.62-13.475 30.095-30.096 30.095zm0-180.572c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 60.19c-16.62 0-30.095-13.473-30.095-30.094 0-16.621 13.474-30.096 30.095-30.096s30.096 13.475 30.096 30.096c0 16.62-13.475 30.095-30.096 30.095zm0 60.191c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zM708.241 90.286c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm0 180.571c-16.62 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 60.19c-16.62 0-30.095-13.473-30.095-30.095 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zM603.911 90.287c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm52.165-210.667c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096C625.981 13.475 639.455 0 656.076 0s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 300.952c-16.621 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096 0-16.621 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zM551.746 120.38c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096C521.65 13.475 535.125 0 551.746 0s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 300.952c-16.621 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zM499.581 90.286c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zM447.416 120.38c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.096 30.095-30.096 30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-60.19c-16.621 0-30.095-13.475-30.095-30.096C417.32 13.475 430.795 0 447.416 0s30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 300.952c-16.621 0-30.095-13.474-30.095-30.095 0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zM395.25 90.286c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.621 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zM343.086 120.38c-16.621 0-30.096-13.474-30.096-30.095s13.475-30.096 30.096-30.096c16.62 0 30.095 13.475 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zm0 240.762c-16.621 0-30.096-13.474-30.096-30.095 0-16.622 13.475-30.096 30.096-30.096 16.62 0 30.095 13.474 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zM134.426 120.38c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.096 30.095-30.096c16.622 0 30.096 13.475 30.096 30.096 0 16.62-13.474 30.095-30.096 30.095zm0 240.762c-16.622 0-30.096-13.474-30.096-30.095 0-16.622 13.474-30.096 30.095-30.096 16.622 0 30.096 13.474 30.096 30.096 0 16.62-13.474 30.095-30.096 30.095zm0-60.19c-16.622 0-30.096-13.475-30.096-30.096 0-16.621 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095s-13.474 30.095-30.096 30.095zM290.92 90.285c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.096-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zm0 240.762c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.096-30.095 16.62 0 30.095 13.474 30.095 30.095 0 16.622-13.474 30.096-30.095 30.096zM82.26 150.476c-16.622 0-30.096-13.474-30.096-30.095S65.64 90.286 82.26 90.286s30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0-60.19c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm0 180.571c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.095-30.095 30.096 13.474 30.096 30.095-13.475 30.095-30.096 30.095zm0 60.19c-16.622 0-30.096-13.473-30.096-30.095 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm0-120.38c-16.622 0-30.096-13.474-30.096-30.096 0-16.62 13.474-30.095 30.095-30.095s30.096 13.474 30.096 30.095c0 16.622-13.475 30.096-30.096 30.096zm156.495-90.286c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.096 30.096-30.096c16.62 0 30.095 13.475 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zm0 60.19c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095zm0 60.19c-16.622 0-30.096-13.473-30.096-30.094 0-16.621 13.474-30.096 30.096-30.096 16.62 0 30.095 13.475 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zm0 60.191c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095zM603.91 150.476c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 60.19c-16.621 0-30.095-13.473-30.095-30.095 0-16.62 13.474-30.095 30.095-30.095s30.095 13.474 30.095 30.095c0 16.622-13.474 30.096-30.095 30.096zm0 60.191c-16.621 0-30.095-13.474-30.095-30.095s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm-365.155 90.286c-16.622 0-30.096-13.474-30.096-30.095 0-16.622 13.474-30.096 30.096-30.096 16.62 0 30.095 13.474 30.095 30.096 0 16.62-13.474 30.095-30.095 30.095zM186.59 90.286c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zm0 240.762c-16.62 0-30.095-13.474-30.095-30.096 0-16.62 13.474-30.095 30.095-30.095 16.622 0 30.096 13.474 30.096 30.095 0 16.622-13.474 30.096-30.096 30.096zM30.095 120.38C13.475 120.381 0 106.907 0 90.286S13.474 60.19 30.095 60.19 60.19 73.665 60.19 90.286c0 16.62-13.474 30.095-30.095 30.095zm0-60.19C13.475 60.19 0 46.715 0 30.094 0 13.475 13.474 0 30.095 0S60.19 13.474 60.19 30.095 46.716 60.19 30.095 60.19zm0 300.952C13.475 361.143 0 347.669 0 331.048c0-16.622 13.474-30.096 30.095-30.096s30.095 13.474 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0-180.572C13.475 180.571 0 167.097 0 150.476s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm0 60.19C13.475 240.762 0 227.289 0 210.668c0-16.621 13.474-30.096 30.095-30.096s30.095 13.475 30.095 30.096c0 16.62-13.474 30.095-30.095 30.095zm0 60.191C13.475 300.952 0 287.478 0 270.857s13.474-30.095 30.095-30.095 30.095 13.474 30.095 30.095-13.474 30.095-30.095 30.095zm260.826-30.095c-16.622 0-30.096-13.474-30.096-30.095s13.474-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095zm52.165 30.095c-16.621 0-30.096-13.474-30.096-30.095s13.475-30.095 30.096-30.095c16.62 0 30.095 13.474 30.095 30.095s-13.474 30.095-30.095 30.095z" id="a" />
+        <circle cx="552" cy="271" fill="#C28A2F" r="30" transform="matrix(-1 0 0 1 1104 0)" />
+        <text fill="#121212" fontSize="36" fontWeight="bold">
+          <tspan x="280" y="170">We haven’t put</tspan>
+        </text>
+        <text fill="#121212" fontSize="36" fontWeight="bold">
+          <tspan x="327" y="205">up a paywall</tspan>
+        </text>
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgPaywallMobile };
