+++ assets/components/svgs/payWallWide.jsx
@@ -1,0 +1,33 @@
+// @flow
+
+import React from 'react';
+
+function SvgPaywallWide() {
+
+  return (
+    <svg
+      className="svg-paywall-wide"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 890 360"
+      width="890"
+      height="360"
+      preserveAspectRatio="xMinYMid"
+      aria-labelledby="title"
+    >
+      <title>We haven&#39;t put up a paywall</title>
+      <g fill="none" fillRule="evenodd">
+        <path fill="#121212" d="M52.961 190.691A29.93 29.93 0 0 1 51 180c0-3.766.694-7.37 1.961-10.691C47.458 175.846 39.214 180 30 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 120c0-3.766.694-7.37 1.961-10.691C47.458 115.846 39.214 120 30 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 60c0-3.766.694-7.37 1.961-10.691C47.458 55.846 39.214 60 30 60 13.431 60 0 46.569 0 30 0 13.431 13.431 0 30 0c16.569 0 30 13.431 30 30 0 3.766-.694 7.37-1.961 10.691C63.542 34.154 71.786 30 81 30c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 90c0 3.766-.694 7.37-1.961 10.691C63.542 94.154 71.786 90 81 90c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 150c0 3.766-.694 7.37-1.961 10.691C63.542 154.154 71.786 150 81 150c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 210c0 3.766-.694 7.37-1.961 10.691C63.542 214.154 71.786 210 81 210c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 270c0 3.766-.694 7.37-1.961 10.691C63.542 274.154 71.786 270 81 270c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 60 330c0 16.569-13.431 30-30 30-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 300c0-3.766.694-7.37 1.961-10.691C47.458 295.846 39.214 300 30 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691A29.93 29.93 0 0 1 51 240c0-3.766.694-7.37 1.961-10.691C47.458 235.846 39.214 240 30 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691zm160.078 120C218.542 304.154 226.786 300 236 300c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-16.569 0-30-13.431-30-30 0-3.766.694-7.37 1.961-10.691C202.458 325.846 194.214 330 185 330c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 3.766-.694 7.37-1.961 10.691zm-5.078-210A29.93 29.93 0 0 1 206 90c0-3.766.694-7.37 1.961-10.691C202.458 85.846 194.214 90 185 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 3.766-.694 7.37-1.961 10.691C218.542 64.154 226.786 60 236 60c16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30-9.214 0-17.458-4.154-22.961-10.691A29.93 29.93 0 0 1 215 120c0 16.569-13.431 30-30 30-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 9.214 0 17.458 4.154 22.961 10.691zM808 150c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM756 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM704 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM600 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM548 120c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM496 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM392 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-210c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zM288 90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 240c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-103-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm-52-90c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 300c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0-180c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30zm0 60c-16.569 0-30-13.431-30-30 0-16.569 13.431-30 30-30 16.569 0 30 13.431 30 30 0 16.569-13.431 30-30 30z" />
+        <circle cx="600" cy="241" fill="#C28A2F" r="30" transform="matrix(-1 0 0 1 1200 0)" />
+        <text fill="#121212" fontSize="56" fontWeight="bold">
+          <tspan x="225" y="173">We haven’t put up</tspan>
+        </text>
+        <text fill="#121212" fontSize="56" fontWeight="bold">
+          <tspan x="270" y="230">a paywall</tspan>
+        </text>
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgPaywallWide };
