+++ assets/components/svgs/scribble.jsx
@@ -1,0 +1,45 @@
+// @flow
+
+import React from 'react';
+
+function SvgScribble() {
+
+  return (
+    <svg
+      className="svg-scribble"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 640 218"
+      preserveAspectRatio="xMinYMid"
+      aria-labelledby="title"
+    >
+      <title>No one edits our editor</title>
+      <g fill="none" fillRule="evenodd" transform="translate(-290) translate(314 9)">
+        <path d="M558.726 55.313l-19.435 12.931 1.98-25.508-27.768 34.594 8.726-50.6-32.131 51.224 9.093-39.001-40.138 57.37 19.98-67.161-34.686 45.065 5.767-31.49-23.185 28.3 6.48-36.477-30.084 42.77L426.929 0l-45.234 77.954 13.42-39.001-29.648 39.001 20.18-62.264-32.397 61.64 12.217-54.721L338.73 83.62l7.342-58.18-27.868 55.662 6.865-46.542-32.961 42.77 17.768-71.81-47.112 78.1 5.787-40.884c-23.274 27.256-34.647 40.884-34.12 40.884.527 0 7.404-20.337 20.632-61.011l-27.77 48.429V42.736L204.284 83.62l16.662-58.18-38.88 55.662 7.144-38.366-30.946 47.803 19.835-65.1C150.594 62.97 136.84 82.362 136.84 83.62V55.313L108.285 94.94l18.248-79.249-41.26 89.944 23.012-83.025-42.054 67.93 13.497-67.93C45.417 72.776 27.694 97.726 26.561 97.458c-1.134-.268 8.056-26.549 27.568-78.842L7.522 77.33 26.56 8.653 0 38.953l4.162-26.407" stroke="#121212" strokeLinecap="round" strokeLinejoin="round" strokeWidth="8" />
+        <path d="M611.067 151.56l-11.02 12.32-5.94-22.847-20.355 22.846 3.398-31.775-34.88 41.1 14.343-52.583-31.688 52.583 21.568-83.946-34.626 83.105 13.058-73.777-28.577 82.257 7.847-78.44-29.786 75.044 7.338-62.749-35.23 57.665 18.992-96.815-50.354 105.295 6.184-55.12c-24.874 36.746-37.03 55.12-36.467 55.12.563 0 7.914-27.42 22.051-82.257l-29.68 65.293v-38.157l-24.592 55.12 17.808-78.44-41.555 75.045 7.635-51.725-33.075 64.449 21.2-87.769c-31.688 56.817-47.532 86.073-47.532 87.769 0 1.696 3.31-27.56 9.931-87.769l-37.017 93.701 4.548-70.38-13.608 31.882L267.828 85.3l-48.834 87.904 18.441-70.802-43.616 81.935 18.756-76.916-39.052 65.783 10.826-65.783-29.442 44.139-6.897-19.456-16.346 14.56" stroke="#121212" strokeLinecap="round" strokeLinejoin="round" strokeWidth="8" />
+        <g fill="#FFF" transform="translate(80.08 24.8)">
+          <circle cx="51.876" cy="51.84" r="50.88" />
+          <circle cx="163.88" cy="51.88" r="50.88" />
+          <circle cx="261.88" cy="51.88" r="50.88" />
+          <circle cx="334.88" cy="50.88" r="50.88" />
+          <circle cx="213.88" cy="71.88" r="50.88" />
+          <circle cx="123.88" cy="61.88" r="50.88" />
+          <circle cx="221.88" cy="99.88" r="50.88" />
+          <circle cx="313.88" cy="98.88" r="50.88" />
+          <circle cx="153.88" cy="111.88" r="50.88" />
+          <circle cx="256.88" cy="108.88" r="50.88" />
+          <circle cx="375.88" cy="112.88" r="50.88" />
+        </g>
+        <text fontSize="56" fontWeight="bold">
+          <tspan fill="#121212" x="110" y="93">No one edits</tspan>
+        </text>
+        <text fontSize="56" fontWeight="bold">
+          <tspan fill="#121212" x="207" y="150">our editor</tspan>
+        </text>
+        <circle cx="513" cy="179" fill="#1896D7" r="30" />
+      </g>
+    </svg>
+  );
+
+}
+
+export { SvgScribble };
