+++ assets/components/svgs/thankYouHeroDesktop.jsx
@@ -1,0 +1,74 @@
+// @flow
+
+import React from 'react';
+
+function SvgThankYouHeroDesktop() {
+
+  return (
+    <svg
+      className="svg-thank-you-hero-desktop"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 1400 308"
+      preserveAspectRatio="xMinYMid slice"
+    >
+      <g fill="none" fillRule="evenodd">
+        <g transform="scale(1 -1) rotate(-55 -173.131 -218.706)">
+          <path
+            d="M174.133 124.74C185.688 114.745 193 99.976 193 83.5c0-28.288-21.552-51.543-49.133-54.239C153.422 20.998 165.877 16 179.5 16c30.1 0 54.5 24.4 54.5 54.5S209.6 125 179.5 125c-1.811 0-3.602-.088-5.367-.26z"
+            fill="#1896D7"
+            opacity=".399"
+          />
+          <path
+            d="M50.647 108.866c1.273.089 2.558.134 3.853.134 30.1 0 54.5-24.4 54.5-54.5a54.256 54.256 0 0 0-10.647-32.366C126.655 24.111 149 47.696 149 76.5c0 30.1-24.4 54.5-54.5 54.5-17.98 0-33.926-8.707-43.853-22.134z"
+            fill="#CE147C"
+            opacity=".8"
+          />
+          <circle cx="54.5" cy="54.5" fill="#CE147C" r="54.5" />
+          <path
+            d="M108.692 129.134C131.91 122.889 149 101.69 149 76.5c0-19.098-9.823-35.901-24.692-45.634A54.571 54.571 0 0 1 138.5 29c30.1 0 54.5 24.4 54.5 54.5S168.6 138 138.5 138c-11.002 0-21.242-3.26-29.808-8.866z"
+            fill="#CE147C"
+            opacity=".604"
+          />
+        </g>
+        <g fill="#F7691C" transform="scale(1 -1) rotate(-22 -554.639 -1815.699)">
+          <path
+            d="M44.104 120.11c4.48-8.004 11.045-14.984 19.47-20.046 18.818-11.307 41.722-9.926 58.756 1.66 9.24-16.506 9.616-37.365-.824-54.74-15.502-25.8-48.985-34.148-74.785-18.646-25.8 15.502-34.148 48.985-18.646 74.785a54.297 54.297 0 0 0 16.03 16.986z"
+            fillOpacity=".4"
+          />
+          <path
+            d="M86.567 201.046c1.836-16.282 10.964-31.555 26.094-40.646a54.256 54.256 0 0 1 33.146-7.55 54.262 54.262 0 0 0-7.448-34.14c-15.502-25.8-48.985-34.149-74.785-18.646-25.8 15.502-34.148 48.985-18.646 74.785 9.275 15.437 24.987 24.626 41.639 26.197z"
+            fillOpacity=".6"
+          />
+          <circle cx="140.73" cy="207.116" fillOpacity=".8" r="54.5" transform="scale(1 -1) rotate(-59 -225.346 0)" />
+          <circle cx="224.643" cy="244.671" r="54.5" transform="scale(1 -1) rotate(-59 -207.811 0)" />
+        </g>
+        <g fill="#FF1A21" transform="rotate(179 230.917 167.735)">
+          <path
+            d="M56.344 127.9c2.87-8.712 7.982-16.816 15.287-23.393 16.314-14.69 39.06-17.704 57.993-9.582 5.92-17.965 2.31-38.513-11.254-53.577-20.14-22.368-54.6-24.174-76.969-4.033-22.368 20.14-24.174 54.6-4.033 76.969a54.297 54.297 0 0 0 18.976 13.615z"
+            opacity=".2"
+          />
+          <path
+            d="M108.34 199.367c.124-16.384 7.606-32.528 21.702-43.15a54.256 54.256 0 0 1 32.176-10.975 54.262 54.262 0 0 0-10.976-33.174c-18.115-24.038-52.286-28.84-76.325-10.726-24.038 18.114-28.841 52.286-10.727 76.324 10.838 14.383 27.424 21.88 44.15 21.701z"
+            opacity=".4"
+          />
+          <path
+            d="M216.938 198.218c-16.643 10.6-25.652 28.838-25.193 47.357-.405.258-.815.511-1.229.76-25.8 15.503-59.283 7.154-74.785-18.646-15.502-25.8-7.154-59.283 18.646-74.785 25.8-15.502 59.283-7.154 74.785 18.646 5.031 8.373 7.55 17.555 7.776 26.668z"
+            opacity=".6"
+          />
+          <circle cx="246.237" cy="244.173" opacity=".8" r="54.5" transform="scale(1 -1) rotate(-59 -185.338 0)" />
+          <circle cx="366.747" cy="273.281" r="54.5" transform="scale(1 -1) rotate(-59 -116.275 0)" />
+        </g>
+        <g transform="rotate(179 415.292 185.413)">
+          <circle cx="327.714" cy="75.779" fill="#BDDBEA" r="54.5" transform="scale(1 -1) rotate(-59 193.775 0)" />
+          <circle cx="292.38" cy="209.183" fill="#94CAE5" r="54.5" transform="scale(1 -1) rotate(-59 -77.349 0)" />
+          <circle cx="235.202" cy="276.195" fill="#7FC1E3" r="54.5" transform="scale(1 -1) rotate(-59 -252.971 0)" />
+          <circle cx="186.985" cy="292.117" fill="#6AB8E0" r="54.5" transform="scale(1 -1) rotate(-59 -329.33 0)" />
+          <circle cx="121.155" cy="285.798" fill="#41A7DB" r="54.5" transform="scale(1 -1) rotate(-59 -383.992 0)" />
+          <circle cx="74.984" cy="238.052" fill="#1896D7" r="54.5" transform="scale(1 -1) rotate(-59 -345.771 0)" />
+        </g>
+      </g>
+    </svg>
+  );
+}
+
+export { SvgThankYouHeroDesktop };
