+++ assets/components/svgs/thankYouHeroMobile.jsx
@@ -1,0 +1,57 @@
+// @flow
+
+import React from 'react';
+
+function SvgThankYouHeroMobile() {
+
+  return (
+    <svg
+      className="svg-thank-you-hero-mobile"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 375 277"
+      preserveAspectRatio="xMinYMid"
+    >
+      <g fill="#CE147C" fillRule="evenodd" transform="rotate(-148 126.06 183.951)">
+        <path
+          d="M56.344 127.9c2.87-8.712 7.982-16.816 15.287-23.393 16.314-14.69 39.06-17.704 57.993-9.582 5.92-17.965 2.31-38.513-11.254-53.577-20.14-22.368-54.6-24.174-76.969-4.033-22.368 20.14-24.174 54.6-4.033 76.969a54.297 54.297 0 0 0 18.976 13.615z"
+          opacity=".2"
+        />
+        <path
+          d="M108.34 199.367c.124-16.384 7.606-32.528 21.702-43.15a54.256 54.256 0 0 1 32.176-10.975 54.262 54.262 0 0 0-10.976-33.174c-18.115-24.038-52.286-28.84-76.325-10.726-24.038 18.114-28.841 52.286-10.727 76.324 10.838 14.383 27.424 21.88 44.15 21.701z"
+          opacity=".4"
+        />
+        <path
+          d="M216.938 198.218c-16.643 10.6-25.652 28.838-25.193 47.357-.405.258-.815.511-1.229.76-25.8 15.503-59.283 7.154-74.785-18.646-15.502-25.8-7.154-59.283 18.646-74.785 25.8-15.502 59.283-7.154 74.785 18.646 5.031 8.373 7.55 17.555 7.776 26.668z"
+          opacity=".6"
+        />
+        <circle cx="246.237" cy="244.173" opacity=".8" r="54.5" transform="scale(1 -1) rotate(-59 -185.338 0)" />
+        <circle cx="366.747" cy="273.281" r="54.5" transform="scale(1 -1) rotate(-59 -116.275 0)" />
+      </g>
+      <g transform="rotate(156 224.236 227.744)">
+        <ellipse cx="278.059" cy="173.895" fill="#94CAE5" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -29.299 0)" />
+        <ellipse cx="231.847" cy="254.693" fill="#7FC1E3" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -218.322 0)" />
+        <ellipse cx="190.741" cy="300.625" fill="#6AB8E0" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -340.611 0)" />
+        <ellipse cx="133.727" cy="326.437" fill="#41A7DB" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -443.248 0)" />
+        <ellipse cx="67.806" cy="339.043" fill="#1896D7" rx="49.226" ry="49.302" transform="scale(1 -1) rotate(-59 -531.45 0)" />
+      </g>
+      <g transform="rotate(119 136.185 213.425)">
+        <path
+          d="M157.017 112.084c10.42-8.98 17.013-22.25 17.013-37.055 0-25.419-19.434-46.314-44.303-48.737a49.053 49.053 0 0 1 32.13-11.915c27.14 0 49.143 21.925 49.143 48.97 0 27.047-22.002 48.972-49.143 48.972a49.91 49.91 0 0 1-4.84-.235z"
+          fill="#F7691C"
+          opacity=".399"
+        />
+        <path
+          d="M45.669 97.822c1.148.08 2.306.12 3.474.12 27.141 0 49.143-21.925 49.143-48.971 0-10.89-3.567-20.95-9.6-29.082 25.52 1.776 45.669 22.968 45.669 48.85 0 27.046-22.002 48.971-49.143 48.971-16.213 0-30.592-7.823-39.543-19.888z"
+          fill="#F48345"
+        />
+        <ellipse cx="49.143" cy="48.971" fill="#F7691C" rx="49.143" ry="48.971" />
+        <path
+          d="M98.009 116.033c20.936-5.611 36.346-24.66 36.346-47.294 0-17.16-8.858-32.26-22.266-41.004a49.37 49.37 0 0 1 12.798-1.677c27.14 0 49.143 21.925 49.143 48.971 0 27.046-22.002 48.971-49.143 48.971-9.92 0-19.154-2.93-26.878-7.967z"
+          fill="#F09D70"
+        />
+      </g>
+    </svg>
+  );
+}
+
+export { SvgThankYouHeroMobile };
