+++ assets/components/svgs/twitter.jsx
@@ -1,0 +1,14 @@
+// @flow
+
+import React from 'react';
+
+// A monochrome version of the Twitter logo.
+function SvgTwitter() {
+  return (
+    <svg className="svg-twitter" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" preserveAspectRatio="xMinYMid">
+      <path d="M13.553 5.05v.399c0 3.746-2.79 8.051-7.892 8.051-1.594 0-3.029-.478-4.224-1.275.239 0 .478.08.638.08 1.275 0 2.471-.478 3.428-1.196-1.196 0-2.232-.797-2.631-1.913.159 0 .319.08.558.08l.718-.08c-1.275-.239-2.232-1.434-2.232-2.79.399.239.797.319 1.275.319-.718-.478-1.275-1.355-1.275-2.312 0-.478.159-1.037.399-1.434 1.355 1.674 3.428 2.87 5.74 2.95-.08-.239-.08-.399-.08-.638 0-1.594 1.275-2.79 2.79-2.79.797 0 1.514.319 1.993.877.638-.08 1.196-.319 1.753-.638-.239.638-.638 1.196-1.196 1.514.558-.08 1.116-.239 1.594-.399-.319.319-.797.797-1.355 1.196l-.001-.001z" />
+    </svg>
+  );
+}
+
+export { SvgTwitter };
