+++ assets/components/svgs/video.jsx
@@ -1,0 +1,16 @@
+// @flow
+
+import React from 'react';
+
+// A video camera icon.
+function SvgVideo() {
+
+  return (
+    <svg className="svg-video" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 12.728" preserveAspectRatio="xMinYMid">
+      <path d="M0 1.818v9.091l1.818 1.819h10.682v-12.728h-10.682zM18.75.165l-5 5v2.5l5 5.001h1.25v-12.501z" />
+    </svg>
+  );
+
+}
+
+export { SvgVideo };
