+++ assets/stylesheets/main.scss
@@ -55,7 +55,7 @@
 @import '../components/secure/secure';
 @import '../components/selectInput/selectInput';
 @import '../components/socialShare/socialShare';
-@import '../components/svg/svg';
+@import '../components/svgs/svg';
 @import '../components/whySupport/whySupport';
 
 // ----- Pages ----- //
